% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{AddOutgroup}
\alias{AddOutgroup}
\title{Function to add an outgroup to any phylogeny, in phylo or newick format}
\usage{
AddOutgroup(phy = NULL, outgroup = "outgroup", check_datelife_query = TRUE)
}
\arguments{
\item{phy}{The constraint tree: a phylo object or a newick character string, with or without branch lengths.}

\item{outgroup}{A character vector with the name of the outgroup. If it has length>1, only first element will be used.}

\item{check_datelife_query}{Boolean. If true, phy will be processed with check_datelife_query function.}
}
\value{
A character vector with a newick string.
}
\description{
Function to add an outgroup to any phylogeny, in phylo or newick format
}
