% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{AllMatching}
\alias{AllMatching}
\title{Are all desired taxa in the patristic.matrix?}
\usage{
AllMatching(patristic.matrix, taxa)
}
\arguments{
\item{patristic.matrix}{A patristic matrix, rownames and colnames must be taxa}

\item{taxa}{A vector of taxon names}
}
\value{
A Boolean
}
\description{
Are all desired taxa in the patristic.matrix?
}
