% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{author.pretty}
\alias{author.pretty}
\title{Authors from studies with chronograms in Open tree of Life}
\format{A character vector with the author names from studies with chronograms that are in OToL}
\source{
\url{http://opentreeoflife.org}
}
\usage{
author.pretty
}
\description{
Authors from studies with chronograms in Open tree of Life
}
\details{
Generated with make_contributor_cache()
}
\keyword{author}
\keyword{chronogram}
\keyword{otol}
\keyword{studies}
\keyword{study}
\keyword{tree}
