% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{author.results}
\alias{author.results}
\title{Information on authors, study ids and clades from chronograms in Open Tree of Life}
\format{A dataframe with three variables
\describe{
  \item{person}{Character vector of authors of the included studies}
  \item{study}{Character vector of study identifiers}
  \item{clade}{Character vector with names of clades with chronograms from the included studies}
}}
\source{
\url{http://opentreeoflife.org}
}
\usage{
author.results
}
\description{
Information on authors, study ids and clades from chronograms in Open Tree of Life
}
\details{
Generated with make_contributor_cache()
}
\keyword{author}
\keyword{chronogram}
\keyword{clade}
\keyword{id}
\keyword{otol}
\keyword{studies}
\keyword{study}
\keyword{tree}
