% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_phyloch_functions.R
\name{axisGeo}
\alias{axisGeo}
\title{Function from Christoph Heibl's phyloch package
Add a Geological Time Axis for Time-Calibrated Phylogenies.}
\usage{
axisGeo(
  GTS,
  tip.time = 0,
  unit = c("epoch", "period"),
  ages = TRUE,
  cex = 1,
  col = "white",
  texcol = "black",
  gridty = 0,
  gridcol = "black"
)
}
\arguments{
\item{GTS}{a data frame containing the geological time scale such as provided by \code{\link{strat2012}}.}

\item{tip.time}{a real number giving time units back from present were the youngest taxa were sampled.}

\item{unit}{a vector of mode character used to select geological time units that shall be displayed. When using \code{gradstein04}, \code{"eon"}, \code{"era"}, \code{"period"}, \code{"epoch"}, and \code{"stage"} are available.}

\item{ages}{logical: if \code{TRUE}, a real-numbered axis of unit boundaries is added to the plot.}

\item{cex}{a real number giving the \bold{c}haracter \bold{ex}pansion.}

\item{col}{a vector of mode character used to select background color(s). Will be recycled if necessary.}

\item{texcol}{a vector of mode character used to select font color(s). Will be recycled if necessary.}

\item{gridty}{an integer representing the line type of the grid exactly in the same way as \code{lty} argument in \code{\link{par}}: 0 or \code{blank}, 1 or \code{"solid"}, 2 or \code{"dashed"}, 3 or \code{"dotted"}, 4 or \code{"dotdash"}, 5 or \code{"longdash"}, and 6 or \code{"twodash"}.}

\item{gridcol}{a vector of mode character used to select grid color(s). Will be recycled if necessary.}
}
\value{
Adds geological time scale to a phylo plot.
}
\description{
Function from Christoph Heibl's phyloch package
Add a Geological Time Axis for Time-Calibrated Phylogenies.
}
\details{
Make sure ape is loaded otherwise it won't find .PlotPhyloEnv
If the name of a time interval does not fit the space on the plotting device provided by that time frame, \code{axisGeo} tries to fit the name string by lowering the original character expansion. In order to avoid undiscernable small font sizes only character expansion of 50 percent or more of the original \code{cex} is allowed; in all other cases strings are abbreviated and plotted with the original character expansion.
}
\references{
Gradstein F.M., J.G. Ogg & A.G. Smith. 2004. A Geologic Time Scale. Cambridge University Press, Cambridge, UK. \url{www.cambridge.org/uk/catalogue/catalogue.asp?isbn=0521786738}.
International Commission on Stratigraphy. 2012. International stratigraphic chart: \url{http://www.stratigraphy.org/upload/ISChart2009.pdf}.
}
\author{
Christoph Heibl
}
