% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{BindMatrices}
\alias{BindMatrices}
\title{Convert list of patristic matrices to a 3D array. Used inside: summarize_datelife_result, GetSubsetArrayCongruify.}
\usage{
BindMatrices(patristic.matrix.list, pad = TRUE)
}
\arguments{
\item{patristic.matrix.list}{List of patristic matrices}

\item{pad}{If TRUE, pad missing entries}
}
\value{
A 3d array of patristic matrices
}
\description{
Convert list of patristic matrices to a 3D array. Used inside: summarize_datelife_result, GetSubsetArrayCongruify.
}
