% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birds_and_cats}
\alias{birds_and_cats}
\title{A multiPhylo object with trees resulting from a datelife search of some birds and cats species}
\format{
A multiPhylo object
}
\usage{
birds_and_cats
}
\description{
A multiPhylo object with trees resulting from a datelife search of some birds and cats species
}
\details{
Generated with:
taxa <- c("Rhea americana", "Pterocnemia pennata", "Struthio camelus", "Gallus", "Felis")
birds_and_cats <- datelife_search(input = taxa, summary_format = "phylo_all", get_spp_from_taxon = TRUE)
usethis::use_data(birds_and_cats)
}
\keyword{datelifeResult}
