% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groves.R
\name{build_grove_list}
\alias{build_grove_list}
\title{Build grove list}
\usage{
build_grove_list(datelife_result, n = 2)
}
\arguments{
\item{datelife_result}{datelifeResult object (named list of patristic matrices)}

\item{n}{Degree of overlap required}
}
\value{
list of vectors; each list element is a grove
}
\description{
Using theorem 1.1 of Ané et al. 10.1007/s00026-009-0017-x Groves of Phylogenetic Trees
}
