% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groves.R
\name{build_groves_matrix}
\alias{build_groves_matrix}
\title{Build grove matrix}
\usage{
build_groves_matrix(datelife_result, n = 2)
}
\arguments{
\item{datelife_result}{datelifeResult object (named list of patristic matrices)}

\item{n}{Degree of overlap required}
}
\value{
matrix; each cell shows whether n-overlap exists between that pair of inputs
}
\description{
Using theorem 1.1 of Ané et al. 10.1007/s00026-009-0017-x Groves of Phylogenetic Trees
}
