% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{check_datelife_result}
\alias{check_datelife_result}
\title{checks if we obtained an empty search with the set of input taxon names}
\usage{
check_datelife_result(datelife_result)
}
\arguments{
\item{datelife_result}{An output of get_datelife_result function: A list of patristic matrices with names corresponding to original study citations.}
}
\value{
a datelifeResult object or a message explaining why it is not a datelifeResul tobject
}
\description{
checks if we obtained an empty search with the set of input taxon names
}
