% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{check_phylo}
\alias{check_phylo}
\title{Checks if phy is a phylo class object and a chronogram}
\usage{
check_phylo(phy = phy, dated = FALSE)
}
\arguments{
\item{phy}{A tree either as a newick character string or as a phylo object}

\item{dated}{Boolean. If TRUE it checks if phylo object has branch lengths and is ultrametric.}
}
\value{
A phylo object
}
\description{
Checks if phy is a phylo class object and a chronogram
}
