% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{CheckFilteredResults}
\alias{CheckFilteredResults}
\title{checks if we obtained an empty search with the set of input taxon names}
\usage{
CheckFilteredResults(filtered.results, usetnrs, verbose = FALSE)
}
\arguments{
\item{filtered.results}{An object output from get_datelife_result function}

\item{usetnrs}{If TRUE, use OpenTree's services to resolve names. This can dramatically improve the chance of matches, but also take much longer.}

\item{verbose}{Boolean. If TRUE, it gives printed updates to the user.}
}
\description{
checks if we obtained an empty search with the set of input taxon names
}
