% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{CheckInput}
\alias{CheckInput}
\title{checks if input has already been processed, otherwise it uses make_datelife_query}
\usage{
CheckInput(input, ...)
}
\arguments{
\item{input}{Target taxa names in the form of a vector of characters, a newick character string, or a phylo object.}

\item{...}{Arguments passed on to \code{make_datelife_query}
\describe{
  \item{input}{Target taxa names in the form of a vector of characters, a newick character string, or a phylo object.}
  \item{usetnrs}{If TRUE, use OpenTree's services to resolve names. This can dramatically improve the chance of matches, but also take much longer.}
  \item{approximatematch}{If TRUE, use a slower TNRS to correct mispellings, increasing the chance of matches (including false matches).}
  \item{sppfromtaxon}{boolean vector, default to FALSE. If TRUE, will get all species names from taxon names given in input. Must have same length as input. If input is a newick string , with some clades it will be converted to phylo object phy, and the order of sppfromtaxon will match phy$tip.label.}
  \item{verbose}{Boolean. If TRUE, it gives printed updates to the user.}
}}
}
\description{
checks if input has already been processed, otherwise it uses make_datelife_query
}
