% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{CheckMissingTaxa}
\alias{CheckMissingTaxa}
\title{Checks that missingTaxa argument is ok to be used by make_mrbayes_runfile function}
\usage{
CheckMissingTaxa(missingTaxa)
}
\arguments{
\item{missingTaxa}{A phylo object, a newick character string or a dataframe with taxonomic assignations}
}
\value{
A phylo object, a newick character string or a dataframe with taxonomic assignations
}
\description{
Checks that missingTaxa argument is ok to be used by make_mrbayes_runfile function
}
