% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.r
\name{checknames}
\alias{checknames}
\title{Function to check names on a species list or phylogeny.}
\usage{
checknames(phylo = NULL, charvector = NULL, source_ = "NCBI",
  writefile = FALSE, writedir = "~/", byfilename = TRUE)
}
\arguments{
\item{phylo}{Phylogeny object, class phylo}

\item{charvector}{Character vector.}

\item{source_}{Source to match names to. One of ncbi, iPlant}

\item{writefile}{Write file to directory or not. Remember to change directory to 
write to in the function. Right now it is set to 
"/Library/WebServer/Sites/datelife.org/datelife/data/"}

\item{writedir}{A directory to write file to (defaults to "~/"); ignore if writefile = FALSE.}

\item{byfilename}{Defaults to TRUE - if TRUE, then interprets the phylo argument input 
using eval() so executes it (used in name fixing on stored trees). If FALSE, 
then you can specify phylo as a phylo object instead of the name of that object.}

\item{splitby}{Length of chunks by which to split species list}
}
\description{
Function to check names on a species list or phylogeny.
}
\examples{
\dontrun{
# Cleaning stored trees
# A phylogeny as input
library(doMC); library(ape)
out <- checknames(phylo="BergmannEtAl2012", source_="NCBI", splitby=200)
out <- checknames(phylo="AlfaroEtAl2009", source_="NCBI", splitby=200)
out <- checknames(phylo="EastmanEtAlUnpublished_tree", source_="NCBI", splitby=100)
out <- checknames(phylo="HeathEtAl2012_tree", source_="NCBI")
out <- checknames(phylo="JaffeEtAl2011", source_="NCBI", splitby=100)
out <- checknames(phylo="Chaetodontidae2011", source_="NCBI")
out <- checknames(phylo="Drosophila2012_large", source_="NCBI", splitby=50, writefile=TRUE)
out <- checknames(phylo="Drosophila2012_small", source_="NCBI", splitby=50)

out <- checknames(phylo="Apogonidae2011", source_="NCBI")
out <- checknames(phylo="BinindaEmondsEtAl2007", source_="NCBI", splitby=500)
out <- checknames(phylo="HardyCook2012", source_="NCBI")
out <- checknames(phylo="FabreEtAl2009", source_="NCBI", splitby=50)
out <- checknames(phylo="PyronWiens2011", source_="NCBI", splitby=500, writefile=TRUE)
out <- checknames(phylo="Unpub3", source_="NCBI", splitby=50, writefile=TRUE)

# Cleaning user input trees
# A character vector as input (of species names)
out <- checknames(charvector=AlfaroEtAl2009$tip.label[1:10], source_="NCBI")

# A phylo object - user suppressMessages() around the fxn
mmm<-drop.tip(AlfaroEtAl2009, 1:190)
out <- suppressMessages(checknames(phylo=phy, source_="NCBI", byfilename=FALSE))
}
}

