% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_phylo.R
\name{choose_cluster}
\alias{choose_cluster}
\title{Choose an ultrametric phylo object from cluster_patristicmatrix obtained with a particular clustering method, or the next best tree.
If there are not any ultrametric trees, it does not force them.}
\usage{
choose_cluster(phycluster, clustering_method = "nj")
}
\arguments{
\item{phycluster}{An output from cluster_patristicmatrix}

\item{clustering_method}{A character vector indicating the method to construct the tree. Options are "nj" for Neighbor-Joining and "upgma" for Unweighted Pair Group Method with Arithmetic Mean.}
}
\value{
A phylo object or NA
}
\description{
Choose an ultrametric phylo object from cluster_patristicmatrix obtained with a particular clustering method, or the next best tree.
If there are not any ultrametric trees, it does not force them.
}
