% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_from_taxonomy.R
\name{classification_paths_from_taxonomy}
\alias{classification_paths_from_taxonomy}
\title{Gets classification paths for a vector of taxa}
\usage{
classification_paths_from_taxonomy(taxa, sources = "Catalogue of Life")
}
\arguments{
\item{taxa}{Vector of taxon names}

\item{sources}{Vector of names of preferred sources; see taxize::gnr_datasources()}
}
\value{
A list with resolved taxa (a tibble, from taxize::gnr_resolve) and a vector of taxa not resolved
}
\description{
This uses the taxize package's wrapper of the Global Names Resolver to get taxonomic paths for the vector of taxa you pass in. Sources is a vector of source labels in order (though it works best if everything uses the same taxonomy, so we recommend doing just one source). You can see options by doing taxize::gnr_datasources(). Our default is Catalogue of Life
}
