% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_chronograms.R
\name{clean_chronogram}
\alias{clean_chronogram}
\title{Clean up some issues with OToL chronograms}
\usage{
clean_chronogram(phy)
}
\arguments{
\item{phy}{A phylo object}
}
\value{
A cleaned up phylo object
}
\description{
Clean up some issues with OToL chronograms
}
