% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_taxonomy.R
\name{clean_taxon_info_children}
\alias{clean_taxon_info_children}
\title{Identifies, extracts and cleans children names from a taxonomy_taxon_info() output}
\usage{
clean_taxon_info_children(taxon_info, invalid = c("barren", "extinct",
  "uncultured", "major_rank_conflict", "incertae_sedis", "unplaced",
  "conflict", "environmental", "not_otu", "hidden", "hybrid"))
}
\arguments{
\item{taxon_info}{An output of rotl::taxonomy_taxon_info()}

\item{invalid}{A character vector of flags used by Open Tree Taxonomy to detect invalid taxon names}
}
\value{
a list with valid children unique OTT names, ott_ids and ranks
}
\description{
Identifies, extracts and cleans children names from a taxonomy_taxon_info() output
}
\details{
eliminates all taxa that will give problems when trying to retrieve an induced subtree from otol.
}
\examples{
# some suppressed taxa within "Gleicheniales" family of ferns:
rotl::taxonomy_taxon_info(3866048)
rotl::taxonomy_taxon_info(5262766)
}
