% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_chronograms.R
\name{CleanChronogram}
\alias{CleanChronogram}
\title{Clean up some issues with OToL chronograms}
\usage{
CleanChronogram(phy)
}
\arguments{
\item{phy}{Input phylo object}
}
\value{
A cleaned up phylo object
}
\description{
Clean up some issues with OToL chronograms
}
