% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrations_congruify.R
\name{congruify_and_mrca}
\alias{congruify_and_mrca}
\title{Congruify nodes of a tree topology to nodes from a source chronogram, and find the mrca nodes}
\usage{
congruify_and_mrca(phy)
}
\arguments{
\item{phy}{A \code{phylo} object.}
}
\value{
An object of class data frame.
}
\description{
Congruify nodes of a tree topology to nodes from a source chronogram, and find the mrca nodes
}
