% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{ConvertSpacesToUnderscores}
\alias{ConvertSpacesToUnderscores}
\title{Convert spaces to underscores in trees}
\usage{
ConvertSpacesToUnderscores(phy)
}
\arguments{
\item{phy}{A phylo object}
}
\value{
A phylo object
}
\description{
Convert spaces to underscores in trees
}
