% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{ConvertUnderscoresToSpaces}
\alias{ConvertUnderscoresToSpaces}
\title{Convert underscores to spaces in trees}
\usage{
ConvertUnderscoresToSpaces(phy)
}
\arguments{
\item{phy}{A phylo object}
}
\value{
A phylo object
}
\description{
Convert underscores to spaces in trees
}

