% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheauthors.R
\name{CreateContributorCache}
\alias{CreateContributorCache}
\title{Create a cache  from Open Tree of Life}
\usage{
CreateContributorCache(outputfile = "contributorcache.RData", verbose = TRUE)
}
\arguments{
\item{outputfile}{Path including file name}

\item{verbose}{If TRUE, give status updates to the user}
}
\value{
List containing author and curator results
}
\description{
Create a cache  from Open Tree of Life
}
