% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheauthors.R
\name{CreateTreeBaseCache}
\alias{CreateTreeBaseCache}
\title{Create a cache from TreeBase}
\usage{
CreateTreeBaseCache(outputfile = "treebasecache.rda", verbose = TRUE)
}
\arguments{
\item{outputfile}{Path including file name}

\item{verbose}{If TRUE, give status updates to the user}
}
\value{
List containing author and curator results
}
\description{
Create a cache from TreeBase
}
