% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curator.pretty}
\alias{curator.pretty}
\title{Curators fof chronograms in Open tree of Life}
\format{A character vector with the names of curators of chronograms that are in OToL}
\source{
\url{http://opentreeoflife.org}
}
\usage{
curator.pretty
}
\description{
Curators are the ones uploading the chronograms to Open Tree of Life
}
\details{
Generated with make_contributor_cache()
}
\keyword{chronogram}
\keyword{curator}
\keyword{otol}
\keyword{studies}
\keyword{study}
\keyword{tree}
