% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_query.R
\name{datelife_query_check}
\alias{datelife_query_check}
\title{Check that input is a datelifeQuery object. To be deprecated.}
\usage{
datelife_query_check(datelife_query = NULL, ...)
}
\arguments{
\item{datelife_query}{A datelifeQuery object, output of make_datelife_query function}

\item{...}{
  Arguments passed on to \code{\link[=make_datelife_query]{make_datelife_query}}
  \describe{
    \item{\code{verbose}}{Boolean. If \code{TRUE}, it gives printed updates to the user.}
  }}
}
\value{
A datelifeQuery object
}
\description{
Check if input is a datelifeQuery object, otherwise use
[make_datelife_query()] to process it and create a \code{datelifeQuery} object.
}
