% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{datelife_query_length_check}
\alias{datelife_query_length_check}
\title{checks that we have at least two taxon names to perform a search}
\usage{
datelife_query_length_check(cleaned_names = NULL,
  get_spp_from_taxon = FALSE, verbose = FALSE)
}
\arguments{
\item{cleaned_names}{A character vector; usually an output from make_datelife_query function}

\item{get_spp_from_taxon}{boolean vector, default to FALSE. If TRUE, will get all species names from taxon names given in input. Must have same length as input. If input is a newick string , with some clades it will be converted to phylo object phy, and the order of get_spp_from_taxon will match phy$tip.label.}

\item{verbose}{Boolean. If TRUE, it gives printed updates to the user.}
}
\description{
checks that we have at least two taxon names to perform a search
}
