% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_search.R
\name{datelife_result_check}
\alias{datelife_result_check}
\title{Check if we obtained an empty search with the set of input taxon names}
\usage{
datelife_result_check(datelife_result, use_tnrs = FALSE, verbose = FALSE)
}
\arguments{
\item{datelife_result}{An output of get_datelife_result function: A list of patristic matrices with names corresponding to original study citations.}

\item{verbose}{Boolean. If \code{TRUE}, it gives printed updates to the user.}
}
\description{
Check if we obtained an empty search with the set of input taxon names
}
