% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_summary.R
\name{datelife_result_median_matrix}
\alias{datelife_result_median_matrix}
\title{Function to compute median of a datelifeResult object.}
\usage{
datelife_result_median_matrix(datelife_result)
}
\arguments{
\item{datelife_result}{An output of get_datelife_result function: A list of patristic matrices with names corresponding to original study citations.}
}
\description{
Function to compute median of a datelifeResult object.
}
