% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{datelife_result_sdm}
\alias{datelife_result_sdm}
\title{Function to compute the SDM supertree (Criscuolo et al. 2006) from a datelifeResult object.}
\usage{
datelife_result_sdm(datelife_result, weighting = "flat")
}
\arguments{
\item{datelife_result}{List of patristic matrices.}

\item{weighting}{A character vector. One of "flat", "taxa" or "inverse".}
}
\value{
A list containing phy, a chronogram from SDM, and the original datelife_result that were actually used.
}
\description{
Function to compute the SDM supertree (Criscuolo et al. 2006) from a datelifeResult object.
}
\details{
Weighting is how much weight to give each input tree.
flat = all trees have equal weighting
taxa = weight is proportional to number of taxa
inverse = weight is proportional to 1 / number of taxa
Criscuolo A, Berry V, Douzery EJ, Gascuel O. SDM: a fast distance-based approach for (super) tree building in phylogenomics. Syst Biol. 2006. 55(5):740–55. doi: 10.1080/10635150600969872.
}
