% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_sdm.R
\name{datelife_result_sdm_matrix}
\alias{datelife_result_sdm_matrix}
\title{Go from a \code{datelifeResult} object to a Super Distance Matrix (SDM) using weighting = "flat"}
\usage{
datelife_result_sdm_matrix(datelife_result)
}
\arguments{
\item{datelife_result}{A \code{datelifeResult} object, usually an output of \code{\link[=get_datelife_result]{get_datelife_result()}}.}
}
\value{
A numeric matrix.
}
\description{
Go from a \code{datelifeResult} object to a Super Distance Matrix (SDM) using weighting = "flat"
}
