% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{datelife_result_study_index}
\alias{datelife_result_study_index}
\title{Find the index of relevant studies in a opentree_chronograms object. Used inside: summarize_datelife_result.}
\usage{
datelife_result_study_index(datelife_result, cache = "opentree_chronograms")
}
\arguments{
\item{datelife_result}{An output of get_datelife_result function: A list of patristic matrices with names corresponding to original study citations.}

\item{cache}{The cache of studies}
}
\value{
A vector with the indices of studies that have relevant info
}
\description{
Find the index of relevant studies in a opentree_chronograms object. Used inside: summarize_datelife_result.
}
