% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_median.R
\name{datelife_result_variance_matrix}
\alias{datelife_result_variance_matrix}
\title{Function to compute variance matrix of a datelifeResult object.}
\usage{
datelife_result_variance_matrix(datelife_result)
}
\arguments{
\item{datelife_result}{An output of get_datelife_result function: A list of patristic matrices with names corresponding to original study citations.}
}
\value{
A variance matrix from a datelifeResult object.
}
\description{
Function to compute variance matrix of a datelifeResult object.
}
