% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_use.R
\name{datelife_use}
\alias{datelife_use}
\title{Generate one or multiple chronograms for a set of input taxa.}
\usage{
datelife_use(input = NULL, dating_method = "bladj", ...)
}
\arguments{
\item{input}{A character vector of taxon names, a tree as a 'phylo' object or
a newick character string, or a 'datelifeQuery' object from make_datelife_query function.}

\item{dating_method}{The method used for tree dating. Options are "bladj" or "pathd8"}

\item{...}{
  Arguments passed on to \code{\link[=get_all_calibrations]{get_all_calibrations}}
  \describe{
    \item{\code{each}}{Boolean, default to \code{FALSE}: all calibrations are returned in
the same data frame. If \code{TRUE}, calibrations from each chronogram are returned
in separate data frames.}
  }}
}
\value{
A list with a chronogram and secondary calibrations used
}
\description{
\code{datelife_use} generates one or multiple phylogenetic trees with branch lengths
proportional to time (i.e., chronograms) for all \code{input} taxa, dated with
bladj or PATHd8, using secondary calibrations available from the chronogram database.
}
\details{
If input is a tree, it will use secondary calibrations to (1) date the tree with bladj
if it has no branch lengths, or (2) date the tree with PATHd8 if it has branch lengths.
If input is a vector of taxon names, it will attempt to reconstruct a BOLD tree first
to get a topology with branch lengths. If it can't, it will get an Open Tree
of Life synthetic tree topology and will date it with bladj.
}
