% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_use.R
\name{datelife_use_datelifequery}
\alias{datelife_use_datelifequery}
\title{Generate one or multiple chronograms from a \code{datelifeQuery} object.}
\usage{
datelife_use_datelifequery(
  datelife_query = NULL,
  dating_method = "bladj",
  each = FALSE
)
}
\arguments{
\item{datelife_query}{A \code{datelifeQuery} object, output of \code{\link[=make_datelife_query]{make_datelife_query()}}.}

\item{dating_method}{Tree dating method to use. Options are "bladj" or "pathd8".}

\item{each}{Boolean, default to \code{FALSE}: all calibrations are returned in
the same data frame. If \code{TRUE}, calibrations from each chronogram are returned
in separate data frames.}
}
\value{
A list with a chronogram and secondary calibrations used.
}
\description{
\code{datelife_use_datelifequery} generates one or multiple chronograms (i.e., phylogenetic
trees with branch lengths proportional to time) for a set of \code{input} taxa,
dated with bladj or PATHd8, using secondary calibrations available for any pair
of \code{input} taxa, mined from the code{\link[=opentree_chronograms]{opentree_chronograms}}
local DateLife database.
}
