% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_use.R
\name{datelife_use_vector}
\alias{datelife_use_vector}
\alias{use_all_calibrations_vector}
\title{Generate one or multiple chronograms for a set of given taxon names.}
\usage{
datelife_use_vector(input = NULL, dating_method = "bladj", ...)
}
\arguments{
\item{input}{A character vector of taxon names}

\item{dating_method}{The method used for tree dating. Options are "bladj" or "pathd8"}

\item{...}{
  Arguments passed on to \code{\link[=get_all_calibrations]{get_all_calibrations}}
  \describe{
    \item{\code{each}}{Boolean, default to \code{FALSE}: all calibrations are returned in
the same data frame. If \code{TRUE}, calibrations from each chronogram are returned
in separate data frames.}
  }}
}
\value{
A list with a chronogram and secondary calibrations used
}
\description{
\code{datelife_use_vector} generates one or multiple chronograms (i.e., phylogenetic
trees with branch lengths proportional to time) for a set of \code{input} taxa,
dated with bladj or PATHd8, using secondary calibrations available for any pair
of \code{input} taxa, mined from the code{\link[=opentree_chronograms]{opentree_chronograms}}
local DateLife database.
}
\details{
If input is a tree, it will use secondary calibrations to (1) date the tree with bladj
if it has no branch lengths, or (2) date the tree with PATHd8 if it has branch lengths.
If input is a vector of taxon names, it will attempt to reconstruct a BOLD tree first
to get a topology with branch lengths. If it can't, it will get an Open Tree
of Life synthetic tree topology and will date it with bladj.
}
