% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dlsearch_subset2}
\alias{dlsearch_subset2}
\title{A list with datelieQuery and datelifeResult objects from a search of taxon names from names_subset2}
\format{A list with two named elements. datelifeResult object with 24 patristic matrices
\describe{
  \item{datelife_query}{A datelifeQuery object using names_subset 2 as input.}
  \item{datelife_result}{A datelifeResult object resulting from a search of names in datelifeQuery}
}}
\usage{
dlsearch_subset2
}
\description{
A list with datelieQuery and datelifeResult objects from a search of taxon names from names_subset2
}
\details{
Generated with:
datelife_query <- make_datelife_query(names_subset2)
datelife_result <- get_datelife_result(datelife_query)
dlsearch_subset2 <- list(datelife_result = datelife_result, datelife_query = datelife_query)
usethis::use_data(dlsearch_subset2)
}
\keyword{datelifeResult}
\keyword{names}
\keyword{subset2}
\keyword{taxon}
