% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_taxonomy.R
\name{.get_ott_lineage}
\alias{.get_ott_lineage}
\title{Get the lineage of a set of taxa.
\code{.get_ott_lineage} uses \code{\link[rotl:taxonomy_taxon_info]{rotl::taxonomy_taxon_info()}} with \code{include_lineage = TRUE}.}
\usage{
.get_ott_lineage(input_ott_match)
}
\arguments{
\item{input_ott_match}{An Output of check_ott_input function.}
}
\value{
A taxonomy_taxon_info object
}
\description{
Get the lineage of a set of taxa.
\code{.get_ott_lineage} uses \code{\link[rotl:taxonomy_taxon_info]{rotl::taxonomy_taxon_info()}} with \code{include_lineage = TRUE}.
}
