% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.r
\name{drop_epithets}
\alias{drop_epithets}
\title{Function to detect messy names and remove them, and remove sp.'s}
\usage{
drop_epithets(taxon)
}
\arguments{
\item{taxon}{Taxonomic name, or a vector of taxonomic names.}
}
\description{
Function to detect messy names and remove them, and remove sp.'s
}
\examples{
\dontrun{
drop_epithets(taxon = Oaks2011$tip.label)
}
}

