% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrations_get.R
\name{extract_calibrations_dateliferesult}
\alias{extract_calibrations_dateliferesult}
\title{Extract secondary calibrations from a given \code{datelifeResult} object}
\usage{
extract_calibrations_dateliferesult(input = NULL, each = FALSE)
}
\arguments{
\item{input}{A \code{datelifeResult} object.}

\item{each}{Boolean, default to \code{FALSE}: all calibrations are returned in
the same data frame. If \code{TRUE}, calibrations from each chronogram are returned
in separate data frames.}
}
\value{
An object of class \code{datelifeCalibrations} -- a \code{data frame}
of secondary calibrations (or list of \code{data frames}, if \code{each = TRUE}),
for each pair of taxon names in \{input]. The attribute "chronograms" stores
the source data from which the calibrations were extracted.
}
\description{
\code{extract_calibrations_dateliferesult} extracts divergence
times (i.e., secondary calibrations) for each taxon pair in a given
'\code{datelifeResult} object.
}
\details{
The function calls summarize_datelife_result()] with
\code{summary_format = "phylo_all"} to go from a \code{datelifeResult} object
to a \code{phylo} or \code{multiPhylo} object that will be passed to
[extract_calibrations_phylo()].
}
