% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{filter_for_grove}
\alias{filter_for_grove}
\title{Filter a datelifeResult to find the largest grove}
\usage{
filter_for_grove(datelife_result, criterion = c("trees", "taxa"))
}
\arguments{
\item{datelife_result}{datelifeResult object (named list of patristic matrices). Only needed for "taxa" criterion}

\item{criterion}{Whether to get the grove with the most trees or the most taxa}
}
\value{
A datelifResult object filtered to only include one grove of trees
}
\description{
Filter a datelifeResult to find the largest grove
}
