% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{FindMatchingStudyIndex}
\alias{FindMatchingStudyIndex}
\title{Find the index of relevant studies in a datelife.cache object}
\usage{
FindMatchingStudyIndex(results.list, taxa = NULL, partial = FALSE)
}
\arguments{
\item{results.list}{A list returned from using GetSubsetArrayDispatch on datelife.cache$trees}

\item{taxa}{A vector of taxa to match}

\item{partial}{If TRUE, return matrices that have only partial matches}
}
\value{
A vector with the indices of studies that have relevant info
}
\description{
Find the index of relevant studies in a datelife.cache object
}

