% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{force_ultrametric}
\alias{force_ultrametric}
\title{Force a non ultrametric phylo object to be ultrametric}
\usage{
force_ultrametric(phy)
}
\arguments{
\item{phy}{A phylo object}
}
\value{
A phylo object
}
\description{
Force a non ultrametric phylo object to be ultrametric
}
