% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groves.R
\name{get_best_grove}
\alias{get_best_grove}
\title{Get the best grove from a datelifeResult object that can be summmarized with median or sdm.}
\usage{
get_best_grove(datelife_result, criterion = "taxa", overlap = 2)
}
\arguments{
\item{datelife_result}{datelifeResult object (named list of patristic matrices). Only needed for "taxa" criterion}

\item{criterion}{Whether to get the grove with the most trees or the most taxa}
}
\value{
A datelifResult object filtered to only include one grove of trees that can be summarized with median or sdm.
}
\description{
Get the best grove from a datelifeResult object that can be summmarized with median or sdm.
}
