% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_summary.R
\name{get_biggest_multiphylo}
\alias{get_biggest_multiphylo}
\title{Get the tree with the most tips from a multiPhylo object: the biggest tree.}
\usage{
get_biggest_multiphylo(trees)
}
\arguments{
\item{trees}{A list of trees as \code{multiPhylo} or as a generic \code{list} object.}
}
\value{
The largest tree from those given in \code{trees}, as a \code{phylo} object with an additional \verb{$citation} element containing the reference of the original publication.
}
\description{
Get the tree with the most tips from a multiPhylo object: the biggest tree.
}
