% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_summary.R
\name{get_biggest_phylo}
\alias{get_biggest_phylo}
\title{Get the tree with the most tips: the biggest tree}
\usage{
get_biggest_phylo(trees)
}
\arguments{
\item{trees}{A list of trees as multiPhylo or as a plain list object.}
}
\value{
A phylo object with a citation slot with the citation of the biggest tree
}
\description{
Get the tree with the most tips: the biggest tree
}
