% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_calibrations_get.R
\name{get_calibrations_dateliferesult}
\alias{get_calibrations_dateliferesult}
\title{Get secondary calibrations from a given \code{datelifeResult} object.}
\usage{
get_calibrations_dateliferesult(input = NULL, each = FALSE)
}
\arguments{
\item{input}{A \code{datelifeResult} object.}

\item{each}{Boolean, default to \code{FALSE}: all calibrations are returned in
the same data frame. If \code{TRUE}, calibrations from each chronogram are returned
in separate data frames.}
}
\value{
A data frame of secondary calibrations (or list of data frames, if \code{each = TRUE}),
for each pair of given taxon names. The attribute "chronograms" contains the
source chronograms from which the calibrations were obtained.
}
\description{
\code{get_calibrations_dateliferesult} gets divergence times (i.e., secondary
calibrations) for each taxon pair in a given \code{datelifeResult} object.
}
\details{
The function calls
\code{\link[=summarize_datelife_result]{summarize_datelife_result}} with
\code{summary_format = "phylo_all"} to go from a \code{datelifeResult} object
to a \code{phylo} or \code{multiPhylo} object that will be passed to
\code{\link[=get_calibrations_phylo]{get_calibrations_phylo}}.
}
