% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_calibrations_get.R
\name{get_calibrations_phylo}
\alias{get_calibrations_phylo}
\title{Get secondary calibrations from a \code{phylo} or \code{multiPhylo} object with
branch lengths proportional to time.}
\usage{
get_calibrations_phylo(input = NULL, each = FALSE)
}
\arguments{
\item{input}{a phylo or multiPhylo object with branch lengths proportional to time.}

\item{each}{Boolean, default to \code{FALSE}: all calibrations are returned in
the same data frame. If \code{TRUE}, calibrations from each chronogram are returned
in separate data frames.}
}
\value{
A data frame (or list of data frames, if \code{each = TRUE}) of secondary
calibrations available for each pair of taxon names in a given \code{phylo}
or \code{multiPhylo} object. The attribute "chronograms" contains the \{input]
chronograms from which the secondary calibrations were extracted.
}
\description{
\code{get_calibrations_phylo} gets divergence times (i.e., secondary
calibrations) for each taxon pair in a given \code{phylo} or \code{multiPhylo}
object.
}
