% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otol_tree.R
\name{get_dated_otol_induced_subtree}
\alias{get_dated_otol_induced_subtree}
\title{Get an otol induced dated subtree from your set of queried taxa}
\usage{
get_dated_otol_induced_subtree(input = NULL, ott_ids = NULL, ...)
}
\arguments{
\item{input}{Optional. A character vector of names or a datelifeQuery object}

\item{ott_ids}{If not NULL, it takes this argument and ignores input. A numeric vector of ott ids obtained with rotl::taxonomy_taxon_info or rolt::tnrs_match_names or datelife::tnrs_match}

\item{...}{Arguments passed on to \code{make_datelife_query}
\describe{
  \item{use_tnrs}{If TRUE, use OpenTree's services to resolve names. This can dramatically improve the chance of matches, but also take much longer.}
  \item{approximate_match}{If TRUE, use a slower TNRS to correct mispellings, increasing the chance of matches (including false matches).}
  \item{get_spp_from_taxon}{boolean vector, default to FALSE. If TRUE, will get all species names from taxon names given in input. Must have same length as input. If input is a newick string , with some clades it will be converted to phylo object phy, and the order of get_spp_from_taxon will match phy$tip.label.}
  \item{verbose}{Boolean. If TRUE, it gives printed updates to the user.}
}}
}
\value{
A phylo object with edge length proportional to time in Myrs. It will return NA if any ott_id is invalid.
}
\description{
Get an otol induced dated subtree from your set of queried taxa
}
\details{
otol dated tree from Stephen Smith's otol scaling service.
# if you want to make an ltt plot of a dated OToL tree you'll need to get rid of singleton nodes with ape::collapse.singles
# and also probably do phytools::force.ultrametric
}
