% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{get_datelife_result}
\alias{get_datelife_result}
\title{Go from a vector of species, newick string, or phylo object to a list of patristic matrices}
\usage{
get_datelife_result(input = c("Rhea americana", "Pterocnemia pennata",
  "Struthio camelus"), partial = TRUE, usetnrs = FALSE,
  approximatematch = TRUE, update_cache = FALSE,
  cache = get("opentree_chronograms"), dating_method = "PATHd8",
  sppfromtaxon = FALSE, verbose = FALSE)
}
\arguments{
\item{input}{Target taxa names in the form of a vector of characters, a newick character string, or a phylo object.}

\item{partial}{If TRUE, use source chronograms even if they only match some of the desired taxa}

\item{usetnrs}{If TRUE, use OpenTree's services to resolve names. This can dramatically improve the chance of matches, but also take much longer.}

\item{approximatematch}{If TRUE, use a slower TNRS to correct mispellings, increasing the chance of matches (including false matches).}

\item{update_cache}{default to FALSE}

\item{cache}{The cached set of chronograms and other info from data(opentree_chronograms).}

\item{dating_method}{The method used for tree dating.}

\item{sppfromtaxon}{boolean vector, default to FALSE. If TRUE, will get all species names from taxon names given in input. Must have same length as input. If input is a newick string , with some clades it will be converted to phylo object phy, and the order of sppfromtaxon will match phy$tip.label.}

\item{verbose}{Boolean. If TRUE, it gives printed updates to the user.}
}
\value{
List of patristic matrices
}
\description{
Go from a vector of species, newick string, or phylo object to a list of patristic matrices
}
