% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_result.R
\name{get_datelife_result_datelifequery}
\alias{get_datelife_result_datelifequery}
\title{Get a list of patristic matrices from a given \code{datelifeQuery} object.}
\usage{
get_datelife_result_datelifequery(
  input = NULL,
  partial = TRUE,
  cache = "opentree_chronograms"
)
}
\arguments{
\item{input}{A \code{datelifeQuery} object from}
}
\value{
A datelifeResult object - a named list of patristic matrices. You can
access the original query with attributes(object)$query
}
\description{
Get a list of patristic matrices from a given \code{datelifeQuery} object.
}
