% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_result.R
\name{get_datelife_result_vector}
\alias{get_datelife_result_vector}
\title{Get a list of patristic matrices from a given set of taxon names.}
\usage{
get_datelife_result_vector(
  input = NULL,
  partial = TRUE,
  cache = "opentree_chronograms",
  ...
)
}
\arguments{
\item{input}{A character vector of taxon names}
}
\value{
A datelifeResult object - a named list of patristic matrices. You can
access the \code{datelifeQuery} object with attributes(object)$query
}
\description{
Get a list of patristic matrices from a given set of taxon names.
}
