% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_sdm.R
\name{get_goodmatrices}
\alias{get_goodmatrices}
\title{Get good matrices for SDM}
\usage{
get_goodmatrices(unpadded.matrices, verbose)
}
\arguments{
\item{unpadded.matrices}{A list of patristic matrices, a datelifeResult object.}

\item{verbose}{Boolean. If TRUE, it gives printed updates to the user.}
}
\value{
A numeric vector of good matrix indices in unpadded.matrices.
}
\description{
Get good matrices for SDM
}
