% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_calibrations.R
\name{get_otol_synthetic_tree}
\alias{get_otol_synthetic_tree}
\title{Gets Open Tree of Life synthetic tree of a set of lineages.}
\usage{
get_otol_synthetic_tree(input = NULL, ott_ids = NULL,
  otol_version = "v2", ...)
}
\arguments{
\item{input}{Optional. A character vector of names or a datelifeQuery object}

\item{ott_ids}{If not NULL, it takes this argument and ignores input. A numeric vector of ott ids obtained with rotl::taxonomy_taxon_info or rolt::tnrs_match_names or datelife::tnrs_match}

\item{otol_version}{Version of OToL to use}

\item{...}{Arguments passed on to \code{check_ott_input}
\describe{
  \item{input}{Optional. A character vector of names or a datelifeQuery object}
  \item{ott_ids}{If not NULL, it takes this argument and ignores input. A numeric vector of ott ids obtained with rotl::taxonomy_taxon_info or rolt::tnrs_match_names or datelife::tnrs_match}
}}
}
\value{
A phylo object
}
\description{
Gets Open Tree of Life synthetic tree of a set of lineages.
}
