% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_taxonomy.R
\name{get_ott_children}
\alias{get_ott_children}
\title{use this instead of rotl::tol_subtree when taxa are not in synthesis tree and you still need to get all species or an induced otol subtree}
\usage{
get_ott_children(input = c("Felis", "Homo", "Malvaceae"),
  ott_id = NULL, ott_rank = "species")
}
\arguments{
\item{input}{A character vector of names}

\item{ott_id}{A numeric vector of ott ids obtained with rotl::taxonomy_taxon_info or rolt::tnrs_match_names or datelife::tnrs_match}

\item{ott_rank}{A character vector with the ranks you wanna get lineage children from.}
}
\description{
use this instead of rotl::tol_subtree when taxa are not in synthesis tree and you still need to get all species or an induced otol subtree
}
\examples{
# try getting an otol tree of a taxon missing from the synthetic tree
# tnrs <- rotl::tnrs_match_names("Mus")
tnrs <- tnrs_match("Canis")
# Mus, Gleicheniales, Polypodiales, etc., are missing in synth tree too
\dontrun{
  rotl::tol_subtree(tnrs$ott_id[1])
  Error: HTTP failure: 400
  [/v3/tree_of_life/subtree] Error: node_id was not found (broken taxon).
}
children <- get_ott_children(ott_id = tnrs$ott_id[1]) # or
children <- get_ott_children("Canis")
rownames(children[[1]])
tree_children <- datelife::get_otol_synthetic_tree(children$Canis)
plot(tree_children, cex = 0.3)
# Other examples:
oo <- get_ott_children(input= "magnoliophyta", ott_rank = "order")
sum(oo$Magnoliophyta$rank == "order") # to know how many orders of flowering plants we have
}
