% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_taxonomy.R
\name{get_ott_clade}
\alias{get_ott_clade}
\title{gets the ott id and name of one or several given taxonomic rank from one or more input taxa}
\usage{
get_ott_clade(input = c("Felis", "Homo"), ott_ids = NULL,
  ott_rank = "family")
}
\arguments{
\item{input}{Optional. A character vector of names or a datelifeQuery object}

\item{ott_ids}{If not NULL, it takes this argument and ignores input. A numeric vector of ott ids obtained with rotl::taxonomy_taxon_info or rolt::tnrs_match_names or datelife::tnrs_match}

\item{ott_rank}{A character vector with the ranks you wanna get lineage children from.}
}
\value{
a list of named numeric vectors with ott ids from input and all requested ranks
}
\description{
gets the ott id and name of one or several given taxonomic rank from one or more input taxa
}
