% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_sdm.R
\name{get_sdm_matrix}
\alias{get_sdm_matrix}
\title{Get good matrices for SDM}
\usage{
get_sdm_matrix(datelife_result)
}
\arguments{
\item{datelife_result}{An output of get_datelife_result function: A list of patristic matrices with names corresponding to original study citations.}
}
\value{
A numeric matrix.
}
\description{
Get good matrices for SDM
}
