% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_chronograms.R
\name{get_study_tree_with_dups}
\alias{get_study_tree_with_dups}
\title{Allow trees with duplicate taxa to be read in, courtesy David Winter}
\usage{
get_study_tree_with_dups(study_id, tree_id, tip_label = "ot:otttaxonname")
}
\arguments{
\item{study_id}{Open Tree study id}

\item{tree_id}{Open Tree tree id}

\item{tip_label}{Which Open Tree tree label format you want}
}
\value{
A phylo object
}
\description{
Allow trees with duplicate taxa to be read in, courtesy David Winter
}

