% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{GetAge}
\alias{GetAge}
\title{Get time of MRCA from patristic matrix}
\usage{
GetAge(patristic.matrix, partial = TRUE)
}
\arguments{
\item{patristic.matrix}{A patristic matrix}

\item{partial}{If TRUE, drop NA from the patristic matrix; if FALSE, will return NA if there are missing entries}
}
\value{
The depth of the MRCA
}
\description{
Get time of MRCA from patristic matrix
}

