% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{GetAges}
\alias{GetAges}
\title{Get vector of MRCA from list of patristic matrices}
\usage{
GetAges(filtered.results, partial = TRUE)
}
\arguments{
\item{filtered.results}{List of patristic matrices}

\item{partial}{If TRUE, drop NA from the patristic matrix; if FALSE, will return NA if there are missing entries}
}
\value{
Vector of MRCA ages with names same as in filtered.results
}
\description{
Get vector of MRCA from list of patristic matrices
}

