% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{GetBladjTree}
\alias{GetBladjTree}
\title{Takes a tree and uses bladj to estimate node ages and branch lengths given a set of fixed node ages and respective node names}
\usage{
GetBladjTree(nodenames = NULL, nodeages = NULL, phy = NULL,
  phyformat = "newick")
}
\arguments{
\item{nodenames}{A character vector with node names from tree with fixed ages}

\item{nodeages}{A numeric vector with known or fixed node ages from tree}

\item{phy}{A tree either as a newick character string or phylo format}

\item{phyformat}{A character vector specifying tree output format, either "newick" (default) or "phylo"}
}
\value{
A newick or phylo tree with non negative branch lengths
}
\description{
Takes a tree and uses bladj to estimate node ages and branch lengths given a set of fixed node ages and respective node names
}
