% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{GetFilteredResults}
\alias{GetFilteredResults}
\title{Go from a vector of species, newick string, or phylo object to a list of patristic matrices}
\usage{
GetFilteredResults(input = c("Rhea americana", "Pterocnemia pennata",
  "Struthio camelus"), partial = TRUE, usetnrs = FALSE,
  approximatematch = TRUE, cache = get("opentree_chronograms"),
  method = "PATHd8", bold = FALSE, ...)
}
\arguments{
\item{input}{Target taxa names in the form of a vector of characters, a newick character string, or a phylo object.}

\item{partial}{If TRUE, use source chronograms even if they only match some of the desired taxa}

\item{usetnrs}{If TRUE, use OpenTree's services to resolve names. This can dramatically improve the chance of matches, but also take much longer.}

\item{approximatematch}{If TRUE, use a slower TNRS to correct mispellings, increasing the chance of matches (including false matches).}

\item{cache}{The cached set of chronograms and other info from data(opentree_chronograms).}

\item{method}{The method used for congruification. PATHd8 only right now, r8s and treePL later.}

\item{bold}{Logical. If TRUE, use Barcode of Life Data Systems (BOLD)  and Open Tree of Life (OToL) backbone to estimate branch lengths of target taxa using GetBoldOToLTree function.}

\item{...}{Arguments passed on to \code{GetBoldOToLTree}
\describe{
  \item{marker}{A character vector with the name of the gene from Barcode of Life Data Systems (BOLD) to be used for branch length estimation.}
  \item{otol_version}{Version of OToL to use}
  \item{chronogram}{Boolean; default to TRUE:  branch lengths represent time estimated with ape::chronoMPL. If FALSE, branch lengths represent relative substitution rates estimated with phangorn::acctran.}
  \item{doML}{Boolean; if TRUE, does ML branch length optimization with phangorn::optim.pml}
  \item{input}{Target taxa names in the form of a vector of characters, a newick character string, or a phylo object.}
  \item{usetnrs}{If TRUE, use OpenTree's services to resolve names. This can dramatically improve the chance of matches, but also take much longer.}
  \item{approximatematch}{If TRUE, use a slower TNRS to correct mispellings, increasing the chance of matches (including false matches).}
}}
}
\value{
List of patristic matrices
}
\description{
Go from a vector of species, newick string, or phylo object to a list of patristic matrices
}
