% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{GetMrBayesTree}
\alias{GetMrBayesTree}
\title{Takes a constraint tree and uses mrBayes to get node ages and branch lengths given a set of node calibrations}
\usage{
GetMrBayesTree(phy = NULL, ncalibration = NULL,
  file = "mrbayes_run.nexus")
}
\arguments{
\item{phy}{The constraint tree: a phylo object or a newick character string, with or without branch lengths.}

\item{ncalibration}{The node calibrations: a phylo object with branch lengths proportional to time; in this case all nodes from ncalibration will be used as calibration points. Alternatively, a list with two elements: the first is a character vector with node names from phy to calibrate; the second is a numeric vector with the corresponding ages to use as calibrations.}

\item{file}{A character vector specifying the name of mrBayes run file and outputs (can specify directory too).}
}
\value{
A phylo tree with branch lengths proportional to time. It will save all mrBayes outputs in the working directory.
}
\description{
Takes a constraint tree and uses mrBayes to get node ages and branch lengths given a set of node calibrations
}
