% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{GetSubsetArrayDispatch}
\alias{GetSubsetArrayDispatch}
\title{Figure out which subset function to use}
\usage{
GetSubsetArrayDispatch(study.element, taxa, phy = NULL, phy4 = NULL,
  dating_method = "PATHd8")
}
\arguments{
\item{study.element}{The thing being passed in: an array or a phylo to serve as reference}

\item{taxa}{Vector of taxon names to get a subset for}

\item{phy}{A user tree to congruify in phylo format (ape)}

\item{phy4}{A user tree to congruify in phylo4 format (phylobase)}

\item{dating_method}{The method used for tree dating.}
}
\value{
A patristic matrix with for the taxa.
}
\description{
Figure out which subset function to use
}
