% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{GetTips}
\alias{GetTips}
\title{To get tip numbers descending from any given node of a tree}
\usage{
GetTips(tree = NULL, node = NULL, curr = NULL)
}
\arguments{
\item{tree}{a phylogenetic tree as an object of class \code{"phylo"}.}

\item{node}{an integer specifying a node number in the tree.}

\item{curr}{the set of previously stored node numbers - used in recursive function calls.}
}
\value{
A numeric vector with tip numbers descending from a node
}
\description{
To get tip numbers descending from any given node of a tree
}
