% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_chronograms.R
\name{phylo_has_brlen}
\alias{phylo_has_brlen}
\title{Check for branch lengths in a tree}
\usage{
phylo_has_brlen(x)
}
\arguments{
\item{x}{A phylo object}
}
\value{
A TRUE or FALSE
}
\description{
Check for branch lengths in a tree
}
