% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloch_functions.R
\name{HPDbars}
\alias{HPDbars}
\title{Function to add interval bars to nodes on a phylo plot.}
\usage{
HPDbars(phy, label = "height_95\%_HPD", tab = NULL, nodes, col, lwd,
  broken = FALSE, ...)
}
\arguments{
\item{phy}{A phylo object with two extra elements specifying the MAX and MIN limits of intervals to plot.}

\item{label}{A character vector corresponding to the name used to label intervals stored in phy}

\item{tab}{Something I'm not sure of, should ask Christoph Heibl.}

\item{nodes}{Numeric. Use this if you wanna plot bars on certain nodes only.}

\item{col}{Color for the bars, default to "grey75"}

\item{lwd}{Line width for the bars, default to 5.}

\item{broken}{Something I'm not sure of, should ask Christoph Heibl.}

\item{...}{Arguments passed on to \code{graphics::segments}
\describe{
  \item{x0}{coordinates of points \bold{from} which to draw.}
  \item{y0}{coordinates of points \bold{from} which to draw.}
  \item{x1}{coordinates of points \bold{to} which to draw.  At least
    one must be supplied.}
  \item{y1}{coordinates of points \bold{to} which to draw.  At least
    one must be supplied.}
  \item{col}{\link{graphical parameters} as in
    \code{\link{par}}, possibly vectors.  \code{NA} values in \code{col}
    cause the segment to be omitted.}
  \item{lty}{\link{graphical parameters} as in
    \code{\link{par}}, possibly vectors.  \code{NA} values in \code{col}
    cause the segment to be omitted.}
  \item{lwd}{\link{graphical parameters} as in
    \code{\link{par}}, possibly vectors.  \code{NA} values in \code{col}
    cause the segment to be omitted.}
}}
}
\value{
Adds bars to nodes on a phylo plot.
}
\description{
Function to add interval bars to nodes on a phylo plot.
}
