% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{IdentifySingletonOutgroup}
\alias{IdentifySingletonOutgroup}
\title{Identifies the presence of a single lineage outgroup in a phylogeny.}
\usage{
IdentifySingletonOutgroup(phy = NULL)
}
\arguments{
\item{phy}{The constraint tree: a phylo object or a newick character string, with or without branch lengths.}
}
\value{
A character vector with the name of the single lineage outgroup. Returns NA if there is none.
}
\description{
Identifies the presence of a single lineage outgroup in a phylogeny.
}
