% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_calibrations.R
\name{input_tnrs}
\alias{input_tnrs}
\title{Taxon name resolution service applied to input by batches}
\usage{
input_tnrs(input, reference_taxonomy = "otl")
}
\arguments{
\item{input}{Target taxa names in the form of a vector of characters, a newick character string, or a phylo object.}

\item{reference_taxonomy}{A character vetor specifying the reference taxonomy to use for tnrs.}
}
\value{
A data frame from tnrs_match_names function
}
\description{
Taxon name resolution service applied to input by batches
}
