% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_query.R
\name{is_datelife_query}
\alias{is_datelife_query}
\title{checks if input is formatted as a datelifeQuery object}
\usage{
is_datelife_query(input)
}
\arguments{
\item{input}{Any object to be tested for 'datelifeQuery' class format}
}
\value{
boolean
}
\description{
checks if input is formatted as a datelifeQuery object
}
