% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_search.R
\name{is_datelife_result_empty}
\alias{is_datelife_result_empty}
\title{Check if we obtained an empty search with the set of input taxon names}
\usage{
is_datelife_result_empty(datelife_result, use_tnrs = FALSE)
}
\arguments{
\item{datelife_result}{A \code{datelifeResult} object, output of \code{\link[=get_datelife_result]{get_datelife_result()}}.}

\item{use_tnrs}{Whether to use OpenTree's Taxonomic Name Resolution Service (TNRS)
to process input taxon names. Default to \code{TRUE}, it corrects misspellings and
taxonomic name variations.}
}
\description{
Check if we obtained an empty search with the set of input taxon names
}
