% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_chronograms.R
\name{is_good_chronogram}
\alias{is_good_chronogram}
\title{Check if a tree is a valid chronogram.}
\usage{
is_good_chronogram(phy)
}
\arguments{
\item{phy}{A \code{phylo} object.}
}
\value{
\code{TRUE} if it is a valid tree.
}
\description{
Check if a tree is a valid chronogram.
}
