% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groves.R
\name{is_n_overlap}
\alias{is_n_overlap}
\title{Function for computing n overlap for two vectors of names (ie., phy1$tip.label, phy2$tip.label) and seeing if they have n overlap}
\usage{
is_n_overlap(names_1, names_2, n = 2)
}
\arguments{
\item{names_1}{First vector of names}

\item{names_2}{Second vector of names}

\item{n}{Degree of overlap required}
}
\value{
Boolean for whether the degree of overlap was met
}
\description{
n-overlap comes from Definition 2.8 of Ané et al. 10.1007/s00026-009-0017-x Groves of Phylogenetic Trees
}
