% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_chronograms.R
\name{IsGoodChronogram}
\alias{IsGoodChronogram}
\title{Check to see that a chronogram is valid}
\usage{
IsGoodChronogram(phy)
}
\arguments{
\item{phy}{Input phylo object}
}
\value{
Boolean: TRUE if good tree
}
\description{
Check to see that a chronogram is valid
}
