% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_ltt_summary.R
\name{ltt_summary}
\alias{ltt_summary}
\title{Lineage through time plots of  all chronograms in multiphylo object and its summaries.}
\usage{
ltt_summary(phy_summ, phy_summ_type = NULL, phy_summ_col = NULL,
  max_tips, length_arrowhead = 0.075, ...)
}
\arguments{
\item{phy_summ}{A multiphylo object with chronograms from a summary.}

\item{phy_summ_type}{A character value.}

\item{phy_summ_col}{A character value.}

\item{max_tips}{A numeric value}

\item{length_arrowhead}{A numeric value}

\item{...}{Arguments passed on to \code{ape::ltt.plot}
\describe{
  \item{phy}{an object of class \code{"phylo"}; this could be an object
    of class \code{"multiPhylo"} in the case of \code{mltt.plot}.}
  \item{xlab}{a character string (or a variable of mode character)
    giving the label for the \eqn{x}-axis (default is "Time").}
  \item{ylab}{idem for the \eqn{y}-axis (default is "N").}
  \item{backward}{a logical value: should the time axis be traced from
    the present (the default), or from the root of the tree?}
  \item{tol}{a numeric value (see details).}
}}
}
\description{
Lineage through time plots of  all chronograms in multiphylo object and its summaries.
}
