% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bladj_tree.R
\name{make_bladj_tree}
\alias{make_bladj_tree}
\title{Takes a tree and uses bladj to estimate node ages and branch lengths given a
set of fixed node ages and respective node names}
\usage{
make_bladj_tree(tree = NULL, nodenames = NULL, nodeages = NULL)
}
\arguments{
\item{tree}{A tree either as a newick character string or as a phylo object}

\item{nodenames}{A character vector with names of nodes in tree with known ages}

\item{nodeages}{A numeric vector with the actual ages of named nodes}
}
\value{
A phylo tree
}
\description{
Takes a tree and uses bladj to estimate node ages and branch lengths given a
set of fixed node ages and respective node names
}
\details{
tree can or cannot be dated, edge.lengths in it will be ignored anyways;
only ages from nodeages will be fixed according to nodenames.
}
