% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_tree.R
\name{make_bold_otol_tree}
\alias{make_bold_otol_tree}
\title{Use Barcode of Life data to get branch lengths on the OToL tree of a set of taxa.}
\usage{
make_bold_otol_tree(input = c("Rhea americana", "Struthio camelus",
  "Gallus gallus"), use_tnrs = FALSE, approximate_match = TRUE,
  marker = "COI", otol_version = "v3", chronogram = TRUE,
  doML = FALSE, get_spp_from_taxon = FALSE, verbose = FALSE)
}
\arguments{
\item{input}{Target taxa names as a character vector, a newick character string, or a phylo object.}

\item{use_tnrs}{If TRUE, use OpenTree's services to resolve names. This can dramatically improve the chance of matches, but also take much longer.}

\item{approximate_match}{If TRUE, use a slower TNRS to correct mispellings, increasing the chance of matches (including false matches).}

\item{marker}{A character vector with the name of the gene from Barcode of Life Data Systems (BOLD) to be used for branch length estimation.}

\item{otol_version}{Version of OToL to use}

\item{chronogram}{Boolean; default to TRUE:  branch lengths represent time estimated with ape::chronoMPL. If FALSE, branch lengths represent relative substitution rates estimated with phangorn::acctran.}

\item{doML}{Boolean; if TRUE, does ML branch length optimization with phangorn::optim.pml}

\item{get_spp_from_taxon}{boolean vector, default to FALSE. If TRUE, will get all species names from taxon names given in input. Must have same length as input. If input is a newick string , with some clades it will be converted to phylo object phy, and the order of get_spp_from_taxon will match phy$tip.label.}

\item{verbose}{Boolean. If TRUE, it gives printed updates to the user.}
}
\value{
A phylogeny with branch lengths proportional to relative substitution rate.
}
\description{
Use Barcode of Life data to get branch lengths on the OToL tree of a set of taxa.
}
\details{
If input is a phylo object, it is used as backbone. If it is a character vector of taxon names, an induced OToL tree is used as backbone.
}
