% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheauthors.R
\name{make_contributor_cache}
\alias{make_contributor_cache}
\title{Create a cache from Open Tree of Life}
\usage{
make_contributor_cache(outputfile = "contributorcache.RData")
}
\arguments{
\item{outputfile}{Path including file name}
}
\value{
List containing author and curator results
}
\description{
Create a cache from Open Tree of Life
}
