% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_query_make.R
\name{make_datelife_query_vector}
\alias{make_datelife_query_vector}
\title{Prepare a set of given taxon names as a \code{datelifeQuery} object}
\usage{
make_datelife_query_vector(
  input = NULL,
  use_tnrs = FALSE,
  approximate_match = TRUE,
  get_spp_from_taxon = FALSE
)
}
\arguments{
\item{input}{A character vector of taxon names.}

\item{use_tnrs}{Boolean; default to \code{FALSE}. If \code{TRUE}, use OpenTree's services
to resolve names. This can dramatically improve the chance of matches, but also
take longer.}

\item{approximate_match}{Boolean; default to \code{TRUE}: use a slower TNRS to correct
misspellings, increasing the chance of matches (including false matches)}

\item{get_spp_from_taxon}{boolean vector, default to \code{FALSE}. If \code{TRUE},
it will get all species within \code{input} taxon names. It can be selectively
applied, in which case it must have the same length as \code{input} taxon names.
It will be applied in the same order. If \code{input} is a newick string, it will
is converted to a \code{phylo} object, and \code{get_spp_from_taxon} is applied
following the order of \code{phylo$tip.label}.}
}
\description{
Prepare a set of given taxon names as a \code{datelifeQuery} object
}
