% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheauthors.R
\name{make_otol_associations}
\alias{make_otol_associations}
\title{Associate Open Tree of Life authors with studies}
\usage{
make_otol_associations()
}
\value{
\code{data.frame} with author last name, author first and other names, and comma delimited URLs for OToL studies
}
\description{
Associate Open Tree of Life authors with studies
}
