% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheauthors.R
\name{make_overlap_table}
\alias{make_overlap_table}
\title{Create an overlap table}
\usage{
make_overlap_table(results.table)
}
\arguments{
\item{results.table}{An author.results or curator.results data.frame}
}
\value{
Data.frame with info on people and what clades they've worked on
}
\description{
Create an overlap table
}
