% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheauthors.R
\name{make_treebase_associations}
\alias{make_treebase_associations}
\title{Associate TreeBase authors with studies}
\usage{
make_treebase_associations()
}
\value{
\code{data.frame} with author last name, author first and other names, and comma delimited URLs for TreeBase studies
}
\description{
Associate TreeBase authors with studies
}
