% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheauthors.R
\name{make_treebase_cache}
\alias{make_treebase_cache}
\title{Create a cache from TreeBase}
\usage{
make_treebase_cache(outputfile = "treebasecache.RData")
}
\arguments{
\item{outputfile}{Path including file name}
}
\value{
List containing author and curator results
}
\description{
Create a cache from TreeBase
}
