% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{MakeMrBayesRunFile}
\alias{MakeMrBayesRunFile}
\title{Makes a mrBayes run block file with a constraint topology and a set of node calibrations}
\usage{
MakeMrBayesRunFile(phy = NULL, ncalibration = NULL,
  file = "mrbayes_run.nexus")
}
\arguments{
\item{phy}{The constraint tree: a phylo object or a newick character string, with or without branch lengths.}

\item{ncalibration}{The node calibrations: a phylo object with branch lengths proportional to time; in this case all nodes from ncalibration will be used as calibration points. Alternatively, a list with two elements: the first is a character vector with node names from phy to calibrate; the second is a numeric vector with the corresponding ages to use as calibrations.}

\item{file}{A character vector specifying the name of mrBayes run file and outputs (can specify directory too).}
}
\value{
A MrBayes block run file in nexus format.
}
\description{
Makes a mrBayes run block file with a constraint topology and a set of node calibrations
}
