% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_calibrations_map.R
\name{map_all_calibrations}
\alias{map_all_calibrations}
\title{Map calibrations to nodes of a tree topology}
\usage{
map_all_calibrations(phy, calibrations)
}
\arguments{
\item{phy}{A phylo object to map calibrations to its nodes}

\item{calibrations}{A data frame of calibrations from get_all_calibrations function}
}
\value{
A list with two elements
\describe{
\item{phy}{A phylo object with a list of calibration distributions per node.}
\item{calibrations}{A data frame of summarized calibrations.}
}
}
\description{
Map calibrations to nodes of a tree topology
}
\details{
If input is a phylo object, it is used as backbone. If it is a character vector of taxon names, an induced OToL tree is used as backbone.
}
