% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_chronograms.R
\name{map_nodes_ott}
\alias{map_nodes_ott}
\title{Map opentree_chronograms tip_label's Open Tree of Life Taxonomy to nodes.}
\usage{
map_nodes_ott(tree)
}
\arguments{
\item{tree}{A tree either as a newick character string or as a phylo object}
}
\value{
A cleaned up phylo object
}
\description{
Map opentree_chronograms tip_label's Open Tree of Life Taxonomy to nodes.
}
