% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_calibrations_map.R
\name{match_all_calibrations}
\alias{match_all_calibrations}
\title{Map calibrations to nodes of a tree topology}
\usage{
match_all_calibrations(phy, calibrations)
}
\arguments{
\item{phy}{A phylo object to map calibrations to its nodes}

\item{calibrations}{A data frame of calibrations from taxon pairs. Usually an output from get_all_calibrations function (see this function for format)}
}
\value{
A list with two elements
\describe{
\item{phy}{A phylo object with a list of calibration distributions per node.}
\item{calibrations}{A data frame of summarized calibrations.}
}
}
\description{
Map calibrations to nodes of a tree topology
}
\details{
The function will check if taxon pairs in calibrations are present in phy.
}
