% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_table.R
\name{matrix_to_table}
\alias{matrix_to_table}
\title{Go from a patristic distance matrix to a node ages table}
\usage{
matrix_to_table(matrix, reference)
}
\arguments{
\item{matrix}{A patristic distance matrix.}

\item{reference}{A character vector with the study reference from where the ages come from.}
}
\value{
A \code{data.frame} of "taxonA", "taxonB", and "age".
}
\description{
Go from a patristic distance matrix to a node ages table
}
