% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{message_multiphylo}
\alias{message_multiphylo}
\title{Message for \code{multiPhylo} input}
\usage{
message_multiphylo()
}
\value{
A relevant message as a character string.
}
\description{
Message for \code{multiPhylo} input
}
