% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{missed_doi}
\alias{missed_doi}
\title{Study ids of chronograms in Open Tree of Life not included in authors.}
\format{A character vector with study ids}
\source{
\url{http://opentreeoflife.org}
}
\usage{
missed_doi
}
\description{
There are different reasons why a doi cannot be found online, so author data
cannot be retrieved for that particular study.
This vector contains all the studies with chrongrams in OToL whose information
was not included automatically in author.pretty and author.results
}
\details{
Generated with make_contributor_cache()
}
\keyword{chronogram}
\keyword{id}
\keyword{otol}
\keyword{studies}
\keyword{study}
\keyword{tree}
