% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{missing_taxa_check}
\alias{missing_taxa_check}
\title{Checks that missing_taxa argument is ok to be used by make_mrbayes_runfile function}
\usage{
missing_taxa_check(missing_taxa = NULL, dated_phy = NULL)
}
\arguments{
\item{missing_taxa}{either a tree (as phylo object or as a newick character string, with or without branch lengths) with all the taxa you want at the end, or a data frame assigning missing taxa to nodes in dated tree.}

\item{dated_phy}{a tree (newick or phylo) with branch lengths proportional to absolute time}
}
\value{
A phylo object, a newick character string or a dataframe with taxonomic assignations
}
\description{
Checks that missing_taxa argument is ok to be used by make_mrbayes_runfile function
}
