% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_subset2}
\alias{names_subset2}
\title{Long list of >2.7k virus, bacteria, plant and animal taxon names}
\format{A character vector of length 2778}
\source{
\url{https://github.com/phylotastic/rphylotastic/tree/master/tests/testthat}
}
\usage{
names_subset2
}
\description{
Long list of >2.7k virus, bacteria, plant and animal taxon names
}
\details{
Generated with:
names_subset2 <- rphylotastic::url_get_scientific_names("https://github.com/phylotastic/rphylotastic/blob/master/tests/testthat/subset2.txt")
}
\keyword{animal}
\keyword{bacteria}
\keyword{names}
\keyword{plant}
\keyword{subset}
\keyword{taxon}
\keyword{virus}
