% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.r
\name{nonmatching_by_row}
\alias{nonmatching_by_row}
\title{Find only rows in a data.frame where the values don't match in a row}
\usage{
nonmatching_by_row(df)
}
\arguments{
\item{df}{A data.frame}
}
\description{
Find only rows in a data.frame where the values don't match in a row
}
\examples{
dff <- data.frame(FabreEtAl2009$tip.label, out$tip.label)
nonmatching_by_row(dff)
}

