% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/opentree_chronograms.R
\docType{data}
\name{opentree_chronograms}
\alias{opentree_chronograms}
\alias{opentree_chronograms}
\title{Cached chronograms and other data from OpenTree}
\format{A list with four lists contained within it
\describe{
\item{authors}{List of lists of authors for the included studies}
\item{curators}{List of lists of curators for the included studies}
\item{studies}{List of study identifiers}
\item{trees}{List storing the chronograms from OpenTree}
}}
\source{
\url{http://opentreeoflife.org}

\url{https://tree.opentreeoflife.org}
}
\usage{
opentree_chronograms
}
\description{
A list with four elements for >100 chronograms from OpenTree

Open Tree of Life chronograms data set
}
\references{
\url{https://tree.opentreeoflife.org}
}
\author{
Luna L. Sánchez-Reyes \email{lsanche7@utk.edu} Brian O'Meara \email{bomeara@utk.edu}
}
\keyword{chronogram}
\keyword{datasets}
\keyword{open}
\keyword{otol}
\keyword{tree}
