% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{patristic_matrix_array_congruify}
\alias{patristic_matrix_array_congruify}
\title{\code{patristic_matrix_array_congruify} is used for patristic_matrix_array_subset_both and patristic_matrix_array_congruify.}
\usage{
patristic_matrix_array_congruify(
  patristic_matrix_array,
  taxa,
  phy = NULL,
  dating_method = "PATHd8"
)
}
\arguments{
\item{patristic_matrix_array}{A patristic matrix array, \code{rownames} and \code{colnames} must be taxa.}

\item{taxa}{Vector of taxon names to get a subset for.}

\item{phy}{A user tree to congruify as \code{phylo} object (ape).}

\item{dating_method}{The method used for tree dating.}
}
\value{
A patristic matrix with ages for the target taxa.
}
\description{
\code{patristic_matrix_array_congruify} is used for patristic_matrix_array_subset_both and patristic_matrix_array_congruify.
}
