% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{patristic_matrix_array_split}
\alias{patristic_matrix_array_split}
\title{Split a patristic matrix array
Used inside: patristic_matrix_array_congruify}
\usage{
patristic_matrix_array_split(patristic_matrix_array)
}
\arguments{
\item{patristic_matrix_array}{A patristic matrix array, \code{rownames} and \code{colnames} must be taxa.}
}
\value{
A patristic matrix 3d array.
}
\description{
Split a patristic matrix array
Used inside: patristic_matrix_array_congruify
}
