% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{patristic_matrix_array_subset}
\alias{patristic_matrix_array_subset}
\title{Subset a patristic matrix array}
\usage{
patristic_matrix_array_subset(patristic_matrix_array, taxa, phy4 = NULL)
}
\arguments{
\item{patristic_matrix_array}{A patristic matrix array, \code{rownames} and \code{colnames} must be taxa.}

\item{taxa}{Vector of taxon names to get a subset for.}

\item{phy4}{A user tree to congruify in \code{phylo4} format (phylobase).}
}
\value{
A list with a patristic matrix array and a \verb{$problem} if any.
}
\description{
Subset a patristic matrix array
}
