% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{patristic_matrix_list_to_array}
\alias{patristic_matrix_list_to_array}
\title{Convert list of patristic matrices to a 3D array.}
\usage{
patristic_matrix_list_to_array(patristic_matrix_list, pad = TRUE)
}
\arguments{
\item{patristic_matrix_list}{List of patristic matrices}

\item{pad}{If TRUE, pad missing entries}
}
\value{
A 3d array of patristic matrices
}
\description{
\code{patristic_matrix_list_to_array} us ised inside \code{\link[=summarize_datelife_result]{summarize_datelife_result()}}, \code{\link[=patristic_matrix_array_congruify]{patristic_matrix_array_congruify()}}.
}
