% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{patristic_matrix_name_order_test}
\alias{patristic_matrix_name_order_test}
\title{Test the name order of a patristic matrix so that row and column labels are in alphabetical order.}
\usage{
patristic_matrix_name_order_test(
  patristic_matrix,
  standard.rownames,
  standard.colnames
)
}
\arguments{
\item{patristic_matrix}{A patristic matrix, \code{rownames} and \code{colnames} must be taxa.}

\item{standard.rownames}{A character vector of row names.}

\item{standard.colnames}{A character vector of column names.}
}
\value{
Boolean.
}
\description{
\code{patristic_matrix_name_order_test} is only used in \code{\link[=patristic_matrix_list_to_array]{patristic_matrix_list_to_array()}}.
}
