% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{patristic_matrix_name_reorder}
\alias{patristic_matrix_name_reorder}
\title{Reorder a matrix so that row and column labels are in alphabetical order.}
\usage{
patristic_matrix_name_reorder(patristic_matrix)
}
\arguments{
\item{patristic_matrix}{A patristic matrix, \code{rownames} and \code{colnames} must be taxa.}
}
\value{
A patristic matrix with row and column names for taxa in alphabetical order.
}
\description{
\code{patristic_matrix_name_reorder} is only used in: \code{\link[=patristic_matrix_pad]{patristic_matrix_pad()}}.
}
