% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{patristic_matrix_pad}
\alias{patristic_matrix_pad}
\title{Fill in empty cells in a patristic matrix for missing taxa.}
\usage{
patristic_matrix_pad(patristic_matrix, all_taxa)
}
\arguments{
\item{patristic_matrix}{A patristic matrix, \code{rownames} and \code{colnames} must be taxa.}

\item{all_taxa}{A vector of names of all taxa you want, including ones not
in the patristic matrix.}
}
\value{
A patristic matrix, with \code{NA} for entries between taxa
where at least one was not in the original patristic matrix.
}
\description{
Used in: \code{\link[=patristic_matrix_list_to_array]{patristic_matrix_list_to_array()}}.
}
