% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{patristic_matrix_taxa_all_matching}
\alias{patristic_matrix_taxa_all_matching}
\title{Are all desired taxa in the patristic matrix?}
\usage{
patristic_matrix_taxa_all_matching(patristic_matrix, taxa)
}
\arguments{
\item{patristic_matrix}{A patristic matrix, \code{rownames} and \code{colnames} must be taxa.}

\item{taxa}{Vector of taxon names to get a subset for.}
}
\value{
A Boolean.
}
\description{
\code{patristic_matrix_taxa_all_matching} is used inside: \code{\link[=results_list_process]{results_list_process()}}.
}
