% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_newick.R
\name{patristic_matrix_to_newick}
\alias{patristic_matrix_to_newick}
\title{Convert patristic matrix to a newick string. Used inside: summarize_datelife_result.}
\usage{
patristic_matrix_to_newick(patristic_matrix)
}
\arguments{
\item{patristic_matrix}{A patristic matrix}
}
\value{
A newick string
}
\description{
Convert patristic matrix to a newick string. Used inside: summarize_datelife_result.
}
