% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{patristic_matrix_to_phylo}
\alias{patristic_matrix_to_phylo}
\title{Convert patristic matrix to a phylo object. Used inside: summarize_datelife_result, CongruiyTree.}
\usage{
patristic_matrix_to_phylo(patristic_matrix)
}
\arguments{
\item{patristic_matrix}{A patristic matrix}
}
\value{
A rooted phylo object
}
\description{
Convert patristic matrix to a phylo object. Used inside: summarize_datelife_result, CongruiyTree.
}
