% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{patristic_matrix_unpad}
\alias{patristic_matrix_unpad}
\title{Function to remove missing taxa from a datelifeResult object. Used in: datelife_result_sdm.}
\usage{
patristic_matrix_unpad(patristic_matrix)
}
\arguments{
\item{patristic_matrix}{A patristic matrix with row and column names for taxa}
}
\value{
patristic_matrix for all_taxa
}
\description{
Function to remove missing taxa from a datelifeResult object. Used in: datelife_result_sdm.
}
