% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{PatristicMatrixToNewick}
\alias{PatristicMatrixToNewick}
\title{Convert patristic matrix to a newick string}
\usage{
PatristicMatrixToNewick(patristic.matrix)
}
\arguments{
\item{patristic.matrix}{A patristic matrix}
}
\value{
A newick string
}
\description{
Convert patristic matrix to a newick string
}
