% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{PatristicMatrixToTree}
\alias{PatristicMatrixToTree}
\title{Convert patristic matrix to a phylo object. Used inside: summarize_datelife_result, CongruiyTree.}
\usage{
PatristicMatrixToTree(patristic.matrix)
}
\arguments{
\item{patristic.matrix}{A patristic matrix}
}
\value{
A rooted phylo object
}
\description{
Convert patristic matrix to a phylo object. Used inside: summarize_datelife_result, CongruiyTree.
}
