% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{phy_process}
\alias{phy_process}
\title{Process phy with input_process and checks if it is a phylo class object and/or a chronogram with phylo_check}
\usage{
phy_process(phy = NULL, ...)
}
\arguments{
\item{phy}{A tree either as a newick character string or as a phylo object}

\item{...}{Arguments passed on to \code{phylo_check}
\describe{
  \item{dated}{Boolean. If TRUE it checks if phylo object has branch lengths and is ultrametric.}
  \item{phy}{A tree either as a newick character string or as a phylo object}
}}
}
\value{
If phy is correctly formatted, it returns a phylo object
}
\description{
Process phy with input_process and checks if it is a phylo class object and/or a chronogram with phylo_check
}
