% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{phylo_add_node_labels}
\alias{phylo_add_node_labels}
\title{Adds labels to nodes with no asigned label}
\usage{
phylo_add_node_labels(phy = NULL, node_prefix = "n",
  node_index = "node_number")
}
\arguments{
\item{phy}{A tree either as a newick character string or as a phylo object}

\item{node_prefix}{Character vector. If length 1, it will be used to name all nodes with no labels, followed by a number which can be the node_number or consecutive, as specified in node_number}

\item{node_index}{Character vector choosing one of "consecutive" or "node_number" as node label index. It will use consecutive numbers from 1 to total node number in the first case and node numbers in the second case.}
}
\value{
A phylo object
}
\description{
Adds labels to nodes with no asigned label
}
