% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{phylo_add_outgroup}
\alias{phylo_add_outgroup}
\title{Function to add an outgroup to any phylogeny, in phylo or newick format}
\usage{
phylo_add_outgroup(phy = NULL, outgroup = "outgroup")
}
\arguments{
\item{phy}{A tree either as a newick character string or as a phylo object}

\item{outgroup}{A character vector with the name of the outgroup. If it has length>1, only first element will be used.}
}
\value{
A phylo object.
}
\description{
Function to add an outgroup to any phylogeny, in phylo or newick format
}
