% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{phylo_congruify}
\alias{phylo_congruify}
\title{Congruify a reference tree and a target tree given as \code{phylo} objects.}
\usage{
phylo_congruify(
  reference_tree,
  target_tree,
  dating_method = "PATHd8",
  attempt_fix = TRUE
)
}
\arguments{
\item{reference_tree}{A \code{phylo} object.}

\item{target_tree}{A \code{phylo} object. Use this in case you want a specific
backbone for the output tree.}

\item{dating_method}{The method used for tree dating.}

\item{attempt_fix}{Default to \code{TRUE}. If congruification results in NA branch
lengths, it will attempt to fix them.}
}
\value{
A matrix.
}
\description{
Congruify a reference tree and a target tree given as \code{phylo} objects.
}
