% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{phylo_fix_brlen}
\alias{phylo_fix_brlen}
\title{Takes a tree and fixes negative branch lengths in several ways}
\usage{
phylo_fix_brlen(phy = NULL, fixing_criterion = "negative",
  fixing_method = 0)
}
\arguments{
\item{phy}{A tree either as a newick character string or as a phylo object}

\item{fixing_criterion}{A character vector specifying the type of branches to be fixed: "negative" or "zero"}

\item{fixing_method}{A character vector specifying the method to fix negative branch lengths: "bladj", "mrbayes" or a number to assign to all branches}
}
\value{
A phylo object with no negative branch lengths
}
\description{
Takes a tree and fixes negative branch lengths in several ways
}
