% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_processing.R
\name{phylo_get_node_numbers}
\alias{phylo_get_node_numbers}
\title{Gets node numbers from any phylogeny}
\usage{
phylo_get_node_numbers(phy)
}
\arguments{
\item{phy}{A \code{phylo} object.}
}
\value{
A numeric vector with node numbers
}
\description{
Gets node numbers from any phylogeny
}
