% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{phylo_get_singleton_outgroup}
\alias{phylo_get_singleton_outgroup}
\title{Identifies the presence of a single lineage outgroup in a phylogeny.}
\usage{
phylo_get_singleton_outgroup(phy = NULL)
}
\arguments{
\item{phy}{A tree either as a newick character string or as a phylo object}
}
\value{
A character vector with the name of the single lineage outgroup. Returns NA if there is none.
}
\description{
Identifies the presence of a single lineage outgroup in a phylogeny.
}
