% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{phylo_get_subset_array}
\alias{phylo_get_subset_array}
\title{Get a subset array from a \code{phylo} object}
\usage{
phylo_get_subset_array(
  reference_tree,
  taxa,
  phy4 = NULL,
  dating_method = "PATHd8"
)
}
\arguments{
\item{reference_tree}{A \code{phylo} object.}

\item{taxa}{Vector of taxon names to get a subset for.}

\item{phy4}{A user tree to congruify in \code{phylo4} format (phylobase).}

\item{dating_method}{The method used for tree dating.}
}
\value{
A list with a patristic matrix array and a \verb{$problem} if any.
}
\description{
Get a subset array from a \code{phylo} object
}
