% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{phylo_get_subset_array_congruify}
\alias{phylo_get_subset_array_congruify}
\title{Get a congruified subset array from a \code{phylo} object}
\usage{
phylo_get_subset_array_congruify(
  reference_tree,
  taxa,
  phy = NULL,
  dating_method = "PATHd8"
)
}
\arguments{
\item{reference_tree}{A \code{phylo} object.}

\item{taxa}{Vector of taxon names to get a subset for.}

\item{phy}{A user tree to congruify as \code{phylo} object (ape).}

\item{dating_method}{The method used for tree dating.}
}
\value{
A list with a patristic matrix array and a \verb{$problem} if any.
}
\description{
Get a congruified subset array from a \code{phylo} object
}
