% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_chronograms.R
\name{phylo_has_brlen}
\alias{phylo_has_brlen}
\title{Check if a tree has branch lengths}
\usage{
phylo_has_brlen(phy)
}
\arguments{
\item{phy}{A phylo object}
}
\value{
A TRUE or FALSE
}
\description{
Check if a tree has branch lengths
}
