% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{phylo_height_omi}
\alias{phylo_height_omi}
\title{get the outer margin of a graphics device from a number of tips}
\usage{
phylo_height_omi(tree)
}
\arguments{
\item{tree}{phylo object to be plotted}
}
\description{
get the outer margin of a graphics device from a number of tips
}
