% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{phylo_prune_missing_taxa}
\alias{phylo_prune_missing_taxa}
\title{Prune missing taxa from a \code{phylo} object
Used inside phylo_get_subset_array and phylo_get_subset_array_congruify.}
\usage{
phylo_prune_missing_taxa(phy, taxa)
}
\arguments{
\item{phy}{A user tree to congruify as \code{phylo} object (ape).}

\item{taxa}{Vector of taxon names to get a subset for.}
}
\value{
A \code{phylo} object.
}
\description{
Prune missing taxa from a \code{phylo} object
Used inside phylo_get_subset_array and phylo_get_subset_array_congruify.
}
