% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{phylo_subset_both}
\alias{phylo_subset_both}
\title{Subset a reference and a target tree given as \code{phylo} objects.}
\usage{
phylo_subset_both(
  reference_tree,
  taxa,
  phy = NULL,
  phy4 = NULL,
  dating_method = "PATHd8"
)
}
\arguments{
\item{reference_tree}{A \code{phylo} object.}

\item{taxa}{Vector of taxon names to get a subset for.}

\item{phy}{A user tree to congruify as \code{phylo} object (ape).}

\item{phy4}{A user tree to congruify in \code{phylo4} format (phylobase).}

\item{dating_method}{The method used for tree dating.}
}
\value{
A list with a patristic matrix array and a \verb{$problem} if any.
}
\description{
Subset a reference and a target tree given as \code{phylo} objects.
}
