% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{phylo_tips_from_node}
\alias{phylo_tips_from_node}
\title{To get tip numbers descending from any given node of a tree}
\usage{
phylo_tips_from_node(phy = NULL, node = NULL, curr = NULL)
}
\arguments{
\item{phy}{A tree either as a newick character string or as a phylo object}

\item{node}{an integer specifying a node number in the tree.}

\item{curr}{the set of previously stored node numbers - used in recursive function calls.}
}
\value{
A numeric vector with tip numbers descending from a node
}
\description{
To get tip numbers descending from any given node of a tree
}
