% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{phylo_to_patristic_matrix}
\alias{phylo_to_patristic_matrix}
\title{Get a patristic matrix from a \code{phylo} object.}
\usage{
phylo_to_patristic_matrix(phy, test = TRUE, tol = 0.01, option = 2)
}
\arguments{
\item{phy}{A \code{phylo} object.}

\item{test}{Default to \code{TRUE}. Whether to test if \code{phy} has branch lengths and is ultrametric or not.}

\item{tol}{branching time in \code{reference} above which secondary constraints will be applied to \code{target}}

\item{option}{an integer (1 or 2; see details).}
}
\value{
A patristic matrix.
}
\description{
Get a patristic matrix from a \code{phylo} object.
}
