% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groves.R
\name{pick_grove}
\alias{pick_grove}
\title{Pick a grove in the case of multiple groves in a set of trees.}
\usage{
pick_grove(grove_list, criterion = "taxa", datelife_result)
}
\arguments{
\item{grove_list}{A list of vectors of tree indices. Each element is a grove.}

\item{criterion}{A character vector indicating whether to get the grove with the most "trees" or the most "taxa"}

\item{datelife_result}{datelifeResult object (named list of patristic matrices). Only needed for "taxa" criterion}
}
\value{
A numeric vector of the elements of the picked grove.
}
\description{
Pick a grove in the case of multiple groves in a set of trees.
}
