% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plant_bold_otol_tree}
\alias{plant_bold_otol_tree}
\title{Some plants chronogram}
\format{
A phylo object with 6 tips and 5 internal nodes
\describe{
\item{edge}{Integer vector with edge (branch) numbers}
\item{tip.label}{Character vector with species names of plants}
\item{Nnode}{Integer vector with the number of nodes}
\item{node.label}{Character vector with node names}
\item{edge.length}{Numeric vector with edge (branch) lengths}
}
}
\source{
\url{http://opentreeoflife.org}

\url{http://www.boldsystems.org}
}
\usage{
plant_bold_otol_tree
}
\description{
Some plants chronogram
}
\details{
Generated with make_bold_otol_tree(input = "((Zea mays,Oryza sativa),((Arabidopsis thaliana,(Glycine max,Medicago sativa)),Solanum lycopersicum)Pentapetalae);")
usethis::use_data(plant_bold_otol_tree)
}
\author{
Luna L. Sanchez-Reyes \email{lsanche7@utk.edu}

Brian O'Meara \email{bomeara@utk.edu}
}
\keyword{bold}
\keyword{chronogram}
\keyword{otol}
\keyword{plant}
\keyword{tree}
