% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_densitree}
\alias{plot_densitree}
\title{get a densiTree plot from a set of opentree_chronograms
if densiTree plot function throws an error, it chooses the tree with the most tips as consensus (using get_biggest_phylo)
we found that densiTree errors commonly from failing to do a consensus tree.}
\usage{
plot_densitree(trees)
}
\arguments{
\item{trees}{A list of trees as multiPhylo or as a plain list object.}
}
\description{
get a densiTree plot from a set of opentree_chronograms
if densiTree plot function throws an error, it chooses the tree with the most tips as consensus (using get_biggest_phylo)
we found that densiTree errors commonly from failing to do a consensus tree.
}
