% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_ltt.R
\name{plot_ltt_phyloall}
\alias{plot_ltt_phyloall}
\title{Lineahe through time plots of  all chronograms in multiphylo object.}
\usage{
plot_ltt_phyloall(taxon = NULL, phy, ltt_colors = NULL,
  tax_datedotol = NULL, file_name = NULL, file_dir = NULL,
  height = 3.5, width = 7, add_legend = FALSE, add_title = FALSE)
}
\arguments{
\item{taxon}{Character vector indicating the name of the taxon or lineage that the chronograms in phy belong to.}

\item{phy}{A phylo or multiphylo object with chronograms (trees with branch lengths proportional to geologic time), ideally.}

\item{ltt_colors}{A character vector indicating the colors to be ised for plotting ltt}

\item{tax_datedotol}{A chronogram to compare trees in phy to.}

\item{file_name}{A character string giving the name of the pdf file.}

\item{file_dir}{A character string giving the path to write the file to.}

\item{add_legend}{Boolean}

\item{add_title}{Boolean}
}
\description{
Lineahe through time plots of  all chronograms in multiphylo object.
}
