% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_phylo}
\alias{plot_phylo}
\title{plot one tree with study title and geochronological axis}
\usage{
plot_phylo(
  tree,
  title = "Tree",
  time_depth = NULL,
  axis_type = 1,
  cex = graphics::par("cex"),
  mai4 = NULL,
  write = "nothing",
  file_name = NULL,
  GTS = getAnywhere("strat2012")
)
}
\arguments{
\item{tree}{A tree either as a newick character string or as a phylo object}

\item{title}{A character string giving the name and path to write the files to.}

\item{time_depth}{A numeric vector indicating the upper limit on the time x axis scale.}

\item{axis_type}{A numeric vector indicating the type of geochronological axis to plot. See examples.}

\item{cex}{a numeric value giving the factor scaling of the tip and
    node labels (Character EXpansion). The default is to take the
    current value from the graphical parameters.}

\item{mai4}{A numeric vector of length one indicating the space needed for plotting whole tip labels (right margin of the plot).}

\item{write}{A character vector of length 1. Use pdf or png to write a file on those formats respectively. Anything else will not write any image file.}

\item{file_name}{A character string giving the name and path to write the files to.}

\item{GTS}{A dataframe of geochronological limits.}
}
\description{
plot one tree with study title and geochronological axis
}
