% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_phylo_all}
\alias{plot_phylo_all}
\title{plot all trees with study titles and geochronological axis}
\usage{
plot_phylo_all(trees, include = TRUE, individually = TRUE)
}
\arguments{
\item{trees}{A list of trees as multiPhylo or as a plain list object.}

\item{include}{Boolean or numeric vector. Default to TRUE, plot all chronograms
in trees. If FALSE, exclude chronograms with only two tips. If numeric, plot
trees matching the numeric vector.}

\item{individually}{Boolean indicating if trees should be plotted one by one or all on the same file}
}
\description{
plot all trees with study titles and geochronological axis
}
