% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_chronograms.R
\docType{data}
\name{problems}
\alias{problems}
\title{Problematic chronograms from Open Tree of Life.}
\format{
A list of trees with unmapped taxa
}
\source{
\url{http://opentreeoflife.org}
}
\usage{
problems
}
\description{
Problematic chronograms from Open Tree of Life.
}
\details{
Before we developed tools to clean and map tip labels for our cached trees
we found some trees that were stored with unmapped tip labels
we extracted them and saved them to be used for testing functions.
Generated with
problems <- opentree_chronograms$trees[sapply(sapply(opentree_chronograms$trees, "[", "tip.label"), function(x) any(grepl("not.mapped", x)))]
usethis::use_data(problems)
opentree_chronograms object from commit https://github.com/phylotastic/datelife/tree/be894448f6fc437241cd0916fab45e84ac3e09c6

[", "tip.label"), function(x) any(grepl("not.mapped", x)))]: R:\%22,\%20\%22tip.label\%22),\%20function(x)\%20any(grepl(\%22not.mapped\%22,\%20x)))
}
\keyword{chronogram}
\keyword{otol}
\keyword{tnrs}
\keyword{tree}
\keyword{unmapped}
