% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{ProcessInput}
\alias{ProcessInput}
\title{Take input string, figure out if it's newick or list of species}
\usage{
ProcessInput(input = c("Rhea americana", "Pterocnemia pennata",
  "Struthio camelus"), usetnrs = FALSE, approximatematch = TRUE)
}
\arguments{
\item{input}{Target taxa names in the form of a vector of characters, a newick character string, or a phylo object.}

\item{usetnrs}{If TRUE, use OpenTree's services to resolve names. This can dramatically improve the chance of matches, but also take much longer.}

\item{approximatematch}{If TRUE, use a slower TNRS to correct mispellings, increasing the chance of matches (including false matches).}
}
\value{
A list with the phy (or NA, if no tree) and cleaned vector of taxa
}
\description{
Take input string, figure out if it's newick or list of species
}
