% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{ProcessResultsList}
\alias{ProcessResultsList}
\title{Take results.list and process it}
\usage{
ProcessResultsList(results.list, taxa = NULL, partial = FALSE)
}
\arguments{
\item{results.list}{A list returned from using GetSubsetArrayDispatch on opentree_chronograms$trees}

\item{taxa}{A vector of taxa to match}

\item{partial}{If TRUE, return matrices that have only partial matches}
}
\value{
A list with the patristic.matrices that are not NA
}
\description{
Take results.list and process it
}
