% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_taxonomy.R
\name{recover_mrcaott}
\alias{recover_mrcaott}
\title{gets an mrcaott tag from an open tree of life tree and gets its name and ott id}
\usage{
recover_mrcaott(tag)
}
\arguments{
\item{tag}{A character vector with the mrca tag}
}
\value{
A numeric vector with ott id from original taxon named with the corresponding ott name
}
\description{
gets an mrcaott tag from an open tree of life tree and gets its name and ott id
}
