% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonely_functions.R
\name{relevant_curators_tabulate}
\alias{relevant_curators_tabulate}
\title{Return the relevant curators for a set of studies.}
\usage{
relevant_curators_tabulate(results.index, cache = "opentree_chronograms")
}
\arguments{
\item{results.index}{A vector from \code{\link[=datelife_result_study_index]{datelife_result_study_index()}} with the indices of the relevant studies.}

\item{cache}{The cached chronogram database.}
}
\value{
A vector with counts of each curator, with names equal to curator names.
}
\description{
Return the relevant curators for a set of studies.
}
