% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{ReorderMatrix}
\alias{ReorderMatrix}
\title{Function to reorder a matrix so that row and column labels are in alphabetical order}
\usage{
ReorderMatrix(patristic.matrix)
}
\arguments{
\item{patristic.matrix}{A patristic matrix with row and column names for taxa}
}
\value{
patristic.matrix A patristic matrix with row and column names for taxa in alphabetial order
}
\description{
Function to reorder a matrix so that row and column labels are in alphabetical order
}

