% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{run}
\alias{run}
\title{Core function to generate results}
\usage{
run(input = c("Rhea americana", "Pterocnemia pennata", "Struthio camelus"),
  format = "html", partial = "liberal", uncertainty = 100,
  randomtreesperstudy = 0, plot.width = 600, plot.height = 600,
  usetnrs = "no", approximatematch = "yes", prunenonmatch = "yes",
  datelife.cache = datelife.cache, do.out = TRUE)
}
\arguments{
\item{input}{A newick string or vector of taxa}

\item{format}{The output format}

\item{partial}{How to deal with trees that have a subset of taxa in the query}

\item{uncertainty}{How much to multiply the range by to represent uncertainty}

\item{randomtreesperstudy}{IDK}

\item{plot.width}{Width in pixels for output plot}

\item{plot.height}{Height in pixels for output plot}

\item{usetnrs}{Whether to use OpenTree's TNRS for the input}

\item{approximatematch}{IDK}

\item{datelife.cache}{The list of lists containing the input trees and other info}

\item{pruneonmatch}{IDK}
}
\value{
Depends on options
}
\description{
Core function to generate results
}

