% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{run_mrbayes}
\alias{run_mrbayes}
\title{Runs MrBayes from R.}
\usage{
run_mrbayes(file = NULL)
}
\arguments{
\item{file}{A character vector specifying the name of mrBayes run file and outputs (can specify directory too).}
}
\value{
MrBayes outputs.
}
\description{
Runs MrBayes from R.
}
