% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{RunSDM}
\alias{RunSDM}
\title{Function to compute the SDM supertree Criscuolo et al. 2006}
\usage{
RunSDM(filtered.results, weighting = "flat")
}
\arguments{
\item{filtered.results}{List of patristic matrices}

\item{weighting}{flat, taxa, inverse}
}
\value{
A list containing phy (a chronogram), and filtered.results that were actually used
}
\description{
Function to compute the SDM supertree Criscuolo et al. 2006
}
\details{
Weighting is how much weight to give each input tree.
   flat = all trees have equal weighting
   taxa = weight is proportional to number of taxa
   inverse = weight is proportional to 1 / number of taxa
Criscuolo A, Berry V, Douzery EJ, Gascuel O. SDM: a fast distance-based approach for (super) tree building in phylogenomics. Syst Biol. 2006;55(5):740–55. doi: 10.1080/10635150600969872.
}

