% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{sample_trees}
\alias{sample_trees}
\title{Function to sample trees from a file containing multiple trees. Usually from a bayesian analysis output trees file.}
\usage{
sample_trees(trees_file, trees_object = NULL, burnin = 0.25, size = 100)
}
\arguments{
\item{trees_file}{A character vector indicating the name and directory of file with trees to sample.}

\item{trees_object}{An R object containing a list of trees already read intro R from a tree file from a bayesian analysis output.}

\item{burnin}{A numeric vector indicating the burnin fraction. It should be a number between 0 and 1. Default to 0.25}

\item{size}{A numeric vector indicating the number of samples to be generated.}
}
\value{
A multiPhylo object with a random sample of trees.
}
\description{
Function to sample trees from a file containing multiple trees. Usually from a bayesian analysis output trees file.
}
