% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_chronograms.R
\name{save_otol_chronograms}
\alias{save_otol_chronograms}
\title{Save all chronograms from Open Tree of Life}
\usage{
save_otol_chronograms(file = "opentree_chronograms.RData")
}
\arguments{
\item{file}{Path including file name}
}
\value{
None
}
\description{
Save all chronograms from Open Tree of Life
}
