% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_chronograms.R
\name{SaveOToLChronograms}
\alias{SaveOToLChronograms}
\title{Save all chronograms from Open Tree of Life}
\usage{
SaveOToLChronograms(file = "opentree_chronograms.RData", verbose = FALSE)
}
\arguments{
\item{file}{Path including file name}

\item{verbose}{If TRUE, give status updates to the user}
}
\value{
None
}
\description{
Save all chronograms from Open Tree of Life
}
