% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_sdm.R
\name{sdm_matrix_to_phylo}
\alias{sdm_matrix_to_phylo}
\title{Go from an SDM matrix to anultrametric phylo object.}
\usage{
sdm_matrix_to_phylo(sdm_matrix)
}
\arguments{
\item{sdm_matrix}{A matrix from get_sdm}
}
\value{
An ultrametric phylo object.
}
\description{
Go from an SDM matrix to anultrametric phylo object.
}
