% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.r
\name{slice}
\alias{slice}
\title{Function to break up a vector by certain N sized chunks.}
\usage{
slice(input, by = 2, equalpiecesof = NULL)
}
\arguments{
\item{input}{Input character vector}

\item{by}{Number by which to split the input character vector}
}
\description{
Function to break up a vector by certain N sized chunks.
}

