% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{some_ants_datelife_result}
\alias{some_ants_datelife_result}
\title{datelifeResult object of some ants}
\format{
A list of one element, containing a named patristic matrix
}
\source{
\url{http://opentreeoflife.org}
}
\usage{
some_ants_datelife_result
}
\description{
datelifeResult object of some ants
}
\details{
Generated with:
some_ants_input <- "(Aulacopone_relicta,(((Myrmecia_gulosa,(Aneuretus_simoni,Dolichoderus_mariae)),((Ectatomma_ruidum,Huberia_brounii),Formica_rufa)),Apomyrma_stygia),Martialis_heureka)Formicidae;"
some_ants_datelife_query <- make_datelife_query(input = some_ants_input)
some_ants_datelife_result <- get_datelife_result(input = some_ants_datelife_query)
usethis::use_data(some_ants_datelife_result)
}
\keyword{ants}
\keyword{chronogram}
\keyword{datelife}
\keyword{otol}
\keyword{subset}
\keyword{tree}
