% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{strat2012}
\alias{strat2012}
\title{A data frame containing the stratigraphic chart by issued in 2012 by the International Commission on Stratigraphy.}
\format{
A data frame
\describe{
  \item{eon}{Eon name.}
  \item{era}{Era name.}
  \item{period}{Period name.}
  \item{epoch}{Epoch name.}
  \item{stage}{Stage name.}
  \item{MA}{Estimated boundary age for the associated interval.}
  \item{error}{Estimated errors associated with each age estimate.}
  \item{GSSP}{Binary response denoting whether the age estimate is defined by a basal Global Standard Section and Point}
}
}
\source{
\url{https://github.com/fmichonneau/phyloch}
}
\usage{
strat2012
}
\description{
A data frame containing the stratigraphic chart by issued in 2012 by the International Commission on Stratigraphy.
}
\details{
See phyloch package data description.
Enhance: There are a couple errors in the original strat2012 object that are not too meaningful but could be improved.
Generated with utils::data("strat2012", package = "phyloch")
use_data(strat2012)
}
\keyword{geochronology}
\keyword{gradstein}
\keyword{plot}
