% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subset2_search}
\alias{subset2_search}
\title{A list with datelifeQuery and datelifeResult objects from a search of taxon names from subset2_taxa}
\format{
A list with two named elements. datelifeResult object with 24 patristic matrices
\describe{
\item{datelife_query}{A datelifeQuery object using names_subset 2 as input.}
\item{datelife_result}{A datelifeResult object resulting from a search of names in datelifeQuery}
}
}
\usage{
subset2_search
}
\description{
A list with datelifeQuery and datelifeResult objects from a search of taxon names from subset2_taxa
}
\details{
Generated with:
datelife_query <- make_datelife_query(subset2_taxa)
datelife_result <- get_datelife_result(datelife_query)
subset2_search <- list(query = datelife_query, result = datelife_result)
usethis::use_data(subset2_search, overwrite = TRUE)
}
\keyword{datelifeResult}
\keyword{names}
\keyword{subset2}
\keyword{taxon}
