% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subset2_taxa}
\alias{subset2_taxa}
\title{Long list of >2.7k virus, bacteria, plant and animal taxon names}
\format{
A character vector of length 2778
}
\source{
\url{https://github.com/phylotastic/rphylotastic/tree/master/tests/testthat}
}
\usage{
subset2_taxa
}
\description{
Long list of >2.7k virus, bacteria, plant and animal taxon names
}
\details{
Generated with:
subset2_taxa <- rphylotastic::url_get_scientific_names("https://github.com/phylotastic/rphylotastic/blob/master/tests/testthat/subset2.txt")
usethis::use_data(subset2_taxa)
}
\keyword{animal}
\keyword{bacteria}
\keyword{names}
\keyword{plant}
\keyword{subset}
\keyword{taxon}
\keyword{virus}
