% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{subset_trees}
\alias{subset_trees}
\title{subset trees for plotting densitree plots and phylo_all plots}
\usage{
subset_trees(trees, include = TRUE)
}
\arguments{
\item{trees}{A list of trees as multiPhylo or as a plain list object.}

\item{include}{Boolean or numeric vector. Default to TRUE, plot all chronograms
in trees. If FALSE, exclude chronograms with only two tips. If numeric, plot
trees matching the numeric vector.}
}
\description{
subset trees for plotting densitree plots and phylo_all plots
}
