% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrations_congruify.R
\name{summarize_congruifiedCalibrations}
\alias{summarize_congruifiedCalibrations}
\title{Get summary statistics of ages in a \code{congruifiedCalibrations} object.}
\usage{
summarize_congruifiedCalibrations(congruified_calibrations, age_column)
}
\arguments{
\item{congruified_calibrations}{A \code{congruifiedCalibrations} object, output of \code{\link[=congruify_and_mrca_multiPhylo]{congruify_and_mrca_multiPhylo()}}.}

\item{age_column}{A character string indicating the name of the column to be summarized.}
}
\value{
A \code{data.frame} of summarized ages.
}
\description{
Function \code{summarize_congruifiedCalibrations} returns a table of
summary statistic for each node in \code{congruified_calibrations} argument.
}
