% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_phylo.R
\name{summarize_summary_matrix}
\alias{summarize_summary_matrix}
\title{Internal for summary_matrix_to_phylo().}
\usage{
summarize_summary_matrix(summ_matrix)
}
\arguments{
\item{summ_matrix}{A summary patristic distance matrix from sdm or median. See details.}
}
\description{
Internal for summary_matrix_to_phylo().
}
