% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{SummarizeResults}
\alias{SummarizeResults}
\title{Summarize a filtered results list in various ways}
\usage{
SummarizeResults(filtered.results, output.format, partial = TRUE,
  cache = datelife.cache, suppress.citations = FALSE)
}
\arguments{
\item{filtered.results}{A list of patristic matrices; labels correspond to citations}

\item{output.format}{The desired output format}

\item{partial}{If TRUE, use source trees even if they only match some of the desired taxa}

\item{cache}{The cached trees and other information}

\item{suppress.citations}{If using a format that would normally print() citations, turn this off}
}
\value{
Depends on output format
}
\description{
Summarize a filtered results list in various ways
}

