% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_summary.R
\name{summary.datelifeResult}
\alias{summary.datelifeResult}
\title{Main function to summarize a datelifeResult object}
\usage{
\method{summary}{datelifeResult}(object, ..., partial = TRUE)
}
\arguments{
\item{object}{A "datelifeResult" object, typically an output of get_datelife_result function.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Main function to summarize a datelifeResult object
}
