% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrations_match.R
\name{summary.matchedCalibrations}
\alias{summary.matchedCalibrations}
\title{Summarize a \code{matchedCalibrations} object
summary.datelifeCalibrations gets the node age distribution from a \code{matchedCalibrations} object.}
\usage{
\method{summary}{matchedCalibrations}(object, ...)
}
\arguments{
\item{object}{A \code{matchedCalibrations} object, usually an output of \code{\link[=match_all_calibrations]{match_all_calibrations()}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summarize a \code{matchedCalibrations} object
summary.datelifeCalibrations gets the node age distribution from a \code{matchedCalibrations} object.
}
