% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_phylo.R
\name{summary_matrix_to_phylo}
\alias{summary_matrix_to_phylo}
\title{Go from a smmary matrix to an ultrametric phylo object.}
\usage{
summary_matrix_to_phylo(summ_matrix)
}
\arguments{
\item{summ_matrix}{A summary patristic distance matrix from sdm or median. See details.}
}
\value{
An ultrametric phylo object.
}
\description{
Go from a smmary matrix to an ultrametric phylo object.
}
\details{
It can take a regular patristic distance matrix, but there are simpler methods for that implemented in patristic_matrix_to_phylo.
}
