% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{summary_patristic_matrix_array}
\alias{summary_patristic_matrix_array}
\title{Summarize patristic matrix array (by default, median). Used inside: summarize_datelife_result.}
\usage{
summary_patristic_matrix_array(patristic_matrix_array, fn = stats::median)
}
\arguments{
\item{patristic_matrix_array}{3D array of patristic matrices}

\item{fn}{The function to use to summarize}
}
\value{
A 2d array with the median (or max, or mean, etc) of the input array
}
\description{
Summarize patristic matrix array (by default, median). Used inside: summarize_datelife_result.
}
