% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonely_functions.R
\name{TabulateRelevantAuthors}
\alias{TabulateRelevantAuthors}
\title{Return the relevant authors for a set of studies}
\usage{
TabulateRelevantAuthors(results.index, cache = get("opentree_chronograms"))
}
\arguments{
\item{results.index}{A vector from FindMatchingStudyIndex() with the indices of the relevant studies}

\item{cache}{The cache}
}
\value{
A vector with counts of each author, with names equal to author names
}
\description{
Return the relevant authors for a set of studies
}
