% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tb.author.pretty}
\alias{tb.author.pretty}
\title{Authors from studies with chronograms in TreeBASE}
\format{A data frame with two elements
\describe{
  \item{person}{Character vector of authors of the included studies}
  \item{study.count}{Numeric vector of studies authored by each person}
}}
\source{
\url{http://treebase.org}
}
\usage{
tb.author.pretty
}
\description{
Authors from studies with chronograms in TreeBASE
}
\details{
Generated with make_treebase_cache()
}
\keyword{author}
\keyword{chronogram}
\keyword{count}
\keyword{study}
\keyword{tree}
\keyword{treebase}
