% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tb.author.results}
\alias{tb.author.results}
\title{Information on authors and study ids from chronograms in TreeBASE}
\format{A dataframe with three variables
\describe{
  \item{person}{Named character vector of authors of the included studies}
  \item{study}{Named character vector of study identifiers}
}}
\source{
\url{http://treebase.org}
}
\usage{
tb.author.results
}
\description{
Information on authors and study ids from chronograms in TreeBASE
}
\details{
Generated with make_treebase_cache()
}
\keyword{chronogram}
\keyword{id}
\keyword{studies}
\keyword{study}
\keyword{tree}
\keyword{treebase}
