% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_search.R
\docType{data}
\name{threebirds_dr}
\alias{threebirds_dr}
\title{\code{datelifeResult} object of three birds "Rhea americana", "Pterocnemia pennata", and "Struthio camelus"}
\format{
A list of 9 named patristic matrix
}
\source{
\url{http://opentreeoflife.org}
}
\usage{
threebirds_dr
}
\description{
\code{datelifeResult} object of three birds "Rhea americana", "Pterocnemia pennata", and "Struthio camelus"
}
\details{
Generated with:
threebirds_dr <- get_datelife_result(input=c("Rhea americana", "Pterocnemia pennata", "Struthio camelus"),
partial = TRUE, use_tnrs = FALSE, approximate_match = TRUE, cache = "opentree_chronograms")
use_data(threebirds_dr)
}
\keyword{ants}
\keyword{chronogram}
\keyword{datelife}
\keyword{otol}
\keyword{subset}
\keyword{tree}
