% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.r
\name{time_cleaning}
\alias{time_cleaning}
\title{Test timing of name cleaning on datelife trees}
\usage{
time_cleaning(listoftrees, plotresults = FALSE)
}
\arguments{
\item{listoftrees}{List of trees, each tree in phylo format, not multiPhylo
format.}

\item{plotresults}{Plot results or not.}
}
\description{
Test timing of name cleaning on datelife trees
}
\examples{
\dontrun{
treefilenames <- dir("/Users/scottmac2/phyloorchard/pkg/data")
l_ply(treefilenames, function(x) load(paste("/Users/scottmac2/phyloorchard/pkg/data/",x,sep=""), .GlobalEnv))
trees <- sapply(treefilenames, function(x) str_replace(x, ".rda", ""), USE.NAMES=F)
binemonds <- BinindaEmondsEtAl2007[[1]]
 makesmallsubtrees <- function(phylo){
 		drops <- c(2000,2500,3000,3500,4000,4100,4200,4300,4350,4400,4420,4450,4470,4500,4505)
 		llply(rev(drops), function(x) drop.tip(phylo, tip=1:x))
}
mytreelist <- makesmallsubtrees(phylo = binemonds)
library(ggplot2); library(plyr); library(lubridate)
out <- time_cleaning(listoftrees = mytreelist, plotresults=TRUE)
out[[1]] # data.frame
out[[2]] # plot
}
}

