% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_taxonomy_general.R
\name{tnrs_match.phylo}
\alias{tnrs_match.phylo}
\title{Taxon name resolution service (tnrs) applied to tips of a phylogeny}
\usage{
tnrs_match.phylo(phy, tip, reference_taxonomy = "otl", ...)
}
\arguments{
\item{phy}{A phylo object}

\item{tip}{A vector of mode numeric or character specifying the tips to match. If left empty all tips will be matched.}

\item{reference_taxonomy}{A character vector specifying the reference taxonomy to use for tnrs.}

\item{...}{Arguments passed on to \code{rotl::tnrs_match_names}
\describe{
  \item{}{}
}}
}
\value{
An object of class phylo and match_names. See details.
}
\description{
Taxon name resolution service (tnrs) applied to tips of a phylogeny
}
\details{
The output will preserve all elements from original input phylo object and will add
\describe{
    \item{phy$mapped}{A character vector indicating the state of mapping of phy$tip.labels:}
		\describe{
		    \item{original}{Tnrs matching was not attempted. Original labeling is preserved.}
		    \item{ott}{Matching was manually made by a curator in Open Tree of Life.}
		    \item{tnrs}{Tnrs matching was attempted and succesful with no approximate matching. Original label is replaced by the matched name.}
		    \item{approximated}{Tnrs matching was attempted and succesful but with approximate matching. Original labeling is preserved.}
		    \item{unmatched}{Tnrs matching was attempted and unsuccesful. Original labeling is preserved.}
		}
    \item{phy$original.tip.label}{A character vector preserving all original labels.}
    \item{phy$ott_ids}{A numeric vector with ott id numbers of matched tips. Unmatched and original tips will be NaN.}
}
if tips are duplicated, tnrs will only be run once (avoiding increases in function running time) but the result will be applied to all duplicated tip labels
}
