% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_taxonomy_general.R
\name{tnrs_match}
\alias{tnrs_match}
\title{Taxon name resolution service (tnrs) applied to query by batches}
\usage{
tnrs_match(names, reference_taxonomy = "otl", ...)
}
\arguments{
\item{names}{A character vector of taxon names to be matched to taxonomy. There is no limit to the numer of names that can be queried.}

\item{reference_taxonomy}{A character vetor specifying the reference taxonomy to use for tnrs.}

\item{...}{Arguments passed on to \code{rotl::tnrs_match_names}
\describe{
  \item{}{}
}}
}
\value{
A data frame from tnrs_match_names function
}
\description{
Taxon name resolution service (tnrs) applied to query by batches
}
