% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{tree_add_dates}
\alias{tree_add_dates}
\title{Fabricates dates of missing taxa (with no data) on an already dated tree.}
\usage{
tree_add_dates(dated_phy = NULL, missing_taxa = NULL,
  dating_method = "mrbayes",
  mrbayes_output_file = "mrbayes_tree_add_dates.nexus")
}
\arguments{
\item{dated_phy}{a tree (newick or phylo) with branch lengths proportional to absolute time}

\item{missing_taxa}{either a tree (as phylo object or as a newick character string, with or without branch lengths) with all the taxa you want at the end, or a data frame assigning missing taxa to nodes in dated tree.}

\item{dating_method}{The method used for tree dating.}

\item{mrbayes_output_file}{A character vector specifying the name of mrBayes run file and outputs (can specify directory too).}
}
\value{
A phylo object
}
\description{
Fabricates dates of missing taxa (with no data) on an already dated tree.
}
