% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_processing.R
\name{tree_add_outgroup}
\alias{tree_add_outgroup}
\title{Function to add an outgroup to any phylogeny, in phylo or newick format}
\usage{
tree_add_outgroup(tree = NULL, outgroup = "outgroup")
}
\arguments{
\item{tree}{A tree either as a newick character string or as a \code{phylo} object.}

\item{outgroup}{A character vector with the name of the outgroup. If it has length>1, only first element will be used.}
}
\value{
A phylo object with no root edge.
}
\description{
Function to add an outgroup to any phylogeny, in phylo or newick format
}
