% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_processing.R
\name{tree_check}
\alias{tree_check}
\title{Checks if a tree is a phylo class object otherwise it uses input_process. Additionally it can check if tree is a chronogram with phylo_check}
\usage{
tree_check(tree = NULL, ...)
}
\arguments{
\item{tree}{A tree either as a newick character string or as a phylo object}

\item{...}{Arguments passed on to \code{phylo_check}
\describe{
  \item{phy}{A phylo object}
  \item{dated}{Boolean. If TRUE it checks if phylo object has branch lengths and is ultrametric.}
}}
}
\value{
If tree is correctly formatted, it returns a phylo object
}
\description{
Checks if a tree is a phylo class object otherwise it uses input_process. Additionally it can check if tree is a chronogram with phylo_check
}
