% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{tree_get_node_data}
\alias{tree_get_node_data}
\title{Gets ages, node numbers, node names and descendant tips number and label of all nodes from a dated tree}
\usage{
tree_get_node_data(tree = NULL, node_data = c("node_number", "node_label",
  "node_age", "descendant_tips_number", "descendant_tips_label"))
}
\arguments{
\item{tree}{A tree either as a newick character string or as a phylo object}

\item{node_data}{A character vector containing one or all from: "node_number", "node_label", "node_age", "descendant_tips_number", "descendant_tips_label"}
}
\value{
A list
}
\description{
Gets ages, node numbers, node names and descendant tips number and label of all nodes from a dated tree
}
