% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_processing.R
\name{tree_node_tips}
\alias{tree_node_tips}
\title{To get tip numbers descending from any given node of a tree}
\usage{
tree_node_tips(tree = NULL, node = NULL, curr = NULL)
}
\arguments{
\item{tree}{A tree either as a newick character string or as a phylo object}
}
\value{
A numeric vector with tip numbers descending from a node
}
\description{
To get tip numbers descending from any given node of a tree
}
