% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treebase_cache}
\alias{treebase_cache}
\title{Information on contributors, authors, study ids and clades from studies with chronograms in Open tree of Life}
\format{
A list of five data sets
\describe{
\item{tb.author.pretty}{A dataframe with two elements: author names and number of studies in TreeBase authored by each}
\item{tb.author.results}{A dataframe with two elements: author names and study identifiers}
}
}
\source{
TreeBASE database, no longer available online \url{https://en.wikipedia.org/wiki/TreeBASE}
}
\usage{
treebase_cache
}
\description{
Information on contributors, authors, study ids and clades from studies with chronograms in Open tree of Life
}
\details{
Generated with make_treebase_cache()
}
\keyword{author}
\keyword{chronogram}
\keyword{count}
\keyword{id}
\keyword{study}
\keyword{tree}
\keyword{treebase}
