% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{UnpadMatrix}
\alias{UnpadMatrix}
\title{Function to remove missing taxa. Used in: RunSDM.}
\usage{
UnpadMatrix(patristic.matrix)
}
\arguments{
\item{patristic.matrix}{A patristic matrix with row and column names for taxa}
}
\value{
Patristic.matrix for all.taxa
}
\description{
Function to remove missing taxa. Used in: RunSDM.
}
