% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_chronograms.R
\name{update_all_cached}
\alias{update_all_cached}
\title{Update all data files as data objects for the package}
\usage{
update_all_cached()
}
\value{
None
}
\description{
This includes opentree chronograms, contributors, treebase and curators
For speed, datelife caches chronograms and other information. Running this (within the checked out version of datelife) will refresh these. Then git commit and git push them back
}
