% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_chronograms.R
\name{UpdateCache}
\alias{UpdateCache}
\title{update Open Tree of Life cache}
\usage{
UpdateCache(save = TRUE, file = "opentree_chronograms.RData",
  verbose = TRUE)
}
\arguments{
\item{save}{Boolean; default TRUE: save all chronograms from Open Tree of Life to an RData file (default to opentree_chronograms.RData)}

\item{file}{Path including file name}

\item{verbose}{If TRUE, give status updates to the user}
}
\value{
A list with elements for the trees, authors, curators, and study ids
}
\description{
update Open Tree of Life cache
}
