% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_calibrations_use.R
\name{use_all_calibrations}
\alias{use_all_calibrations}
\title{Perform a dating analysis on a set of input taxa.
It searches and uses all calibrations from chronograms in the DateLife database.}
\usage{
use_all_calibrations(input = NULL, dating_method = "bladj", ...)
}
\arguments{
\item{input}{A character vector of taxon names, a tree as a 'phylo' object or a newick character string, or a 'datelifeQuery' object from make_datelife_query function.}

\item{...}{
  Arguments passed on to \code{\link[=get_all_calibrations]{get_all_calibrations}}
  \describe{
    \item{\code{each}}{Boolean, default to FALSE: all calibrations are returned in the same data frame. If TRUE, calibrations from each chronogram are returned in a separate data frame.}
  }}
}
\value{
A list with a chronogram and all calibrations found
}
\description{
Perform a dating analysis on a set of input taxa.
It searches and uses all calibrations from chronograms in the DateLife database.
}
\details{
If input is a tree, it will get all calibrations and date it with bladj if it
has no branch lengths or with PATHd8 if it has branch lengths.
If input is a vector of names, it will try to construct a bold tree first to get
a topology with branch lengths. If it can't, it will get the open tree of life
topology only and date it with bladj.
}
