% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrations_use.R
\name{use_calibrations}
\alias{use_calibrations}
\title{Date a given tree topology using a combined set of given calibrations}
\usage{
use_calibrations(
  phy = NULL,
  calibrations = NULL,
  dating_method = "bladj",
  type = "median",
  ...
)
}
\arguments{
\item{phy}{A \code{phylo} object with or without branch lengths}

\item{calibrations}{An object of class \code{datelifeCalibrations},
typically an output of [get_all_calibrations()].}

\item{dating_method}{Tree dating method. Options are "bladj" or "pathd8"}

\item{...}{
  Arguments passed on to \code{\link[=use_calibrations_pathd8]{use_calibrations_pathd8}}
  \describe{
    \item{\code{expand}}{How much to expand by each step to get consistent calibrations. Should be between 0 and 1.}
    \item{\code{giveup}}{How many expansions to try before giving up}
  }}
}
\value{
A \code{phylo} object with branch lengths proportional to time.
}
\description{
\code{use_calibrations} combines all given calibrations and uses them as
constraints to perform a dating analysis on a given tree topology, using BLADJ
if it has no branch lengths, or PATHd8 if the given tree topology has initial
branch lengths.
}
\details{
If \code{input} tree does not have branch lengths, \code{dating_method} is
ignored and the function will use secondary calibrations to date the \code{input}
tree with [BLADJ](http://phylodiversity.net/bladj/)
using [phylocomr::ph_bladj()] (via [use_calibrations_bladj()]). If \code{input}
tree has have branch lengths it can be dated
with [PATHd8](https://www2.math.su.se/PATHd8/) using [geiger::PATHd8.phylo()]
(via [use_calibrations_pathd8()]).
}
