% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrations_use.R
\name{use_calibrations_each}
\alias{use_calibrations_each}
\title{Date a given tree topology by using a given list of calibrations independently,
to generate multiple hypothesis of time of divergence}
\usage{
use_calibrations_each(phy = NULL, calibrations = NULL, ...)
}
\arguments{
\item{phy}{A \code{phylo} object with or without branch lengths}

\item{calibrations}{An object of class \code{datelifeCalibrations},
typically an output of [get_all_calibrations()].}

\item{...}{Arguments to pass to use_calibrations}
}
\value{
A list with a multiPhylo object of chronograms and a list of all calibrations used
}
\description{
\code{use_calibrations_each} wraps \code{use_calibrations} to take each set of
given calibrations and use it independently as constraints for BLADJ or PATHd8
to date a given tree topology.
}
\details{
If \code{input} tree does not have branch lengths, \code{dating_method} is
ignored and the function will use secondary calibrations to date the \code{input}
tree with [BLADJ](http://phylodiversity.net/bladj/)
using [phylocomr::ph_bladj()] (via [use_calibrations_bladj()]). If \code{input}
tree has have branch lengths it can be dated
with [PATHd8](https://www2.math.su.se/PATHd8/) using [geiger::PATHd8.phylo()]
(via [use_calibrations_pathd8()]).
}
