% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_calibrations_pathd8.R
\name{use_calibrations_pathd8}
\alias{use_calibrations_pathd8}
\title{Use calibrations to date a tree with branch lengths with PATHd8.}
\usage{
use_calibrations_pathd8(phy, calibrations, expand = 0.1, giveup = 100)
}
\arguments{
\item{phy}{A phylo object}

\item{calibrations}{A data frame of calibrations from get_all_calibrations function, or a subset of it.}

\item{expand}{How much to expand by each step to get consistent calibrations. Should be between 0 and 1.}

\item{giveup}{How many expansions to try before giving up}
}
\value{
A phylo object
}
\description{
Use calibrations to date a tree with branch lengths with PATHd8.
}
\details{
This function will try to use the calibrations as fixed ages.
If that fails (often due to conflict between calibrations), it will expand the
range of the minimum age and maximum age and try again. And repeat.
If expand = 0, it uses the summarized calibrations.
In some cases, it returns edge lengths in relative time (with maximum tree depth = 1)
instead of absolute time, as given by calibrations. In this case, the function returns NA.
This is an issue from PATHd8.
}
