% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_calibrations_treePL.R
\name{use_calibrations_treePL}
\alias{use_calibrations_treePL}
\title{Use calibrations to date a tree with branch lengths with treePL.}
\usage{
use_calibrations_treePL(phy, calibrations)
}
\arguments{
\item{phy}{A phylo object with or wothput branch lengths}

\item{calibrations}{A data frame of calibrations from get_all_calibrations function, or a subset of it.}
}
\value{
A phylo object
}
\description{
Use calibrations to date a tree with branch lengths with treePL.
}
\details{
This function will try to use the calibrations as fixed ages.
If that fails (often due to conflict between calibrations), it will expand the
range of the minimum age and maximum age and try again. And repeat.
If expand = 0, it uses the summarized calibrations.
In some cases, it returns edge lengths in relative time (with maximum tree depth = 1)
instead of absolute time, as given by calibrations. In this case, the function returns NA.
This is an issue from PATHd8.
}
