% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_calibrations_use.R
\name{use_each_calibration}
\alias{use_each_calibration}
\title{Use calibrations from each source chronogram to date a tree.}
\usage{
use_each_calibration(phy = NULL, phylo_all = NULL,
  calibrations = NULL, ...)
}
\arguments{
\item{phy}{A tree with or without branch lengths to be dated. Only as phylo object for now.}

\item{phylo_all}{Can be NULL. A datelifeResult list of patristic matrices, or a chronogram as phylo or multiPhylo.}

\item{calibrations}{Can be NULL. A list of dataframes from get_all_calibrations function.}
}
\value{
A list with a multiPhylo object of chronograms and a list of all calibrations used
}
\description{
Use calibrations from each source chronogram to date a tree.
}
\details{
You can  get the datelifeResult object or the list of chronograms first
phylo_all1 <- get_datelife_result(input = my_phy)
phylo_all2 <- summarize_datelife_result(phylo_all1)
Either will work the same:
use_each_calibration(phy = my_phy, phylo_all = phylo_all1)
use_each_calibration(phy = my_phy, phylo_all = phylo_all2)
You can also get the list of calibrations outside
my_calibrations <- get_all_calibrations(input = phylo_all1, each = TRUE)
use_each_calibration(phy = my_phy, calibrations = my_calibrations)
All this means that you can use your own set of calibrations, not necessarily the ones found only in datelife.
}
