% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{BindMatrices}
\alias{BindMatrices}
\title{Convert list of patristic matrices to a 3D array}
\usage{
BindMatrices(patristic.matrix.list)
}
\arguments{
\item{patristic.matrix.list}{List of patristic matrices}
}
\value{
A 3d array of patristic matrices
}
\description{
Convert list of patristic matrices to a 3D array
}
