% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groves.R
\name{build_grove_list}
\alias{build_grove_list}
\title{Build grove list}
\usage{
build_grove_list(datelife_result, n = 2)
}
\arguments{
\item{datelife_result}{A \code{datelifeResult} object.}

\item{n}{The degree of taxon name overlap among input chronograms. Defaults
to \code{n = 2}, i.e., at least two overlapping taxon names.}
}
\value{
A list of vectors; each list element is a grove.
}
\description{
This function implements theorem 1.1 of Ané et al. (2009) \doi{10.1007/s00026-009-0017-x}
to find a grove for a given group of chronograms.
}
