% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_taxonomy.R
\name{check_ott_input}
\alias{check_ott_input}
\title{checks input for get_ott_clade,  get_ott_children functions
returns a numeric vector of ott ids}
\usage{
check_ott_input(input = NULL, ott_ids = NULL, ...)
}
\arguments{
\item{input}{Optional. A character vector of names or a datelifeQuery object}

\item{ott_ids}{If not NULL, it takes this argument and ignores input. A numeric vector of ott ids obtained with rotl::taxonomy_taxon_info or rolt::tnrs_match_names or datelife::tnrs_match}

\item{...}{Arguments passed on to \code{datelife_query_check}
\describe{
  \item{datelife_query}{A datelifeQuery object, output of make_datelife_query function}
}}
}
\value{
A named numeric vector of valid ott IDs
}
\description{
checks input for get_ott_clade,  get_ott_children functions
returns a numeric vector of ott ids
}
\details{
By default it uses ott_id argument if it is not NULL.
}
