% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_taxonomy.R
\name{check_ott_input}
\alias{check_ott_input}
\title{checks input for get_ott_clade and get_ott_children functions
returns a numeric vector of ott ids}
\usage{
check_ott_input(input, ott_id)
}
\arguments{
\item{input}{A character vector of names}

\item{ott_id}{A numeric vector of ott ids obtained with rotl::taxonomy_taxon_info or rolt::tnrs_match_names or datelife::tnrs_match}
}
\description{
checks input for get_ott_clade and get_ott_children functions
returns a numeric vector of ott ids
}
