% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{choose_cluster}
\alias{choose_cluster}
\title{Choose a phylo object from cluster_patristicmatrix. If not ultrametric, it does not force it.}
\usage{
choose_cluster(phycluster, clustering_method)
}
\arguments{
\item{phycluster}{An output from cluster_patristicmatrix}

\item{clustering_method}{A character vector indicating the method to construct the tree. Options are "nj" for Neighbor-Joining and "upgma" for Unweighted Pair Group Method with Arithmetic Mean.}
}
\value{
A phylo object or NA
}
\description{
Choose a phylo object from cluster_patristicmatrix. If not ultrametric, it does not force it.
}
