% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_phylo.R
\name{cluster_patristicmatrix}
\alias{cluster_patristicmatrix}
\title{Cluster a patristic matrix with several methods}
\usage{
cluster_patristicmatrix(patristic_matrix)
}
\arguments{
\item{patristic_matrix}{A patristic matrix}
}
\value{
A list of results from clustering with NJ and UPGMA
}
\description{
Cluster a patristic matrix with several methods
}
