% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_phylo.R
\name{cluster_patristicmatrix}
\alias{cluster_patristicmatrix}
\title{Cluster a patristic matrix into a tree with various methods.}
\usage{
cluster_patristicmatrix(patristic_matrix, variance_matrix = NULL)
}
\arguments{
\item{patristic_matrix}{A patristic matrix}

\item{variance_matrix}{A variance matrix from a \code{datelifeResult} object,
usually an output from \code{\link[=datelife_result_variance_matrix]{datelife_result_variance_matrix()}}.
Only used if \code{clustering_method = "mvr"}.}
}
\value{
A list of trees obtained with clustering methods detailed in \code{\link[=patristic_matrix_to_phylo]{patristic_matrix_to_phylo()}}.
}
\description{
Cluster a patristic matrix into a tree with various methods.
}
\details{
If clustering method fails, \code{NA} is returned.
}
