% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contributor_cache}
\alias{contributor_cache}
\title{Information on contributors, authors, study ids and clades from studies with chronograms in Open tree of Life}
\format{
A list of five data sets
\describe{
\item{author.pretty}{A character vector with the author names from studies with chronograms that are in OToL}
\item{author.results}{A dataframe with three variables: authors, study ids and clades}
\item{curator.pretty}{A character vector with the names of curators of chronograms that are in OToL}
\item{curator.results}{A dataframe with three variables: curators, study ids and clades}
\item{missed_doi}{A character vector with study ids whose doi could not be retrieved}
}
}
\source{
\url{http://opentreeoflife.org}
}
\usage{
contributor_cache
}
\description{
Information on contributors, authors, study ids and clades from studies with chronograms in Open tree of Life
}
\details{
Generated with make_contributor_cache()
}
\keyword{author}
\keyword{chronogram}
\keyword{otol}
\keyword{studies}
\keyword{study}
\keyword{tree}
