% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{datelife_result_check}
\alias{datelife_result_check}
\title{checks if we obtained an empty search with the set of input taxon names}
\usage{
datelife_result_check(datelife_result, use_tnrs, verbose = FALSE)
}
\arguments{
\item{datelife_result}{An object output from get_datelife_result function}

\item{use_tnrs}{If TRUE, use OpenTree's services to resolve names. This can dramatically improve the chance of matches, but also take much longer.}

\item{verbose}{Boolean. If TRUE, it gives printed updates to the user.}
}
\description{
checks if we obtained an empty search with the set of input taxon names
}
