% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_median.R
\name{datelife_result_median}
\alias{datelife_result_median}
\title{Get a median summary chronogram from a datelifeResult object.}
\usage{
datelife_result_median(datelife_result)
}
\arguments{
\item{datelife_result}{An output of get_datelife_result function: A list of patristic matrices with names corresponding to original study citations.}
}
\value{
A phylo object
}
\description{
Get a median summary chronogram from a datelifeResult object.
}
