% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_median.R
\name{datelife_result_median_matrix}
\alias{datelife_result_median_matrix}
\title{Compute a median matrix of a \code{datelifeResult} object.}
\usage{
datelife_result_median_matrix(datelife_result)
}
\arguments{
\item{datelife_result}{A \code{datelifeResult} object, usually an output of \code{\link[=get_datelife_result]{get_datelife_result()}}.}
}
\value{
A patristic distance summary matrix from a \code{datelifeResult} object.
}
\description{
Compute a median matrix of a \code{datelifeResult} object.
}
