% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{datelife_result_MRCA}
\alias{datelife_result_MRCA}
\title{Get vector of MRCAs from a datelifeResult object. Used in: summarize_datelife_result.}
\usage{
datelife_result_MRCA(datelife_result, partial = TRUE)
}
\arguments{
\item{datelife_result}{An object from get_datelife_result function.}

\item{partial}{If TRUE, drop NA from the patristic matrix; if FALSE, will return NA if there are missing entries}
}
\value{
Vector of MRCA ages with names same as in datelife_result
}
\description{
Get vector of MRCAs from a datelifeResult object. Used in: summarize_datelife_result.
}
