% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{datelife_result_MRCA}
\alias{datelife_result_MRCA}
\title{Get a numeric vector of MRCAs from a \code{datelifeResult} object. Used in \code{\link[=summarize_datelife_result]{summarize_datelife_result()}}.}
\usage{
datelife_result_MRCA(datelife_result, na_rm = TRUE)
}
\arguments{
\item{datelife_result}{A \code{datelifeResult} object, usually an output of \code{\link[=get_datelife_result]{get_datelife_result()}}.}

\item{na_rm}{If \code{TRUE}, it drops rows containing \code{NA}s from the \code{datelifeResult}
patristic matrix; if \code{FALSE}, it returns \code{NA} where there are missing entries.}
}
\value{
A named numeric vector of MRCA ages for each element given in \code{datelife_result}.
}
\description{
Get a numeric vector of MRCAs from a \code{datelifeResult} object. Used in \code{\link[=summarize_datelife_result]{summarize_datelife_result()}}.
}
