% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_summary.R
\name{datelife_result_sdm}
\alias{datelife_result_sdm}
\title{Function to compute the SDM supertree (Criscuolo et al. 2006) from a datelifeResult object.}
\usage{
datelife_result_sdm(datelife_result, weighting = "flat",
  verbose = TRUE)
}
\arguments{
\item{datelife_result}{An output of get_datelife_result function: A list of patristic matrices with names corresponding to original study citations.}

\item{weighting}{A character vector indicating how much weight to give to each input tree in the SDM analysis.
     Choose one of:
\describe{
\item{weighting = "flat"}{All trees have equal weighting.
}
\item{weighting = "taxa"}{Weight is proportional to number of taxa.
}
\item{weighting = "inverse"}{Weight is proportional to 1 / number of taxa.
}
}}

\item{verbose}{Boolean. If TRUE, it gives printed updates to the user.}
}
\value{
A list of two elements:
\describe{
\item{phy}{A phylo object with the output chronogram from SDM analysis.
}
\item{data}{A datelifeResult object with the chronograms that were used to construct the SDM tree.
}
}
}
\description{
Function to compute the SDM supertree (Criscuolo et al. 2006) from a datelifeResult object.
}
\details{
Criscuolo A, Berry V, Douzery EJ, Gascuel O. SDM: a fast distance-based approach for (super) tree building in phylogenomics. Syst Biol. 2006. 55(5):740. doi: 10.1080/10635150600969872.
}
