% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_sdm.R
\name{datelife_result_sdm}
\alias{datelife_result_sdm}
\title{Function to compute the SDM supertree (Criscuolo et al. 2006) from a
\code{datelifeResult} object}
\usage{
datelife_result_sdm(datelife_result, weighting = "flat", verbose = TRUE, ...)
}
\arguments{
\item{datelife_result}{An output of get_datelife_result function: A list of patristic matrices with names corresponding to original study citations.}

\item{weighting}{A character vector indicating how much weight to give to each
tree in \code{input} during the SDM analysis. Choose one of:
\describe{
\item{weighting = "flat"}{All trees have equal weighting.}
\item{weighting = "taxa"}{Weight is proportional to number of taxa.}
\item{weighting = "inverse"}{Weight is proportional to 1 / number of taxa.}
}}

\item{verbose}{Boolean. If \code{TRUE}, it gives printed updates to the user.}

\item{...}{
  Arguments passed on to \code{\link[=summary_matrix_to_phylo]{summary_matrix_to_phylo}}
  \describe{
    \item{\code{summ_matrix}}{A summary patristic distance matrix from sdm or median. See details.}
    \item{\code{total_distance}}{Boolean. If TRUE it will divide the matrix in half, if FALSE it will take it as is.}
    \item{\code{use}}{A character vector indicating what type of age to use for summary. One of the following
\describe{
\item{mean}{It will use the mean of the node age distributions.}
\item{min}{It will use the minimum age from the node age distributions.}
\item{max}{Choose this if you wanna be conservative; it will use the maximum age from the node age distributions.}
}}
    \item{\code{target_tree}}{A phylo object. Use this in case you want a particular backbone for the output tree.}
    \item{\code{datelife_query}}{A \code{datelifeQuery} object, output of \code{\link[=make_datelife_query]{make_datelife_query}}.}
  }}
}
\value{
A \code{phylo} object from SDM with an extra element:
\describe{
\item{data}{A \code{datelifeResult} object with the chronograms that were used
 	          to construct the SDM tree.
}
}
}
\description{
Function to compute the SDM supertree (Criscuolo et al. 2006) from a
\code{datelifeResult} object
}
\details{
Criscuolo A, Berry V, Douzery EJ, Gascuel O. SDM: a fast distance-based approach for (super) tree building in phylogenomics. Syst Biol. 2006. 55(5):740. doi: 10.1080/10635150600969872.
}
