% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{datelife_result_study_index}
\alias{datelife_result_study_index}
\title{Find the index of relevant studies in a opentree_chronograms object. Used inside: summarize_datelife_result.}
\usage{
datelife_result_study_index(datelife_result,
  cache = get("opentree_chronograms"))
}
\arguments{
\item{datelife_result}{The patristic.matrices that will be used}

\item{cache}{The cache of studies}
}
\value{
A vector with the indices of studies that have relevant info
}
\description{
Find the index of relevant studies in a opentree_chronograms object. Used inside: summarize_datelife_result.
}
