% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{datelife_result_study_index}
\alias{datelife_result_study_index}
\title{Find the index of relevant studies in a cached chronogram database.}
\usage{
datelife_result_study_index(datelife_result, cache = "opentree_chronograms")
}
\arguments{
\item{datelife_result}{A \code{datelifeResult} object, usually an output of \code{\link[=get_datelife_result]{get_datelife_result()}}.}

\item{cache}{The cached chronogram database.}
}
\value{
A vector of indices of studies that have relevant information.
}
\description{
\code{datelife_result_study_index} is used in \code{\link[=summarize_datelife_result]{summarize_datelife_result()}}.
}
