% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_use.R
\name{datelife_use}
\alias{datelife_use}
\alias{use_all_calibrations}
\title{Generate one or multiple chronograms for a set of given taxa}
\usage{
datelife_use(input = NULL, each = FALSE, dating_method = "bladj")
}
\arguments{
\item{input}{One of the following:
\describe{
\item{Taxon names}{As a character vector.}
\item{A tree with taxon names as tip labels}{As a \code{phylo} or \code{multiPhylo}
            object, OR as a newick character string.}
\item{A \code{datelifeQuery} object}{An output from [make_datelife_query_vector()]
                \code{\link[=make_datelife_query]{make_datelife_query}}.}
}}

\item{dating_method}{Tree dating method. Options are "bladj" or "pathd8"}
}
\value{
A list with a chronogram and secondary calibrations used for dating.
}
\description{
\code{datelife_use} generates one or multiple chronograms (i.e., phylogenetic
trees with branch lengths proportional to time) for a set of \code{input} taxa,
dated with bladj or PATHd8, using secondary calibrations available for any pair
of \code{input} taxa, mined from the code{\link[=opentree_chronograms]{opentree_chronograms}}
local DateLife database.
}
\details{
If input is a tree, it will use secondary calibrations to (1) date the tree with bladj
if it has no branch lengths, or (2) date the tree with PATHd8 if it has branch lengths.
If input is a vector of taxon names, it will attempt to reconstruct a BOLD tree first
to get a topology with branch lengths. If it can't, it will get an Open Tree
of Life synthetic tree topology and will date it with bladj.
}
