% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_use.R
\name{datelife_use}
\alias{datelife_use}
\alias{use_all_calibrations}
\title{Generate one or multiple chronograms for a set of given taxon names}
\usage{
datelife_use(input = NULL, each = FALSE, dating_method = "bladj", ...)
}
\arguments{
\item{input}{One of the following:
\describe{
\item{Taxon names}{As a character vector.}
\item{One or more trees with taxon names as tip labels}{As a \code{phylo} or \code{multiPhylo}
            object, OR as a newick character string.}
\item{A \code{datelifeQuery} object}{An output from [make_datelife_query()]
                \code{\link[=make_datelife_query]{make_datelife_query}}.}
}}

\item{dating_method}{Tree dating method. Options are "bladj" or "pathd8"}

\item{...}{
  Arguments passed on to \code{\link[=make_datelife_query]{make_datelife_query}}
  \describe{
    \item{\code{verbose}}{Boolean. If \code{TRUE}, it gives printed updates to the user.}
  }}
}
\value{
A list with one or multiple chronograms and secondary calibrations used for dating.
}
\description{
\code{datelife_use} gets secondary calibrations available for any
pair of \code{input} taxon names, mined from code{\link[=opentree_chronograms]{opentree_chronograms}}
local DateLife database, and uses them to date a given tree topology with BLADJ
or PATHd8. If no tree topology is provided, it will attempt to generate one with
BOLD-OpenTree through [make_bold_otol_tree()].
}
\details{
If input is a tree, it will use secondary calibrations to (1) date the tree with bladj
if it has no branch lengths, or (2) date the tree with PATHd8 if it has branch lengths.
If input is a vector of taxon names, it will attempt to reconstruct a BOLD tree first
to get a topology with branch lengths. If it can't, it will get an Open Tree
of Life synthetic tree topology and will date it with bladj.
}
