% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_use.R
\name{datelife_use_datelifequery}
\alias{datelife_use_datelifequery}
\title{Generate one or multiple chronograms from a \code{datelifeQuery} object.}
\usage{
datelife_use_datelifequery(input = NULL, dating_method = "bladj", each = FALSE)
}
\arguments{
\item{input}{A \code{datelifeQuery} object.}

\item{...}{
  Arguments passed on to \code{\link[=get_all_calibrations]{get_all_calibrations}}, \code{\link[=use_all_calibrations]{use_all_calibrations}}
  \describe{
    \item{\code{each}}{Boolean, default to \code{FALSE}: all calibrations are returned in
the same data frame. If \code{TRUE}, calibrations from each chronogram are returned
in separate data frames.}
    \item{\code{calibrations}}{An object of class \code{datelifeCalibrations},
typically an output of \code{\link[=get_all_calibrations]{get_all_calibrations()}}.}
    \item{\code{dating_method}}{Tree dating method. Options are "bladj" or "pathd8"}
    \item{\code{phy}}{A \code{phylo} object with or without branch lengths}
  }}
}
\value{
A list with a chronogram and secondary calibrations used.
}
\description{
\code{datelife_use_datelifequery} generates one or multiple chronograms (i.e., phylogenetic
trees with branch lengths proportional to time) for a set of \code{input} taxa,
dated with bladj or PATHd8, using secondary calibrations available for any pair
of \code{input} taxa, mined from the code{\link[=opentree_chronograms]{opentree_chronograms}}
local DateLife database.
}
