% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_taxonomy.R
\name{.get_ott_lineage}
\alias{.get_ott_lineage}
\title{Get the lineage of a set of taxa using rotl:taxonomy_taxon_info(include_lineage = TRUE)}
\usage{
.get_ott_lineage(input_ott_match)
}
\arguments{
\item{input_ott_match}{An Output of check_ott_input function.}
}
\value{
A taxonomy_taxon_info object
}
\description{
Get the lineage of a set of taxa using rotl:taxonomy_taxon_info(include_lineage = TRUE)
}
