% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{EstimateDates}
\alias{EstimateDates}
\alias{datelife}
\title{Core function to input a vector of species, newick string, or phylo object to get a chronogram or dates back.}
\usage{
EstimateDates(input = c("Rhea americana", "Pterocnemia pennata",
  "Struthio camelus"), output.format = "phylo.sdm", partial = TRUE,
  usetnrs = FALSE, approximatematch = TRUE, update_cache = FALSE,
  cache = get("opentree_chronograms"), method = "PATHd8", bold = FALSE,
  verbose = c("citations", "taxa"), missing.taxa = c("none", "summary",
  "matrix"), marker = "COI", ...)
}
\arguments{
\item{input}{Target taxa names in the form of a vector of characters, a newick character string, or a phylo object.}

\item{output.format}{The desired output format for target chronograms (chronograms of target taxa). See details.}

\item{partial}{If TRUE, use source chronograms even if they only match some of the desired taxa}

\item{usetnrs}{If TRUE, use OpenTree's services to resolve names. This can dramatically improve the chance of matches, but also take much longer.}

\item{approximatematch}{If TRUE, use a slower TNRS to correct mispellings, increasing the chance of matches (including false matches).}

\item{update_cache}{default to FALSE}

\item{cache}{The cached set of chronograms and other info from data(opentree_chronograms).}

\item{method}{The method used for congruification. PATHd8 only right now, r8s and treePL later.}

\item{bold}{Logical. If TRUE, use Barcode of Life Data Systems (BOLD)  and Open Tree of Life (OToL) backbone to estimate branch lengths of target taxa using GetBoldOToLTree function.}

\item{verbose}{A character vector specifying type of information to be printed: "citations" for the references of chronograms from cache where target taxa are found, "taxa" for a summary of the number of chronograms where each target taxon is found, or "none" if nothing should be printed. Default to display both c("citations", "taxa").}

\item{missing.taxa}{A character vector specifying if data on target taxa missing in source chronograms should be added to the output as a "summary" or as a presence/absence "matrix". Default to "none", no information on missing.taxa added to the output.}

\item{marker}{A character vector with the name of the gene from Barcode of Life Data Systems (BOLD) to be used for branch length estimation.}

\item{...}{Arguments passed on to \code{GetBoldOToLTree}
\describe{
  \item{otol_version}{Version of OToL to use}
  \item{chronogram}{Boolean; default to TRUE:  branch lengths represent time estimated with ape::chronoMPL. If FALSE, branch lengths represent relative substitution rates estimated with phangorn::acctran.}
  \item{doML}{Boolean; if TRUE, does ML branch length optimization with phangorn::optim.pml}
}}
}
\description{
Core function to input a vector of species, newick string, or phylo object to get a chronogram or dates back.
}
\details{
Available output formats are:

citations: A character vector of references where chronograms with some or all of the target taxa are published (source chronograms).

mrca: A named numeric vector of most recent common ancestor (mrca) ages of target taxa defined in input, obtained from the source chronograms. Names of mrca vector are equal to citations.

newick.all: A named character vector of newick strings corresponding to target chronograms derived from source chronograms. Names of newick.all vector are equal to citations.

newick.sdm: Only if multiple source chronograms are available. A character vector with a single newick string corresponding to a target chronogram obtained with SDM supertree method (Criscuolo et al. 2006).

newick.median: Only if multiple source chronograms are available. A character vector with a single newick string corresponding to a target chronogram from the median of all source chronograms.

phylo.sdm: Only if multiple source chronograms are available. A phylo object with a single target chronogram obtained with SDM supertree method (Criscuolo et al. 2006).

phylo.median: Only if multiple source chronograms are available. A phylo object with a single target chronogram obtained from source chronograms with median method.

phylo.all: A named list of phylo objects corresponding to each target chronogram obtained from available source chronograms. Names of phylo.all list correspond to citations.

html: A character vector with an html string that can be saved and then opened in any web browser. It contains a 4 column table with data on target taxa: mrca, number of taxa, citations of source chronogram and newick target chronogram.

data.frame A data.frame with data on target taxa: mrca, number of taxa, citations of source chronograms and newick string.
}
\examples{
# obtain median ages from a set of source chronograms in newick format:
ages <- EstimateDates(c("Rhea americana", "Pterocnemia pennata", "Struthio camelus",
		"Mus musculus"), output.format="newick.median")
# save the tree in newick format
write(ages, file="some.bird.ages.txt")

# obtain median ages from a set of source chronograms in phylo format
# will produce same tree as above but in r phylo format:
ages.again <- EstimateDates(c("Rhea americana", "Pterocnemia pennata", "Struthio camelus",
		"Mus musculus"), output.format="phylo.median")
plot(ages.again)
library(ape)
ape::axisPhylo()
mtext("Time (million years ago)", side=1, line=2, at = (max(get("last_plot.phylo",
		envir = .PlotPhyloEnv)$xx) * 0.5))
write.tree(ages.again, file="some.bird.tree.again.txt") # saves phylo object in newick format

# obtain mrca ages and target chronograms from all source chronograms
# generate an html  output readable in any web browser:
ages.html <- EstimateDates(c("Rhea americana", "Pterocnemia pennata", "Struthio camelus",
		"Mus musculus"), output.format="html")
write(ages.html, file="some.bird.trees.html")
system("open some.bird.trees.html")
}
