% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrations_get.R
\name{extract_calibrations_phylo}
\alias{extract_calibrations_phylo}
\title{Extract secondary calibrations from a \code{phylo} or \code{multiPhylo}
object with branch lengths proportional to time}
\usage{
extract_calibrations_phylo(input = NULL, each = FALSE)
}
\arguments{
\item{input}{a \code{phylo} or \code{multiPhylo} object with branch lengths
proportional to time.}

\item{each}{Boolean, default to \code{FALSE}: all calibrations are returned in
the same data frame. If \code{TRUE}, calibrations from each chronogram are returned
in separate data frames.}
}
\value{
An object of class \code{datelifeCalibrations} -- a \code{data frame}
of secondary calibrations (or list of \code{data frames}, if \code{each = TRUE}),
for each pair of taxon names in \{input]. The attribute "chronograms" stores
the source data from which the calibrations were extracted.
}
\description{
\code{extract_calibrations_phylo} extracts divergence times
(i.e., secondary calibrations) for each taxon pair in a given
\code{phylo} or \code{multiPhylo} object.
}
