% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{FindMatchingStudyIndex}
\alias{FindMatchingStudyIndex}
\title{Find the index of relevant studies in a opentree_chronograms object}
\usage{
FindMatchingStudyIndex(filtered.results, cache = get("opentree_chronograms"))
}
\arguments{
\item{filtered.results}{The patristic.matrices that will be used}

\item{cache}{The cache of studies}
}
\value{
A vector with the indices of studies that have relevant info
}
\description{
Find the index of relevant studies in a opentree_chronograms object
}

