% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_phylo.R
\name{force_ultrametric}
\alias{force_ultrametric}
\title{Force a non-ultrametric \code{phylo} object to be ultrametric with \code{\link[phytools:force.ultrametric]{phytools::force.ultrametric()}}.}
\usage{
force_ultrametric(phy)
}
\arguments{
\item{phy}{A phylo object}
}
\value{
A \code{phylo} object.
}
\description{
Force a non-ultrametric \code{phylo} object to be ultrametric with \code{\link[phytools:force.ultrametric]{phytools::force.ultrametric()}}.
}
