% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_calibrations.R
\name{get_all_calibrations}
\alias{get_all_calibrations}
\title{Get secondary calibrations from a set of given taxon names}
\usage{
get_all_calibrations(input = NULL, each = FALSE)
}
\arguments{
\item{input}{\describe{
\item{Taxon names}{As a character vector.}
\item{A tree with taxon names as tip labels}{As a \code{phylo} or \code{multiPhylo} object, OR as a newick character string.}
\item{A \code{datelifeResult} object}{From \code{\link[=get_datelife_result]{get_datelife_result}}.}
}}

\item{each}{Boolean, default to \code{FALSE}: all calibrations are returned in
the same data frame. If \code{TRUE}, calibrations from each chronogram are returned
in separate data frames.}
}
\value{
A data frame of secondary calibrations (or list of data frames, if \code{each = TRUE}),
for each pair of given taxon names. The attribute "chronograms" contains the
source chronograms from which the calibrations were obtained.
}
\description{
\code{get_all_calibrations} gets divergence times (i.e., secondary
calibrations) for each taxon name pair given in \code{input}.
}
