% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_calibrations_get.R
\name{get_all_calibrations}
\alias{get_all_calibrations}
\title{Get all secondary calibrations available for a set of taxa.}
\usage{
get_all_calibrations(
  input = c("Rhea americana", "Pterocnemia pennata", "Struthio camelus"),
  ...,
  each = FALSE
)
}
\arguments{
\item{input}{A character vector of taxon names; a newick string, OR a phylo OR
multiPhylo object with branch lengths proportional to time; OR a datelifeResult object.}

\item{...}{
  Arguments passed on to \code{\link[=datelife_search]{datelife_search}}
  \describe{
    \item{\code{summary_format}}{The desired output format for target chronograms (chronograms
of target taxa). See details.}
    \item{\code{summary_print}}{A character vector specifying type of summary information
to be printed: "citations" for the references of chronograms from cache where target
taxa are found, "taxa" for a summary of the number of chronograms where each target
taxon is found, or "none" if nothing should be printed. Default to display both
c("citations", "taxa").}
    \item{\code{taxon_summary}}{A character vector specifying if data on target taxa missing
in source chronograms should be added to the output as a "summary" or as a
presence/absence "matrix". Default to "none", no information on taxon_summary
added to the output.}
    \item{\code{partial}}{Boolean; default to TRUE: use source trees even if they only
match some of the desired taxa.}
    \item{\code{use_tnrs}}{Boolean; default to FALSE. If TRUE, use OpenTree's services
to resolve names. This can dramatically improve the chance of matches, but also
take longer.}
    \item{\code{approximate_match}}{Boolean; default to TRUE: use a slower TNRS to correct
misspellings, increasing the chance of matches (including false matches)}
    \item{\code{update_opentree_chronograms}}{default to FALSE}
    \item{\code{cache}}{A character vector of length one, with the name of the data object
to cache. Default to "opentree_chronograms", a data object storing Open Tree of
Life's database chronograms and other associated information.}
    \item{\code{get_spp_from_taxon}}{boolean vector, default to FALSE. If TRUE, will get
all species names from taxon names given in input. Must have same length as input.
If input is a newick string , with some clades it will be converted to phylo
object phy, and the order of get_spp_from_taxon will match phy$tip.label.}
    \item{\code{verbose}}{Boolean. If TRUE, it gives printed updates to the user.}
    \item{\code{criterion}}{Whether to get the grove with the most trees or the most taxa}
  }}

\item{each}{Boolean, default to \code{FALSE}: all calibrations are returned in
the same data frame. If \code{TRUE}, calibrations from each chronogram are returned
in separate data frames.}
}
\value{
A data frame of secondary calibrations (or list of data frames, if \code{each = TRUE}),
for each pair of given taxon names. The attribute "chronograms" contains the
source chronograms from which the calibrations were obtained.
}
\description{
\code{get_all_calibrations} returns secondary calibrations for each pair of given
taxon names.
}
\details{
If input is a character vector of taxon names, the function calls a
\code{\link[=datelife_search]{datelife_search}} with \code{summary_format}
argument set to \code{"phylo_all"}, to get all chronograms containing at
least 2 of the taxa given in \code{input}.
The function retrieves divergence times for each pair of taxa in any given tree with
branch lengths proportional to time (chronogram).
}
