% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_calibrations_get.R
\name{get_all_calibrations}
\alias{get_all_calibrations}
\title{Get all calibrations available for a set of taxa.}
\usage{
get_all_calibrations(
  input = c("Rhea americana", "Pterocnemia pennata", "Struthio camelus"),
  partial = TRUE,
  use_tnrs = FALSE,
  approximate_match = TRUE,
  update_cache = FALSE,
  cache = getAnywhere("opentree_chronograms"),
  verbose = FALSE,
  each = FALSE
)
}
\arguments{
\item{input}{vector of names, a newick string, a phylo or multiPhylo object with branch lengths proportional to time, a datelifeResult object}

\item{partial}{Boolean; default TRUE: use source trees even if they only match some of the desired taxa}

\item{use_tnrs}{Boolean; default False. If TRUE, use OpenTree's services to resolve names. This can dramatically improve the chance of matches, but also take much longer}

\item{approximate_match}{Boolean; default TRUE: use a slower TNRS to correct misspellings, increasing the chance of matches (including false matches)}

\item{update_cache}{default to FALSE}

\item{cache}{The cached set of chronograms and other info from data(opentree_chronograms)}

\item{verbose}{Boolean. If TRUE, it gives printed updates to the user.}

\item{each}{Boolean, default to FALSE all calibrations are returned in the same data frame. If TRUE, calibrations from each chronogram are returned in a separate data frame.}
}
\value{
A data frame (or list of data frames if each = TRUE) of calibrations, with attribute "chronograms" containing the dated trees that were used to get calibrations from.
}
\description{
Get all calibrations available for a set of taxa.
}
