% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groves.R
\name{get_best_grove}
\alias{get_best_grove}
\title{Get grove from a datelifeResult object that can be converted to phylo from a median summary matrix.}
\usage{
get_best_grove(datelife_result, criterion = "taxa", n = 2)
}
\arguments{
\item{datelife_result}{datelifeResult object (named list of patristic matrices). Only needed for "taxa" criterion}

\item{criterion}{Whether to get the grove with the most trees or the most taxa}

\item{n}{Degree of overlap required}
}
\value{
A list of two elements:
\describe{
\item{best_grove}{A datelifeResult object filtered to only include one grove of trees that can be summarized with median or sdm.
}
\item{overlap}{The degree of taxon names overlap among trees in the best grove.
}
}
}
\description{
Get grove from a datelifeResult object that can be converted to phylo from a median summary matrix.
}
