% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrations_get.R
\name{get_calibrations_datelifequery}
\alias{get_calibrations_datelifequery}
\title{Search and extract available secondary calibrations from a given
'\code{datelifeQuery} object}
\usage{
get_calibrations_datelifequery(input = NULL, each = FALSE)
}
\arguments{
\item{input}{A \code{datelifeQuery} object.}

\item{each}{Boolean, default to \code{FALSE}: all calibrations are returned in
the same data frame. If \code{TRUE}, calibrations from each chronogram are returned
in separate data frames.}
}
\value{
An object of class \code{datelifeCalibrations} -- a \code{data frame}
of secondary calibrations (or list of \code{data frames}, if \code{each = TRUE}),
for each pair of taxon names in \{input]. The attribute "chronograms" stores
the source data from which the calibrations were extracted.
}
\description{
The function searches DateLife's local
database of phylogenetic trees with branch lengths proportional to time (aka,
chronograms) with [datelife_search()], and extracts divergence times
(i.e., secondary calibrations) from chronograms for each pair of given
taxon names with [extract_calibrations_phylo()].
}
\details{
The function calls [datelife_search()]
with \code{summary_format = "phylo_all"} to get all chronograms in database
containing at least two taxa from \code{input}, and generates a \code{phylo}
or \code{multiPhylo} object object that will be passed to
[extract_calibrations_phylo()].
}
