% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_summary.R
\name{get_dated_otol_induced_subtree}
\alias{get_dated_otol_induced_subtree}
\title{Get an otol induced dated subtree from your set of queried taxa}
\usage{
get_dated_otol_induced_subtree(input = c("Felis silvestris",
  "Homo sapiens"), ott_id = NULL)
}
\arguments{
\item{input}{Target taxa names as a character vector, a newick character string, or a phylo object.}

\item{ott_id}{Numeric vector of Open Tree Taxonomy ids. Input argumentt will be ignored.}
}
\value{
A phylo object with edge length proportional to time in Myrs. NA if 1 or no inputs are valid.
}
\description{
Get an otol induced dated subtree from your set of queried taxa
}
\details{
otol dated tree from Stephen Smith's otol scaling service.
}
