% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_result.R
\name{get_datelife_result}
\alias{get_datelife_result}
\title{Get a list of patristic matrices from a given set of taxon names}
\usage{
get_datelife_result(
  input = NULL,
  partial = TRUE,
  cache = "opentree_chronograms",
  ...
)
}
\arguments{
\item{input}{One of the following:
\describe{
\item{Taxon names}{As a character vector.}
\item{One or more trees with taxon names as tip labels}{As a \code{phylo} or \code{multiPhylo}
            object, OR as a newick character string.}
\item{A \code{datelifeQuery} object}{An output from [make_datelife_query()]
                \code{\link[=make_datelife_query]{make_datelife_query}}.}
}}

\item{partial}{Boolean; default to \code{TRUE}: use source trees even if they only
match some of the desired taxa.}

\item{cache}{A character vector of length one, with the name of the data object
to cache. Default to \code{"opentree_chronograms"}, a data object storing Open Tree of
Life's database chronograms and other associated information.}

\item{...}{
  Arguments passed on to \code{\link[=make_datelife_query]{make_datelife_query}}
  \describe{
    \item{\code{verbose}}{Boolean. If \code{TRUE}, it gives printed updates to the user.}
  }}
}
\description{
Go from a vector of species, newick string, or phylo object
}
