% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_sdm.R
\name{get_goodmatrices}
\alias{get_goodmatrices}
\title{Get indices of good matrices to apply Super Distance Matrix (SDM) method with \code{\link[=make_sdm]{make_sdm()}}.}
\usage{
get_goodmatrices(unpadded.matrices)
}
\arguments{
\item{unpadded.matrices}{A list of patristic matrices, a \code{datelifeResult} object.}
}
\value{
A numeric vector of good matrix indices in unpadded.matrices.
}
\description{
Get indices of good matrices to apply Super Distance Matrix (SDM) method with \code{\link[=make_sdm]{make_sdm()}}.
}
