% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_chronograms.R
\name{get_otol_chronograms}
\alias{get_otol_chronograms}
\title{Get all chronograms from Open Tree of Life}
\usage{
get_otol_chronograms(verbose = FALSE, max_tree_count = 500)
}
\arguments{
\item{verbose}{If TRUE, give updates to the user}

\item{max_tree_count}{Numeric indicating the max number of trees to be cached. For testing purposes only.}
}
\value{
A list with elements for the trees, authors, curators, and study ids
}
\description{
Get all chronograms from Open Tree of Life
}
