% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_calibrations.R
\name{get_otol_synthetic_tree}
\alias{get_otol_synthetic_tree}
\title{Gets Open Tree of Life synthetic tree of a set of lineages.}
\usage{
get_otol_synthetic_tree(input, otol_version = "v2")
}
\arguments{
\item{input}{Target taxa names as a character vector, a newick character string, or a phylo object.}

\item{otol_version}{Version of OToL to use}
}
\value{
A phylo object
}
\description{
Gets Open Tree of Life synthetic tree of a set of lineages.
}
