% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otol_tree.R
\name{get_otol_synthetic_tree}
\alias{get_otol_synthetic_tree}
\title{Gets Open Tree of Life synthetic tree of a set of lineages.}
\usage{
get_otol_synthetic_tree(
  input = NULL,
  ott_ids = NULL,
  otol_version = "v3",
  resolve = FALSE,
  ...
)
}
\arguments{
\item{input}{Optional. A character vector of names or a datelifeQuery object}

\item{ott_ids}{If not NULL, it takes this argument and ignores input. A numeric vector of ott ids obtained with rotl::taxonomy_taxon_info or rolt::tnrs_match_names or datelife::tnrs_match}

\item{otol_version}{Version of OToL to use}

\item{resolve}{boolean default to TRUE. Whether to resolve the tree at random or not.}

\item{...}{
  Arguments passed on to \code{\link[=make_datelife_query]{make_datelife_query}}
  \describe{
    \item{\code{use_tnrs}}{If TRUE, use OpenTree's services to resolve names. This can dramatically improve the chance of matches, but also take much longer.}
    \item{\code{approximate_match}}{If TRUE, use a slower TNRS to correct misspellings, increasing the chance of matches (including false matches).}
    \item{\code{get_spp_from_taxon}}{boolean vector, default to FALSE. If TRUE, will get all species names from taxon names given in input. Must have same length as input. If input is a newick string , with some clades it will be converted to phylo object phy, and the order of get_spp_from_taxon will match phy$tip.label.}
    \item{\code{verbose}}{Boolean. If TRUE, it gives printed updates to the user.}
  }}
}
\value{
A phylo object
}
\description{
Gets Open Tree of Life synthetic tree of a set of lineages.
}
