% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_taxonomy.R
\name{get_ott_children}
\alias{get_ott_children}
\title{Use this instead of rotl::tol_subtree when taxa are not in synthesis tree and
you still need to get all species or an induced OpenTree subtree}
\usage{
get_ott_children(input = NULL, ott_ids = NULL, ott_rank = "species", ...)
}
\arguments{
\item{input}{Optional. A character vector of names or a datelifeQuery object}

\item{ott_ids}{If not NULL, it takes this argument and ignores input. A numeric vector of ott ids obtained with rotl::taxonomy_taxon_info or rolt::tnrs_match_names or datelife::tnrs_match}

\item{ott_rank}{A character vector with the ranks you wanna get lineage children from.}

\item{...}{Other arguments to pass to get_valid_children()}
}
\value{
A data frame
}
\description{
Use this instead of rotl::tol_subtree when taxa are not in synthesis tree and
you still need to get all species or an induced OpenTree subtree
}
\examples{
# try getting an otol tree of a taxon missing from the synthetic tree
# tnrs <- rotl::tnrs_match_names("Mus")
tnrs <- tnrs_match("Canis")
# Mus, Gleicheniales, Polypodiales, etc., are missing in synth tree too
\dontrun{
  rotl::tol_subtree(tnrs$ott_id[1])
  Error: HTTP failure: 400
  [/v3/tree_of_life/subtree] Error: node_id was not found (broken taxon).
}
ids <- tnrs$ott_id[1]
names(ids) <- tnrs$unique_name
children <- get_ott_children(ott_ids = ids) # or
children <- get_ott_children(input = "Canis")
str(children)
ids <- children$Canis$ott_id
names(ids) <- rownames(children$Canis)
tree_children <- datelife::get_otol_synthetic_tree(ott_ids = ids)
plot(tree_children, cex = 0.3)
# Other examples:
oo <- get_ott_children(input= "magnoliophyta", ott_rank = "order")
sum(oo$Magnoliophyta$rank == "order") # to know how many orders of flowering plants we have
}
