% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_taxonomy.R
\name{get_ott_lineage}
\alias{get_ott_lineage}
\title{gets the ott id and name of all lineages from one or more input taxa}
\usage{
get_ott_lineage(input, ott_id = NULL)
}
\arguments{
\item{input}{A character vector of names}

\item{ott_id}{A numeric vector of ott ids obtained with rotl::taxonomy_taxon_info or rolt::tnrs_match_names or datelife::tnrs_match}
}
\value{
a list of named numeric vectors of ott ids from input and all the clades it belongs to.
}
\description{
gets the ott id and name of all lineages from one or more input taxa
}
\examples{
taxa <- c("Homo", "Bacillus anthracis", "Apis", "Salvia")
lin <- get_ott_lineage(taxa)
lin
}
