% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{get_subset_array_dispatch}
\alias{get_subset_array_dispatch}
\title{Figure out which subset function to use. Used inside \code{\link[=get_datelife_result]{get_datelife_result()}}}
\usage{
get_subset_array_dispatch(
  study_element,
  taxa,
  phy = NULL,
  phy4 = NULL,
  dating_method = "PATHd8"
)
}
\arguments{
\item{study_element}{The thing being passed in: an array or a phylo object to serve as reference for congruification}

\item{taxa}{Vector of taxon names to get a subset for}

\item{phy}{A user tree to congruify as phylo object (ape)}

\item{phy4}{A user tree to congruify in phylo4 format (phylobase)}

\item{dating_method}{The method used for tree dating.}
}
\value{
A patristic matrix with ages for the target taxa.
}
\description{
Figure out which subset function to use. Used inside \code{\link[=get_datelife_result]{get_datelife_result()}}
}
