% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_summary.R
\name{get_taxon_summary}
\alias{get_taxon_summary}
\title{Get taxon summary of a datelifeResult object}
\usage{
get_taxon_summary(datelife_result = NULL, datelife_query = NULL)
}
\arguments{
\item{datelife_result}{An output of get_datelife_result function: A list of patristic matrices with names corresponding to original study citations.}

\item{datelife_query}{A datelifeQuery object, output of make_datelife_query function}
}
\description{
Get taxon summary of a datelifeResult object
}
