% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{GetAllCalibrations}
\alias{GetAllCalibrations}
\title{Get all calibrations given trees in database}
\usage{
GetAllCalibrations(input = c("Rhea americana", "Pterocnemia pennata",
  "Struthio camelus"), partial = TRUE, usetnrs = FALSE,
  approximatematch = TRUE, update_cache = FALSE,
  cache = get("opentree_chronograms"))
}
\arguments{
\item{input}{vector of names, a newick string, or a phylo object}

\item{partial}{Boolean; default TRUE: use source trees even if they only match some of the desired taxa}

\item{usetnrs}{Boolean; default False. If TRUE, use OpenTree's services to resolve names. This can dramatically improve the chance of matches, but also take much longer}

\item{approximatematch}{Boolean; default TRUE: use a slower TNRS to correct mispellings, increasing the chance of matches (including false matches)}

\item{update_cache}{default to FALSE}

\item{cache}{The cached set of chronograms and other info from data(opentree_chronograms)}
}
\value{
data.frame of calibrations
}
\description{
Get all calibrations given trees in database
}
