% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{GetBoldOToLTree}
\alias{GetBoldOToLTree}
\title{Use Barcode of Life data to get branch lengths on the OToL tree of a set of taxa.}
\usage{
GetBoldOToLTree(input = c("Rhea americana", "Struthio camelus",
  "Gallus gallus"), usetnrs = FALSE, approximatematch = TRUE,
  marker = "COI", otol_version = "v2", chronogram = TRUE, doML = FALSE,
  process_input = TRUE, sppfromtaxon = FALSE)
}
\arguments{
\item{input}{Target taxa names in the form of a vector of characters, a newick character string, or a phylo object.}

\item{usetnrs}{If TRUE, use OpenTree's services to resolve names. This can dramatically improve the chance of matches, but also take much longer.}

\item{approximatematch}{If TRUE, use a slower TNRS to correct mispellings, increasing the chance of matches (including false matches).}

\item{marker}{A character vector with the name of the gene from Barcode of Life Data Systems (BOLD) to be used for branch length estimation.}

\item{otol_version}{Version of OToL to use}

\item{chronogram}{Boolean; default to TRUE:  branch lengths represent time estimated with ape::chronoMPL. If FALSE, branch lengths represent relative substitution rates estimated with phangorn::acctran.}

\item{doML}{Boolean; if TRUE, does ML branch length optimization with phangorn::optim.pml}

\item{process_input}{default to TRUE; specifies wether the input should be processed by ProcessInput function}

\item{sppfromtaxon}{boolean vector, default to FALSE. If TRUE, will get all species names from taxon names given in input. Must have same length as input. If input is a newick string , with some clades it will be converted to phylo object phy, and the order of sppfromtaxon will match phy$tip.label.}
}
\value{
A phylogeny with ML branch lengths
}
\description{
Use Barcode of Life data to get branch lengths on the OToL tree of a set of taxa.
}
