% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{GetBoldOToLTree}
\alias{GetBoldOToLTree}
\title{Use Barcode of Life data to get branch lengths on the OToL tree}
\usage{
GetBoldOToLTree(input = c("Rhea americana", "Struthio camelus",
  "Gallus gallus"), marker = "COI", otol_version = "v2", doML = FALSE)
}
\arguments{
\item{input}{A vector of names}

\item{marker}{Gene name to select}

\item{otol_version}{Version of OToL to use}

\item{doML}{Boolean; if TRUE, does ML brlen optimization}
}
\value{
A phylogeny with ML branch lengths
}
\description{
Use Barcode of Life data to get branch lengths on the OToL tree
}

