% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_chronograms.R
\name{GetOToLChronograms}
\alias{GetOToLChronograms}
\title{Get all chronograms from Open Tree of Life}
\usage{
GetOToLChronograms(verbose = FALSE)
}
\arguments{
\item{verbose}{If TRUE, give updates to the user}
}
\value{
A list with elements for the trees, authors, curators, and study ids
}
\description{
Get all chronograms from Open Tree of Life
}
