% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_chronograms.R
\name{HasBrlen}
\alias{HasBrlen}
\title{Check for branch lengths in a tree}
\usage{
HasBrlen(x)
}
\arguments{
\item{x}{A phylo object}
}
\value{
A TRUE or FALSE
}
\description{
Check for branch lengths in a tree
}
