% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_phyloch_functions.R
\name{HPDbars}
\alias{HPDbars}
\title{Function from phyloch package
 Add interval bars to nodes on a phylo plot.}
\usage{
HPDbars(
  phy,
  label = "height_95\%_HPD",
  tab = NULL,
  nodes,
  col,
  lwd,
  broken = FALSE,
  ...
)
}
\arguments{
\item{phy}{A phylo object with two extra elements specifying the MAX and MIN limits of intervals to plot.}

\item{label}{A character vector corresponding to the name used to label intervals stored in phy}

\item{tab}{Something I'm not sure of, should ask Christoph Heibl.}

\item{nodes}{Numeric. Use this if you wanna plot bars on certain nodes only.}

\item{col}{Color for the bars, default to "grey75"}

\item{lwd}{Line width for the bars, default to 5.}

\item{broken}{Something I'm not sure of, should ask Christoph Heibl.}

\item{...}{
  Arguments passed on to \code{\link[graphics:segments]{graphics::segments}}
  \describe{
    \item{\code{x0}}{coordinates of points \bold{from} which to draw.}
    \item{\code{y0}}{coordinates of points \bold{from} which to draw.}
    \item{\code{x1}}{coordinates of points \bold{to} which to draw.  At least
    one must be supplied.}
    \item{\code{y1}}{coordinates of points \bold{to} which to draw.  At least
    one must be supplied.}
    \item{\code{lty}}{\link[graphics]{graphical parameters} as in
    \code{\link[graphics]{par}}, possibly vectors.  \code{NA} values in \code{col}
    cause the segment to be omitted.}
  }}
}
\value{
Adds bars to nodes on a phylo plot.
}
\description{
Function from phyloch package
 Add interval bars to nodes on a phylo plot.
}
\details{
Make sure ape is loaded otherwise it won't find .PlotPhyloEnv
}
\author{
Christoph Heibl
}
