% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_query.R
\name{input_process}
\alias{input_process}
\title{Process a phylo object or a character string to determine if it's correct newick}
\usage{
input_process(input)
}
\arguments{
\item{input}{One of the following:
\describe{
\item{Taxon names}{As a character vector.}
\item{One or more trees with taxon names as tip labels}{As a \code{phylo} or \code{multiPhylo}
object, OR as a newick character string.}
\item{A \code{datelifeQuery} object}{An output from \code{\link[=make_datelife_query]{make_datelife_query()}}
\code{\link[=make_datelife_query]{make_datelife_query}}.}
}}
}
\value{
A phylo object or NA if no tree
}
\description{
Process a phylo object or a character string to determine if it's correct newick
}
