% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{input_process}
\alias{input_process}
\title{Takes a phylo object or a character string and figure out if it's correct newick format or a list of species}
\usage{
input_process(input, verbose = FALSE)
}
\arguments{
\item{input}{Target taxa names in the form of a vector of characters, a newick character string, or a phylo object.}

\item{verbose}{Boolean. If TRUE, it gives printed updates to the user.}
}
\value{
A phylo object or NA if no tree
}
\description{
Takes a phylo object or a character string and figure out if it's correct newick format or a list of species
}
