% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_query.R
\name{input_process}
\alias{input_process}
\title{Process a phylo object or a character string to determine if it's correct newick}
\usage{
input_process(input)
}
\arguments{
\item{input}{Taxon names as one of the following:
\describe{
\item{A character vector of taxon names}{With taxon names as a single comma separated starting or concatenated with \code{\link[=c]{c()}}.}
\item{A phylogenetic tree with taxon names as tip labels}{As a \code{phylo} or \code{multiPhylo}
object, OR as a newick character string.}
}}
}
\value{
A \code{phylo} object or \code{NA} if input is not a tree .
}
\description{
Process a phylo object or a character string to determine if it's correct newick
}
