% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_query.R
\name{is_datelife_query}
\alias{is_datelife_query}
\title{Check if input is a \code{datelifeQuery} object}
\usage{
is_datelife_query(input)
}
\arguments{
\item{input}{An object to be checked as an object with essential properties of a 'datelifeQuery' object.}
}
\value{
Is determined by the second condition.
}
\description{
\code{is_datelife_query} checks for two things to be \code{TRUE} or \code{FALSE}.
First, that \code{input} is of class {datelifeQuery}.
Second, that \code{input} is a list that contains at least two elements of a \code{datelifeQuery} object:
\describe{
\item{cleaned_names}{A character vector of taxon names.}
\item{phy}{Either NA or a \code{phylo} object.}
}
}
\details{
If the object has the correct format but it has a class different than
\code{datelifeQuery}, the class is not modified.
}
