% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_chronograms.R
\name{is_good_chronogram}
\alias{is_good_chronogram}
\title{Check to see that a chronogram is valid}
\usage{
is_good_chronogram(phy)
}
\arguments{
\item{phy}{Input phylo object}
}
\value{
Boolean: TRUE if good tree
}
\description{
Check to see that a chronogram is valid
}
