% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_ltt_summary.R
\name{ltt_summary}
\alias{ltt_summary}
\title{Lineage through time plots of  all chronograms in multiphylo object and its summaries.}
\usage{
ltt_summary(
  phy_summ,
  phy_summ_type = NULL,
  phy_summ_col = NULL,
  max_tips,
  length_arrowhead = 0.075,
  ...
)
}
\arguments{
\item{phy_summ}{A multiphylo object with chronograms from a summary.}

\item{phy_summ_type}{A character value.}

\item{phy_summ_col}{A character value.}

\item{max_tips}{A numeric value}

\item{length_arrowhead}{A numeric value}

\item{...}{parameters passed to ape::ltt.lines}
}
\description{
Lineage through time plots of  all chronograms in multiphylo object and its summaries.
}
