% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheauthors.R
\name{make_all_associations}
\alias{make_all_associations}
\title{Find all authors and where they have deposited their trees}
\usage{
make_all_associations(outputfile = "depositorcache.RData")
}
\arguments{
\item{outputfile}{Path including file name. NULL to prevent saving.}
}
\value{
a \code{data.frame} of "person" and "urls".
}
\description{
Find all authors and where they have deposited their trees
}
