% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{make_bladj_tree}
\alias{make_bladj_tree}
\title{Takes a tree and uses bladj to estimate node ages and branch lengths given a set of fixed node ages and respective node names}
\usage{
make_bladj_tree(tree = NULL, nodenames = NULL, nodeages = NULL)
}
\arguments{
\item{tree}{A tree either as a newick character string or as a phylo object}

\item{nodenames}{A character vector with node names from tree with fixed ages}

\item{nodeages}{A numeric vector with known or fixed node ages from tree}
}
\value{
A phylo tree
}
\description{
Takes a tree and uses bladj to estimate node ages and branch lengths given a set of fixed node ages and respective node names
}
