% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_query.R
\name{make_datelife_query}
\alias{make_datelife_query}
\title{Go from taxon names from a character vector, a phylo object or a newick character
string to a \code{datelifeQuery} object}
\usage{
make_datelife_query(
  input = c("Rhea americana", "Pterocnemia pennata", "Struthio camelus"),
  use_tnrs = FALSE,
  approximate_match = TRUE,
  get_spp_from_taxon = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{input}{One of the following:
\describe{
\item{Taxon names}{As a character vector.}
\item{A tree with taxon names as tip labels}{As a \code{phylo} or \code{multiPhylo}
object, OR as a newick character string.}
}}

\item{verbose}{Boolean. If \code{TRUE}, it gives printed updates to the user.}
}
\value{
A datelifeQuery object, a list of three elements: $phy a phylo object or NA, if input is not a tree; a cleaned vector of taxon names; and $ott_ids a numeric vector of OTT ids if use_tnrs = TRUE, or NULL if use_tnrs = FALSE.
}
\description{
It processes \code{phylo} and \code{newick character} string inputs
with \code{\link[=input_process]{input_process()}}.
}
\details{
If \code{input} is a \code{multiPhylo} object, only the first \code{phylo}
element will be used.
}
