% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_query.R
\name{make_datelife_query}
\alias{make_datelife_query}
\title{Cleans taxon names from input character vector, phylo object or newick character string. Process the two latter with input_process first.}
\usage{
make_datelife_query(
  input = c("Rhea americana", "Pterocnemia pennata", "Struthio camelus"),
  use_tnrs = FALSE,
  approximate_match = TRUE,
  get_spp_from_taxon = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{input}{A character vector of taxon names, a tree as a 'phylo' object or a newick character string, or a 'datelifeQuery' object from make_datelife_query function.}

\item{use_tnrs}{If TRUE, use OpenTree's services to resolve names. This can dramatically improve the chance of matches, but also take much longer.}

\item{approximate_match}{If TRUE, use a slower TNRS to correct misspellings, increasing the chance of matches (including false matches).}

\item{get_spp_from_taxon}{boolean vector, default to FALSE. If TRUE, will get all species names from taxon names given in input. Must have same length as input. If input is a newick string , with some clades it will be converted to phylo object phy, and the order of get_spp_from_taxon will match phy$tip.label.}

\item{verbose}{Boolean. If TRUE, it gives printed updates to the user.}
}
\value{
A datelifeQuery object, a list of three elements: $phy a phylo object or NA, if input is not a tree; a cleaned vector of taxon names; and $ott_ids a numeric vector of OTT ids if use_tnrs = TRUE, or NULL if use_tnrs = FALSE.
}
\description{
Cleans taxon names from input character vector, phylo object or newick character string. Process the two latter with input_process first.
}
\details{
If input has length 1, get_spp_from_taxon is always set to TRUE (in datelife_search, not in here, because of function dependencies)
}
