% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_query.R
\name{make_datelife_query}
\alias{make_datelife_query}
\title{Cleans taxon names from input character vector, phylo object or newick character string. Process the two latter with input_process first.}
\usage{
make_datelife_query(
  input = c("Rhea americana", "Pterocnemia pennata", "Struthio camelus"),
  use_tnrs = FALSE,
  approximate_match = TRUE,
  get_spp_from_taxon = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{input}{Target taxa names as a character vector, a newick character string, or a phylo object.}

\item{use_tnrs}{If TRUE, use OpenTree's services to resolve names. This can dramatically improve the chance of matches, but also take much longer.}

\item{approximate_match}{If TRUE, use a slower TNRS to correct mispellings, increasing the chance of matches (including false matches).}

\item{get_spp_from_taxon}{boolean vector, default to FALSE. If TRUE, will get all species names from taxon names given in input. Must have same length as input. If input is a newick string , with some clades it will be converted to phylo object phy, and the order of get_spp_from_taxon will match phy$tip.label.}

\item{verbose}{Boolean. If TRUE, it gives printed updates to the user.}

\item{...}{
  Arguments passed on to \code{\link[rphylotastic:taxon_get_species]{rphylotastic::taxon_get_species}}
  \describe{
    \item{\code{filters}}{A character vector of strings to exclude.}
  }}
}
\value{
A list with the phy (or NA, if no tree) and cleaned vector of taxa
}
\description{
Cleans taxon names from input character vector, phylo object or newick character string. Process the two latter with input_process first.
}
\details{
If input has length 1, get_spp_from_taxon is always set to TRUE from datelife_search, not in here bc other things depend on this function.
}
