% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheauthors.R
\name{make_overlap_table}
\alias{make_overlap_table}
\title{Create an overlap table}
\usage{
make_overlap_table(results_table)
}
\arguments{
\item{results_table}{An "author.results" or "curator.results" \code{data.frame}}
}
\value{
A \code{data.frame} with information on curators and what clades they've worked on
}
\description{
Create an overlap table
}
