% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{opentree_chronograms}
\alias{opentree_chronograms}
\title{Cached chronograms and other data from OpenTree}
\format{A list with four lists contained within it
\describe{
\item{authors}{List of lists of authors for the included studies}
\item{curators}{List of lists of curators for the included studies}
\item{studies}{List of study identifiers}
\item{trees}{List storing the chronograms from OpenTree}
}}
\source{
\url{http://opentreeoflife.org}
}
\description{
A list with four elements for >100 chronograms from OpenTree
}

