% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{patristic_matrix_list_to_array}
\alias{patristic_matrix_list_to_array}
\title{Convert list of patristic matrices to a 3D array. Used inside: summarize_datelife_result, patristic_matrix_array_congruify.}
\usage{
patristic_matrix_list_to_array(patristic_matrix_list, pad = TRUE)
}
\arguments{
\item{patristic_matrix_list}{List of patristic matrices}

\item{pad}{If TRUE, pad missing entries}
}
\value{
A 3d array of patristic matrices
}
\description{
Convert list of patristic matrices to a 3D array. Used inside: summarize_datelife_result, patristic_matrix_array_congruify.
}
