% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{patristic_matrix_MRCA}
\alias{patristic_matrix_MRCA}
\title{Get time of MRCA from patristic matrix. Used in \code{\link[=datelife_result_MRCA]{datelife_result_MRCA()}}.}
\usage{
patristic_matrix_MRCA(patristic_matrix, na_rm = TRUE)
}
\arguments{
\item{patristic_matrix}{A patristic matrix (aka a \code{datelifeResult} object of length 1)}

\item{na_rm}{If \code{TRUE}, it drops rows containing \code{NA}s from the \code{datelifeResult}
patristic matrix; if \code{FALSE}, it returns \code{NA} where there are missing entries.}
}
\value{
The depth of the MRCA as a numeric vector.
}
\description{
Get time of MRCA from patristic matrix. Used in \code{\link[=datelife_result_MRCA]{datelife_result_MRCA()}}.
}
