% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{patristic_matrix_name_reorder}
\alias{patristic_matrix_name_reorder}
\title{Function to reorder a matrix so that row and column labels are in alphabetical order. Used in: patristic_matrix_pad.}
\usage{
patristic_matrix_name_reorder(patristic_matrix)
}
\arguments{
\item{patristic_matrix}{A patristic matrix with row and column names for taxa}
}
\value{
patristic_matrix A patristic matrix with row and column names for taxa in alphabetical order
}
\description{
Function to reorder a matrix so that row and column labels are in alphabetical order. Used in: patristic_matrix_pad.
}
