% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{patristic_matrix_taxa_all_matching}
\alias{patristic_matrix_taxa_all_matching}
\title{Are all desired taxa in the patristic_matrix? Used inside: results_list_process.}
\usage{
patristic_matrix_taxa_all_matching(patristic_matrix, taxa)
}
\arguments{
\item{patristic_matrix}{A patristic matrix, rownames and colnames must be taxa}

\item{taxa}{A vector of taxon names}
}
\value{
A Boolean
}
\description{
Are all desired taxa in the patristic_matrix? Used inside: results_list_process.
}
