% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_phylo.R
\name{patristic_matrix_to_phylo}
\alias{patristic_matrix_to_phylo}
\title{Convert patristic matrix to a phylo object. Used inside: summarize_datelife_result, CongruiyTree.}
\usage{
patristic_matrix_to_phylo(patristic_matrix, clustering_method = "nj",
  fix_negative_brlen = TRUE, fixing_method = 0, ultrametric = TRUE,
  variance_matrix = NULL)
}
\arguments{
\item{patristic_matrix}{A patristic matrix}

\item{clustering_method}{A character vector indicating the method to construct the tree. Options are "nj" for Neighbor-Joining and "upgma" for Unweighted Pair Group Method with Arithmetic Mean.}

\item{fix_negative_brlen}{Boolean indicating whether to fix negative branch lengths in resulting tree or not. Default to TRUE.}

\item{fixing_method}{A character vector specifying the method to fix branch lengths: "bladj", "mrbayes" or a number to be assigned to all branches meeting fixing_criterion}

\item{ultrametric}{Boolean indicating whether to force ultrametric or not.}

\item{variance_matrix}{A variance matrix from a datelifeResult list of patristic matrices. Usually an output from datelife_result_variance_matrix function.}
}
\value{
A rooted phylo object
}
\description{
Convert patristic matrix to a phylo object. Used inside: summarize_datelife_result, CongruiyTree.
}
