% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_sdm.R
\name{patristic_matrix_unpad}
\alias{patristic_matrix_unpad}
\title{Function to remove missing taxa from a \code{datelifeResult} object.}
\usage{
patristic_matrix_unpad(patristic_matrix)
}
\arguments{
\item{patristic_matrix}{A patristic matrix with row and column names for taxa}
}
\value{
patristic_matrix for all_taxa
}
\description{
Used in \code{\link[=datelife_result_sdm_phylo]{datelife_result_sdm_phylo()}}.
}
