% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{PatristicMatrixToTree}
\alias{PatristicMatrixToTree}
\title{Convert patristic matrix to a phylo object}
\usage{
PatristicMatrixToTree(patristic.matrix)
}
\arguments{
\item{patristic.matrix}{A patristic matrix}
}
\value{
A rooted phylo object
}
\description{
Convert patristic matrix to a phylo object
}
