% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_processing.R
\name{phylo_check}
\alias{phylo_check}
\title{Checks if phy is a phylo class object and/or a chronogram}
\usage{
phylo_check(phy = NULL, brlen = FALSE, dated = FALSE)
}
\arguments{
\item{phy}{A phylo object}

\item{brlen}{Boolean. If TRUE it checks if phylo object has branch lengths.}

\item{dated}{Boolean. If TRUE it checks if phylo object is ultrametric.}
}
\value{
Nothing
}
\description{
Checks if phy is a phylo class object and/or a chronogram
}
