% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{phylo_generate_uncertainty}
\alias{phylo_generate_uncertainty}
\title{Function to generate uncertainty in branch lengths using a lognormal}
\usage{
phylo_generate_uncertainty(phy, num = 100, age_sd = NULL, age_var = 0.1,
  age_scale = 0, alpha = 0.025, rescale = TRUE, verbose = FALSE)
}
\arguments{
\item{phy}{The input tree}

\item{num}{A numeric vector indicating the number of samples to be generated.}

\item{age_sd}{The standard deviation around the age to use for generating the uncertainty. If not a numeric value, var will be used to calculate it.}

\item{age_var}{The variance to calculate age_sd and generate unvcertainty.}

\item{age_scale}{How to scale sd by the depth of the node. If 0, same sd for all. If not, older nodes have more uncertainty}

\item{alpha}{The significance level on uncertainty to generate. By default 0.025}

\item{rescale}{Boolean. If true, observed age will be rescaled each round.}

\item{verbose}{Boolean. If TRUE, it gives printed updates to the user.}
}
\value{
A phylo or multiPhylo object with the same topology as phy but different branch lengths
}
\description{
Function to generate uncertainty in branch lengths using a lognormal
}
