% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{phylo_tiplabel_underscore_to_space}
\alias{phylo_tiplabel_underscore_to_space}
\title{Convert underscores to spaces in trees. Used inside: patristic_matrix_array_phylo_congruify, congruify_and_check.}
\usage{
phylo_tiplabel_underscore_to_space(phy)
}
\arguments{
\item{phy}{A phylo object}
}
\value{
A phylo object
}
\description{
Convert underscores to spaces in trees. Used inside: patristic_matrix_array_phylo_congruify, congruify_and_check.
}
