% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{phylo_tiplabel_underscore_to_space}
\alias{phylo_tiplabel_underscore_to_space}
\title{Convert underscores to spaces in trees.}
\usage{
phylo_tiplabel_underscore_to_space(phy)
}
\arguments{
\item{phy}{A \code{phylo} object.}
}
\value{
A \code{phylo} object.
}
\description{
\code{phylo_tiplabel_underscore_to_space} is used inside \code{\link[=patristic_matrix_array_phylo_congruify]{patristic_matrix_array_phylo_congruify()}}, \code{\link[=congruify_and_check]{congruify_and_check()}}.
}
