% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{pick_grove}
\alias{pick_grove}
\title{Find the grove in the case of multiple groves}
\usage{
pick_grove(grove_list, criterion = c("trees", "taxa"), datelife_result)
}
\arguments{
\item{grove_list}{A list of vectors of tree indices. Each element is a grove}

\item{criterion}{Whether to get the grove with the most trees or the most taxa}

\item{datelife_result}{datelifeResult object (named list of patristic matrices). Only needed for "taxa" criterion}
}
\value{
a vector of the elements of the chosen grove
}
\description{
Find the grove in the case of multiple groves
}
