% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_ltt.R
\name{plot_ltt_phyloall}
\alias{plot_ltt_phyloall}
\title{Easy visualization of lineage through time plots of all chronograms in a multiphylo object.}
\usage{
plot_ltt_phyloall(
  taxon = NULL,
  phy,
  ltt_colors = NULL,
  tax_datedotol = NULL,
  file_name = NULL,
  file_dir = NULL,
  height = 3.5,
  width = 7,
  add_legend = FALSE,
  add_title = FALSE,
  title_text = NULL,
  study_number_cex = 0.75,
  lwd_arrows = 2,
  ...
)
}
\arguments{
\item{taxon}{Character vector indicating the name of the taxon or lineage that the chronograms in phy belong to.}

\item{phy}{A phylo or multiphylo object with chronograms (trees with branch lengths proportional to geologic time), ideally.}

\item{ltt_colors}{A character vector indicating the colors to be used for plotting ltt}

\item{tax_datedotol}{A chronogram to compare trees in phy to.}

\item{file_name}{A character string giving the name of the pdf file.}

\item{file_dir}{A character string giving the path to write the file to.}

\item{height}{Height of the plot output}

\item{width}{width of the plot output}

\item{add_legend}{Boolean}

\item{add_title}{Boolean}

\item{title_text}{Character vector}

\item{study_number_cex}{A numeric value}

\item{lwd_arrows}{A numeric value}

\item{...}{
  Arguments passed on to \code{\link[ape:ltt.plot]{ape::ltt.plot}}
  \describe{
    \item{\code{xlab}}{a character string (or a variable of mode character)
    giving the label for the \eqn{x}-axis (default is "Time").}
    \item{\code{ylab}}{idem for the \eqn{y}-axis (default is "N").}
    \item{\code{backward}}{a logical value: should the time axis be traced from
    the present (the default), or from the root of the tree?}
    \item{\code{tol}}{a numeric value (see details).}
  }}
}
\description{
Easy visualization of lineage through time plots of all chronograms in a multiphylo object.
}
