% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_ltt_summary.R
\name{plot_ltt_summary}
\alias{plot_ltt_summary}
\title{Lineage through time plots of all chronograms in multiphylo object and its summaries.}
\usage{
plot_ltt_summary(taxon, phy, phy_sdm, phy_median, file_name = NULL,
  file_dir = NULL, height = 3.5, width = 7, add_legend = TRUE,
  tax_datedotol = NULL)
}
\arguments{
\item{taxon}{Character vector indicating the name of the taxon or lineage that the chronograms in phy belong to.}

\item{phy}{A phylo or multiphylo object with chronograms (trees with branch lengths proportional to geologic time), ideally.}

\item{phy_sdm}{A multiphylo object with chronograms from and SDM summary.}

\item{phy_median}{A multiphylo object with chronograms from an median summary.}

\item{file_name}{A character string giving the name of the pdf file.}

\item{file_dir}{A character string giving the path to write the file to.}

\item{add_legend}{Boolean}

\item{tax_datedotol}{A chronogram to compare trees in phy to.}

\item{add_title}{Boolean}
}
\description{
Lineage through time plots of all chronograms in multiphylo object and its summaries.
}
