% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_ltt_summary.R
\name{plot_ltt_summary}
\alias{plot_ltt_summary}
\title{Lineage through time plots of all chronograms in multiphylo object and its summaries.}
\usage{
plot_ltt_summary(
  taxon,
  phy,
  phy_sdm,
  phy_median,
  file_name = NULL,
  file_dir = NULL,
  height = 3.5,
  width = 7,
  add_legend = TRUE,
  add_title = FALSE,
  col_sdm = "#00AFBB",
  col_median = "#CC79A7",
  tax_datedotol = NULL,
  ...
)
}
\arguments{
\item{taxon}{Character vector indicating the name of the taxon or lineage that the chronograms in phy belong to.}

\item{phy}{A phylo or multiphylo object with chronograms (trees with branch lengths proportional to geologic time), ideally.}

\item{phy_sdm}{A multiphylo object with chronograms from a SDM summary.}

\item{phy_median}{A multiphylo object with chronograms from a median summary.}

\item{file_name}{A character string giving the name of the pdf file.}

\item{file_dir}{A character string giving the path to write the file to.}

\item{height}{Height of the plot}

\item{width}{Width of the plot}

\item{add_legend}{Boolean}

\item{add_title}{Boolean}

\item{col_sdm}{Color of the SDM tree}

\item{col_median}{Color of the median tree}

\item{tax_datedotol}{A chronogram to compare other chronograms to.}

\item{...}{
  Arguments passed on to \code{\link[ape:ltt.plot]{ape::ltt.plot}}
  \describe{
    \item{\code{xlab}}{a character string (or a variable of mode character)
    giving the label for the \eqn{x}-axis (default is "Time").}
    \item{\code{ylab}}{idem for the \eqn{y}-axis (default is "N").}
    \item{\code{backward}}{a logical value: should the time axis be traced from
    the present (the default), or from the root of the tree?}
    \item{\code{tol}}{a numeric value (see details).}
  }}
}
\description{
Lineage through time plots of all chronograms in multiphylo object and its summaries.
}
