% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_phylo_all}
\alias{plot_phylo_all}
\title{plot all trees with study titles and geochronological axis}
\usage{
plot_phylo_all(trees, include = TRUE, individually = TRUE,
  cex = graphics::par("cex"), write = "no", file = "phylo_all")
}
\arguments{
\item{trees}{A list of trees as multiPhylo or as a plain list object.}

\item{include}{Boolean or numeric vector. Default to TRUE, plot all chronograms
in trees. If FALSE, exclude chronograms with only two tips. If numeric, plot
trees matching the numeric vector.}

\item{individually}{Boolean indicating if trees should be plotted one by one or all on the same file}

\item{cex}{a numeric value giving the factor scaling of the tip and
    node labels (Character EXpansion). The default is to take the
    current value from the graphical parameters.}

\item{write}{A character vector of length 1. Use pdf or png to write a file on those formats respectively. Anything else will not write any image file.}

\item{file}{A character string giving the name of the files. It will be the name of the directory too.}
}
\description{
plot all trees with study titles and geochronological axis
}
