% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{ProcessInput}
\alias{ProcessInput}
\title{Take input string, figure out if it's newick or list of species}
\usage{
ProcessInput(input = c("Rhea americana", "Pterocnemia pennata",
  "Struthio camelus"), usetnrs = "no", approximatematch = "yes")
}
\arguments{
\item{input}{A newick string or vector of taxa}

\item{usetnrs}{Whether to use OpenTree's TNRS for the input}

\item{approximatematch}{IDK}
}
\value{
A list with the phy (or NA, if no tree) and cleaned vector of taxa
}
\description{
Take input string, figure out if it's newick or list of species
}

