% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_taxonomy.R
\name{recover_mrcaott}
\alias{recover_mrcaott}
\title{Get an mrcaott tag from an OpenTree induced synthetic tree and get its name and ott id}
\usage{
recover_mrcaott(tag)
}
\arguments{
\item{tag}{A character vector with the mrca tag}
}
\value{
A numeric vector with ott id from original taxon named with the corresponding ott name
}
\description{
Get an mrcaott tag from an OpenTree induced synthetic tree and get its name and ott id
}
