% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife_internal.R
\name{results_list_process}
\alias{results_list_process}
\title{Take results_list and process it. Used inside: get_datelife_result}
\usage{
results_list_process(results_list, taxa = NULL, partial = FALSE)
}
\arguments{
\item{results_list}{A list returned from using get_subset_array_dispatch on opentree_chronograms$trees}

\item{taxa}{A vector of taxa to match}

\item{partial}{If TRUE, return matrices that have only partial matches}
}
\value{
A list with the patristic.matrices that are not NA
}
\description{
Take results_list and process it. Used inside: get_datelife_result
}
