% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{run}
\alias{run}
\title{Core function to generate results}
\usage{
run(
  input = c("Rhea americana", "Pterocnemia pennata", "Struthio camelus"),
  format = "citations",
  partial = "yes",
  plot.width = 600,
  plot.height = 600,
  use_tnrs = "no",
  opentree_chronograms = NULL
)
}
\arguments{
\item{input}{A newick string or vector of taxa}

\item{format}{The output format}

\item{partial}{How to deal with trees that have a subset of taxa in the query}

\item{plot.width}{Width in pixels for output plot}

\item{plot.height}{Height in pixels for output plot}

\item{use_tnrs}{Whether to use OpenTree's TNRS for the input}

\item{opentree_chronograms}{The list of lists containing the input trees and other info}
}
\value{
results in the desired format
}
\description{
Core function to generate results
}
