% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mrbayes_tree.R
\name{run_mrbayes}
\alias{run_mrbayes}
\title{Runs MrBayes from R}
\usage{
run_mrbayes(mrbayes_output_file = NULL)
}
\arguments{
\item{mrbayes_output_file}{A character vector specifying the name of mrBayes run file and outputs (can specify directory too).}
}
\value{
A phylo object with the consensus tree. MrBayes output files are stored in the working directory.
}
\description{
Runs MrBayes from R
}
