% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_phylo.R
\name{summarize_summary_matrix}
\alias{summarize_summary_matrix}
\title{Internal for summary_matrix_to_phylo().}
\usage{
summarize_summary_matrix(summ_matrix)
}
\description{
Internal for summary_matrix_to_phylo().
}
