% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_phylo.R
\name{summarize_summary_matrix}
\alias{summarize_summary_matrix}
\title{Internal function used in summary_matrix_to_phylo().}
\usage{
summarize_summary_matrix(summ_matrix)
}
\arguments{
\item{summ_matrix}{Any summary patristic distance matrix, such as the ones obtained with \code{\link[=datelife_result_sdm_matrix]{datelife_result_sdm_matrix()}} or \code{\link[=datelife_result_median_matrix]{datelife_result_median_matrix()}}.}
}
\description{
Internal function used in summary_matrix_to_phylo().
}
