% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{TabulateRelevantCurators}
\alias{TabulateRelevantCurators}
\title{Return the relevant curators for a set of studies}
\usage{
TabulateRelevantCurators(results.index, cache = get("opentree_chronograms"))
}
\arguments{
\item{results.index}{A vector from FindMatchingStudyIndex() with the indices of the relevant studies}

\item{cache}{The cache}
}
\value{
A vector with counts of each curator, with names equal to curator names
}
\description{
Return the relevant curators for a set of studies
}
