% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_taxonomy_general.R
\name{tnrs_match}
\alias{tnrs_match}
\title{Taxon name resolution service (tnrs) applied to query by batches}
\usage{
tnrs_match(names, reference_taxonomy = "otl", ...)
}
\arguments{
\item{names}{A character vector of taxon names to be matched to taxonomy. There is no limit to the numer of names that can be queried.}

\item{reference_taxonomy}{A character vetor specifying the reference taxonomy to use for tnrs.}

\item{...}{Arguments passed on to \code{rotl::tnrs_match_names}
\describe{
  \item{context_name}{name of the taxonomic context to be searched
(length-one character vector). Must match (case sensitive) one
of the values returned by \code{\link{tnrs_contexts}}.}
  \item{do_approximate_matching}{A logical indicating whether or not
to perform approximate string (a.k.a. \dQuote{fuzzy})
matching. Using \code{FALSE} will greatly improve
speed. Default, however, is \code{TRUE}.}
  \item{ids}{A vector of ids to use for identifying names. These
will be assigned to each name in the names array. If ids is
provided, then ids and names must be identical in length.}
  \item{include_suppressed}{Ordinarily, some quasi-taxa, such as
incertae sedis buckets and other non-OTUs, are suppressed from
TNRS results. If this parameter is true, these quasi-taxa are
allowed as possible TNRS results.}
}}
}
\value{
A data frame from tnrs_match_names function
}
\description{
Taxon name resolution service (tnrs) applied to query by batches
}
