% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_processing.R
\name{tree_check}
\alias{tree_check}
\title{Checks if a tree is a phylo class object otherwise it uses input_process.
Additionally it can check if tree is a chronogram with phylo_check}
\usage{
tree_check(tree = NULL, ...)
}
\arguments{
\item{tree}{A tree either as a newick character string or as a \code{phylo} object.}

\item{...}{
  Arguments passed on to \code{\link[=phylo_check]{phylo_check}}
  \describe{
    \item{\code{brlen}}{Boolean. If \code{TRUE} it checks if \code{phylo} object has branch lengths.}
    \item{\code{dated}}{Boolean. If \code{TRUE} it checks if \code{phylo} object is ultrametric.}
  }}
}
\value{
If tree is correctly formatted, it returns a \code{phylo} object.
}
\description{
Checks if a tree is a phylo class object otherwise it uses input_process.
Additionally it can check if tree is a chronogram with phylo_check
}
