% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_processing.R
\name{tree_get_node_data}
\alias{tree_get_node_data}
\title{Get node numbers, node names, descendant tip numbers and labels of nodes from any tree, and node ages from dated trees.}
\usage{
tree_get_node_data(
  tree = NULL,
  nodes = NULL,
  node_data = c("node_number", "node_label", "node_age", "descendant_tips_number",
    "descendant_tips_label")
)
}
\arguments{
\item{tree}{A tree either as a newick character string or as a \code{phylo} object.}

\item{nodes}{Numeric vector with node numbers from which you want to obtain data. Default to NULL: obtain data for all nodes in the tree.}

\item{node_data}{A character vector containing one or all from: "node_number", "node_label", "node_age", "descendant_tips_number", "descendant_tips_label"}
}
\value{
A list
}
\description{
Get node numbers, node names, descendant tip numbers and labels of nodes from any tree, and node ages from dated trees.
}
