% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_chronograms.R
\name{update_datelife_cache}
\alias{update_datelife_cache}
\title{Update chronogram database cache}
\usage{
update_datelife_cache(
  save = TRUE,
  file = "opentree_chronograms.RData",
  verbose = TRUE
)
}
\arguments{
\item{save}{Boolean; default TRUE: save all chronograms from Open Tree of Life to an RData file (default to opentree_chronograms.RData)}

\item{file}{Path including file name}

\item{verbose}{If TRUE, give status updates to the user}
}
\value{
None
}
\description{
Update chronogram database cache
}
