% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_chronograms.R
\name{update_datelife_cache}
\alias{update_datelife_cache}
\title{Update chronogram database cache}
\usage{
update_datelife_cache(save = TRUE, file = "opentree_chronograms.RData")
}
\arguments{
\item{save}{Default to \code{TRUE}, it saves all chronograms from Open Tree of Life to a ".RData" file (default to opentree_chronograms.RData)}

\item{file}{Path including file name}
}
\value{
None
}
\description{
Update chronogram database cache
}
