% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_calibrations.R
\name{use_all_calibrations}
\alias{use_all_calibrations}
\title{Use all calibrations from chronograms in a database to date a tree.}
\usage{
use_all_calibrations(phy = NULL, all_calibrations = NULL,
  partial = TRUE, use_tnrs = FALSE, approximate_match = TRUE,
  update_cache = FALSE, cache = get("opentree_chronograms"),
  expand = 0.1, giveup = 100, verbose = FALSE)
}
\arguments{
\item{phy}{A phylo object}

\item{all_calibrations}{A data frame of calibrations from get_all_calibrations function}

\item{partial}{If TRUE, use source trees even if they only match some of the desired taxa}

\item{use_tnrs}{If TRUE, use OpenTree's services to resolve names. This can dramatically improve the chance of matches, but also take much longer}

\item{approximate_match}{If TRUE, use a slower TNRS to correct mispellings, increasing the chance of matches (including false matches)}

\item{update_cache}{default to FALSE}

\item{cache}{The cached set of chronograms and other info from data(opentree_chronograms)}

\item{expand}{How much to expand by each step to get consistent calibrations}

\item{giveup}{How many expansions to try before giving up}

\item{verbose}{Boolean. If TRUE, it gives printed updates to the user.}
}
\value{
list with chronogram, original calibrations, and expanded calibrations
}
\description{
Use all calibrations from chronograms in a database to date a tree.
}
\details{
This will try to use the calibrations as fixed ages.
If that fails (often due to conflict between calibrations), it will expand the range of the minage and maxage and try again. And repeat.
expand sets the expansion value: should be between 0 and 1
}
