% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_calibrations_use.R
\name{use_all_calibrations}
\alias{use_all_calibrations}
\title{Perform a dating analysis on a set of input taxa.
It searches and uses all calibrations from chronograms in the DateLife database.}
\usage{
use_all_calibrations(input = NULL, ...)
}
\arguments{
\item{input}{Target taxa names as a character vector, a newick character string, or a phylo object.}

\item{...}{Arguments passed on to \code{get_all_calibrations}
\describe{
  \item{input}{vector of names, a newick string, a phylo or multiPhylo object, a datelifeResult object}
  \item{partial}{Boolean; default TRUE: use source trees even if they only match some of the desired taxa}
  \item{use_tnrs}{Boolean; default False. If TRUE, use OpenTree's services to resolve names. This can dramatically improve the chance of matches, but also take much longer}
  \item{approximate_match}{Boolean; default TRUE: use a slower TNRS to correct mispellings, increasing the chance of matches (including false matches)}
  \item{cache}{The cached set of chronograms and other info from data(opentree_chronograms)}
  \item{update_cache}{default to FALSE}
  \item{verbose}{Boolean. If TRUE, it gives printed updates to the user.}
}}
}
\value{
A list with a chronogram and all calibrations found
}
\description{
Perform a dating analysis on a set of input taxa.
It searches and uses all calibrations from chronograms in the DateLife database.
}
\details{
If input is a tree, it will get all calibrations and date it with bladj if it
has no branch lengths or with PATHd8 if it has branch lengths.
If input is a vector of names, it will try to construct a bold tree first to get
a topology with branch lengths. If it can't, it will get the open tree of life
topology only and date it with bladj.
}
