% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrations_use.R
\name{use_all_calibrations}
\alias{use_all_calibrations}
\title{Date a given tree topology using a given set of calibrations}
\usage{
use_all_calibrations(
  phy = NULL,
  calibrations = NULL,
  each = FALSE,
  dating_method = "bladj"
)
}
\arguments{
\item{phy}{A \code{phylo} object with or without branch lengths}

\item{calibrations}{An object of class \code{datelifeCalibrations},
typically an output of [get_all_calibrations()].}

\item{each}{Boolean, default to \code{FALSE}: all calibrations are returned in
the same data frame. If \code{TRUE}, calibrations from each chronogram are returned
in separate data frames.}

\item{dating_method}{Tree dating method. Options are "bladj" or "pathd8"}
}
\value{
A list with a chronogram and secondary calibrations used.
}
\description{
\code{use_all_calibrations} generates one or multiple chronograms (i.e., phylogenetic
trees with branch lengths proportional to time) by dating a tree topology given
in \code{phy}, and secondary calibrations given in \code{calibrations},
using bladj or PATHd8.
}
\details{
If \code{input} tree does not have branch lengths, \code{dating_method} is
ignored and the function will use secondary calibrations to date the \code{input}
tree with [BLADJ](http://phylodiversity.net/bladj/)
using [phylocomr::ph_bladj()] (via [use_calibrations_bladj()]). If \code{input}
tree has have branch lengths it can be dated
with [PATHd8](https://www2.math.su.se/PATHd8/) using [geiger::PATHd8.phylo()]
(via [use_calibrations_pathd8()]).
}
