% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrations_use.R
\name{use_all_calibrations}
\alias{use_all_calibrations}
\title{Date a given tree topology using a given set of calibrations}
\usage{
use_all_calibrations(phy = NULL, calibrations = NULL, dating_method = "bladj")
}
\arguments{
\item{phy}{A \code{phylo} object with or without branch lengths}

\item{calibrations}{A data frame (or list of data frames) of secondary calibrations
for any pair of taxon names in \code{phy}. See \code{\link[=get_all_calibrations]{get_all_calibrations}}.}

\item{dating_method}{Tree dating method. Options are "bladj" or "pathd8"}
}
\value{
A list with a chronogram and secondary calibrations used.
}
\description{
\code{use_all_calibrations} generates one or multiple chronograms (i.e., phylogenetic
trees with branch lengths proportional to time) by dating a tree topology given
in \code{phy}, and secondary calibrations given in \code{calibrations},
using bladj or PATHd8.
}
\details{
If input is a tree, it will use secondary calibrations to (1) date the tree with bladj
if it has no branch lengths, or (2) date the tree with PATHd8 if it has branch lengths.
If input is a vector of taxon names, it will attempt to reconstruct a BOLD tree first
to get a topology with branch lengths. If it can't, it will get an Open Tree
of Life synthetic tree topology and will date it with bladj.
}
