% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrations_use.R
\name{use_calibrations}
\alias{use_calibrations}
\title{Date a given tree topology using a combined set of given calibrations.}
\usage{
use_calibrations(
  phy = NULL,
  calibrations = NULL,
  dating_method = "bladj",
  type = "median",
  ...
)
}
\arguments{
\item{phy}{A \code{phylo} object with or without branch lengths}

\item{calibrations}{A data frame (or list of data frames) of secondary calibrations
for any pair of taxon names in \code{phy}. See \code{\link[=get_all_calibrations]{get_all_calibrations}}.}

\item{dating_method}{Tree dating method. Options are "bladj" or "pathd8"}

\item{...}{
  Arguments passed on to \code{\link[=use_calibrations_pathd8]{use_calibrations_pathd8}}
  \describe{
    \item{\code{expand}}{How much to expand by each step to get consistent calibrations. Should be between 0 and 1.}
    \item{\code{giveup}}{How many expansions to try before giving up}
  }}
}
\value{
A \code{phylo} object with branch lengths proportional to time.
}
\description{
\code{use_calibrations} combines all given calibrations and uses them as
constraints to perform a dating analysis on a given tree topology, using BLADJ
if it has no branch lengths, or PATHd8 if the given tree topology has initial
branch lengths.
}
\details{
If phy does not have branch lengths, dating_method is ignored and BLADJ will be used.
}
