% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_calibrations_bladj.R
\name{use_calibrations_bladj}
\alias{use_calibrations_bladj}
\title{Use calibrations to date a topology with bladj.}
\usage{
use_calibrations_bladj(phy, calibrations, type = "median")
}
\arguments{
\item{phy}{A phylo object}

\item{calibrations}{A data frame of calibrations from get_all_calibrations function}

\item{type}{The type of age to use as calibration.}
}
\value{
A phylo object
}
\description{
Use calibrations to date a topology with bladj.
}
