% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datelife.R
\name{use_calibrations_bladj}
\alias{use_calibrations_bladj}
\title{Use calibrations to date a topology with bladj.}
\usage{
use_calibrations_bladj(phy, calibrations, type = "median")
}
\arguments{
\item{phy}{A phylo object}

\item{calibrations}{A data frame of calibrations from get_all_calibrations function}

\item{type}{A character vector}
}
\value{
A phylo object
}
\description{
Use calibrations to date a topology with bladj.
}
