% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrations_use.R
\name{use_calibrations_each}
\alias{use_calibrations_each}
\title{Use given calibrations to generate one or multiple chronograms for a given
tree topology.}
\usage{
use_calibrations_each(phy = NULL, calibrations = NULL, ...)
}
\arguments{
\item{phy}{A tree with or without branch lengths to be dated. Only as phylo object for now.}

\item{calibrations}{Can be NULL. A list of dataframes from get_all_calibrations function.}

\item{...}{Arguments to pass to use_calibrations}
}
\value{
A list with a multiPhylo object of chronograms and a list of all calibrations used
}
\description{
\code{use_calibrations_each} uses each set of calibrations given by the user
independently as constraints for BLADJ or PATHd8 to date a given tree topology.
}
\details{
You can get the datelifeResult object or the list of chronograms first
phylo_all1 <- get_datelife_result(input = my_phy)
phylo_all2 <- summarize_datelife_result(phylo_all1)
Either will work the same:
use_calibrations_each(phy = my_phy, phylo_all = phylo_all1)
use_calibrations_each(phy = my_phy, phylo_all = phylo_all2)
You can also get the list of calibrations outside
my_calibrations <- get_all_calibrations(input = phylo_all1, each = TRUE)
use_calibrations_each(phy = my_phy, calibrations = my_calibrations)
All this means that you can use your own set of calibrations, not necessarily the ones found only in datelife.
}
