% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrations_use.R
\name{use_calibrations_each}
\alias{use_calibrations_each}
\title{Date a given tree topology by using a given list of calibrations independently,
to generate multiple hypothesis of time of divergence}
\usage{
use_calibrations_each(phy = NULL, calibrations = NULL, ...)
}
\arguments{
\item{phy}{A \code{phylo} object with or without branch lengths}

\item{calibrations}{An object of class \code{datelifeCalibrations},
typically an output of \code{\link[=get_all_calibrations]{get_all_calibrations()}}.}

\item{...}{Arguments to pass to \code{use_calibrations}.}
}
\value{
A \code{multiPhylo} object of trees with branch lengths proportional
to time.
}
\description{
\code{use_calibrations_each} wraps \code{use_calibrations} to take each set of
given calibrations and use it independently as constraints for BLADJ or PATHd8
to date a given tree topology.
}
\details{
\code{calibrations} and \code{dating method} are stored as attributes.
They can be accessed with \code{attributes(return)$datelife_calibrations} and
\code{attributes(return)$dating_method}, respectively. \cr
If \code{input} tree does not have branch lengths, \code{dating_method} is
ignored and the function will use secondary calibrations to date the
\code{input} tree with \href{http://phylodiversity.net/bladj/}{BLADJ},
using \code{\link[phylocomr:ph_bladj]{phylocomr::ph_bladj()}}, via \code{\link[=use_calibrations_bladj]{use_calibrations_bladj()}}. If
\code{input} tree has have branch lengths it can be dated
with \href{https://www2.math.su.se/PATHd8/}{PATHd8}, using \code{\link[geiger:geiger-internal]{geiger::PATHd8.phylo()}},
via \code{\link[=use_calibrations_pathd8]{use_calibrations_pathd8()}}.
}
