% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_calibrations_pathd8.R
\name{use_calibrations_pathd8}
\alias{use_calibrations_pathd8}
\title{Date a tree with secondary calibrations using PATHd8}
\usage{
use_calibrations_pathd8(
  phy = NULL,
  calibrations = NULL,
  expand = 0.1,
  giveup = 100
)
}
\arguments{
\item{phy}{A \code{phylo} object with branch lengths.}

\item{calibrations}{A data frame of secondary calibrations for any pair of taxon
names in \code{phy}. See \code{\link[=get_all_calibrations]{get_all_calibrations}}.}

\item{expand}{How much to expand by each step to get consistent calibrations. Should be between 0 and 1.}

\item{giveup}{How many expansions to try before giving up}
}
\value{
A phylo object with branch length proportional to time.
}
\description{
\code{use_calibrations_pathd8} uses secondary calibrations to date a tree with initial branch lengths using PATHd8.
}
\details{
This function will try to use the calibrations as fixed ages.
If that fails (often due to conflict between calibrations), it will expand the
range of the minimum age and maximum age and try again. And repeat.
If expand = 0, it uses the summarized calibrations.
In some cases, it returns edge lengths in relative time (with maximum tree depth = 1)
instead of absolute time, as given by calibrations. In this case, the function returns NA.
This is an issue from PATHd8.
}
