% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_calibrations_treePL.R
\name{use_calibrations_treePL}
\alias{use_calibrations_treePL}
\title{Use calibrations to date a tree with branch lengths with treePL.}
\usage{
use_calibrations_treePL(phy, calibrations)
}
\arguments{
\item{phy}{A \code{phylo} object with or without branch lengths.}

\item{calibrations}{A data frame of secondary calibrations for any pair of taxon
names in \code{phy}. See \code{\link[=get_all_calibrations]{get_all_calibrations}}.}
}
\value{
A phylo object
}
\description{
Use calibrations to date a tree with branch lengths with treePL.
}
\details{
This function will try to use the calibrations as fixed ages.
If that fails (often due to conflict between calibrations), it will expand the
range of the minimum age and maximum age and try again. And repeat.
If expand = 0, it uses the summarized calibrations.
In some cases, it returns edge lengths in relative time (with maximum tree depth = 1)
instead of absolute time, as given by calibrations. In this case, the function returns NA.
This is an issue from PATHd8.
}
